@@ -9,16 +9,16 @@ license = "MIT"
 
 # Requirements
 [tool.poetry.dependencies]
-python = "~2.7 || ^3.6"
+python = "^3.6"
 
 [tool.poetry.dev-dependencies]
 
 [[tool.poetry.source]]
 name = "foo"
 url = "https://foo.bar/simple/"
-secondary = true
+priority = "secondary"
 
 [[tool.poetry.source]]
 name = "bar"
 url = "https://bar.baz/simple/"
-secondary = true
+priority = "secondary"
