@@ -1,17 +1,19 @@
+from __future__ import annotations
+
+import logging
 import os
 import re
 import shutil
 import sys
-import tempfile
 
+from contextlib import suppress
 from pathlib import Path
+from typing import TYPE_CHECKING
 from typing import Any
-from typing import Dict
 
 import httpretty
 import pytest
 
-from cleo.testers.command_tester import CommandTester
 from keyring.backend import KeyringBackend
 
 from poetry.config.config import Config as BaseConfig
@@ -19,110 +21,172 @@ from poetry.config.dict_config_source import DictConfigSource
 from poetry.factory import Factory
 from poetry.inspection.info import PackageInfo
 from poetry.inspection.info import PackageInfoError
-from poetry.installation import Installer
 from poetry.layouts import layout
-from poetry.repositories import Pool
 from poetry.repositories import Repository
+from poetry.repositories import RepositoryPool
+from poetry.utils.cache import ArtifactCache
 from poetry.utils.env import EnvManager
 from poetry.utils.env import SystemEnv
 from poetry.utils.env import VirtualEnv
-from tests.helpers import TestExecutor
+from tests.helpers import MOCK_DEFAULT_GIT_REVISION
 from tests.helpers import TestLocker
 from tests.helpers import TestRepository
 from tests.helpers import get_package
+from tests.helpers import isolated_environment
 from tests.helpers import mock_clone
 from tests.helpers import mock_download
 
 
+if TYPE_CHECKING:
+    from collections.abc import Iterator
+    from collections.abc import Mapping
+
+    from _pytest.config import Config as PyTestConfig
+    from _pytest.config.argparsing import Parser
+    from pytest_mock import MockerFixture
+
+    from poetry.poetry import Poetry
+    from tests.types import FixtureCopier
+    from tests.types import FixtureDirGetter
+    from tests.types import ProjectFactory
+
+
+def pytest_addoption(parser: Parser) -> None:
+    parser.addoption(
+        "--integration",
+        action="store_true",
+        dest="integration",
+        default=False,
+        help="enable integration tests",
+    )
+
+
+def pytest_configure(config: PyTestConfig) -> None:
+    config.addinivalue_line("markers", "integration: mark integration tests")
+
+    if not config.option.integration:
+        config.option.markexpr = "not integration"
+
+
 class Config(BaseConfig):
+    _config_source: DictConfigSource
+    _auth_config_source: DictConfigSource
+
     def get(self, setting_name: str, default: Any = None) -> Any:
         self.merge(self._config_source.config)
         self.merge(self._auth_config_source.config)
 
-        return super(Config, self).get(setting_name, default=default)
+        return super().get(setting_name, default=default)
 
-    def raw(self) -> Dict[str, Any]:
+    def raw(self) -> dict[str, Any]:
         self.merge(self._config_source.config)
         self.merge(self._auth_config_source.config)
 
-        return super(Config, self).raw()
+        return super().raw()
 
-    def all(self) -> Dict[str, Any]:
+    def all(self) -> dict[str, Any]:
         self.merge(self._config_source.config)
         self.merge(self._auth_config_source.config)
 
-        return super(Config, self).all()
+        return super().all()
 
 
 class DummyBackend(KeyringBackend):
-    def __init__(self):
-        self._passwords = {}
+    def __init__(self) -> None:
+        self._passwords: dict[str, dict[str | None, str | None]] = {}
 
     @classmethod
-    def priority(cls):
+    def priority(cls) -> int:
         return 42
 
-    def set_password(self, service, username, password):
+    def set_password(self, service: str, username: str | None, password: Any) -> None:
         self._passwords[service] = {username: password}
 
-    def get_password(self, service, username):
+    def get_password(self, service: str, username: str | None) -> Any:
         return self._passwords.get(service, {}).get(username)
 
-    def get_credential(self, service, username):
+    def get_credential(self, service: str, username: str | None) -> Any:
         return self._passwords.get(service, {}).get(username)
 
-    def delete_password(self, service, username):
+    def delete_password(self, service: str, username: str | None) -> None:
         if service in self._passwords and username in self._passwords[service]:
             del self._passwords[service][username]
 
 
 @pytest.fixture()
-def dummy_keyring():
+def dummy_keyring() -> DummyBackend:
     return DummyBackend()
 
 
 @pytest.fixture()
-def with_simple_keyring(dummy_keyring):
+def with_simple_keyring(dummy_keyring: DummyBackend) -> None:
     import keyring
 
-    keyring.set_keyring(dummy_keyring)
+    keyring.set_keyring(dummy_keyring)  # type: ignore[no-untyped-call]
 
 
 @pytest.fixture()
-def with_fail_keyring():
+def with_fail_keyring() -> None:
     import keyring
 
     from keyring.backends.fail import Keyring
 
-    keyring.set_keyring(Keyring())
+    keyring.set_keyring(Keyring())  # type: ignore[no-untyped-call]
 
 
 @pytest.fixture()
-def with_chained_keyring(mocker):
+def with_null_keyring() -> None:
+    import keyring
+
+    from keyring.backends.null import Keyring
+
+    keyring.set_keyring(Keyring())  # type: ignore[no-untyped-call]
+
+
+@pytest.fixture()
+def with_chained_fail_keyring(mocker: MockerFixture) -> None:
     from keyring.backends.fail import Keyring
 
-    mocker.patch("keyring.backend.get_all_keyring", [Keyring()])
+    mocker.patch(
+        "keyring.backend.get_all_keyring",
+        lambda: [Keyring()],  # type: ignore[no-untyped-call]
+    )
     import keyring
 
     from keyring.backends.chainer import ChainerBackend
 
-    keyring.set_keyring(ChainerBackend())
+    keyring.set_keyring(ChainerBackend())  # type: ignore[no-untyped-call]
+
+
+@pytest.fixture()
+def with_chained_null_keyring(mocker: MockerFixture) -> None:
+    from keyring.backends.null import Keyring
+
+    mocker.patch(
+        "keyring.backend.get_all_keyring",
+        lambda: [Keyring()],  # type: ignore[no-untyped-call]
+    )
+    import keyring
+
+    from keyring.backends.chainer import ChainerBackend
+
+    keyring.set_keyring(ChainerBackend())  # type: ignore[no-untyped-call]
 
 
 @pytest.fixture
-def config_cache_dir(tmp_dir):
-    path = Path(tmp_dir) / ".cache" / "pypoetry"
+def config_cache_dir(tmp_path: Path) -> Path:
+    path = tmp_path / ".cache" / "pypoetry"
     path.mkdir(parents=True)
     return path
 
 
 @pytest.fixture
-def config_virtualenvs_path(config_cache_dir):
+def config_virtualenvs_path(config_cache_dir: Path) -> Path:
     return config_cache_dir / "virtualenvs"
 
 
 @pytest.fixture
-def config_source(config_cache_dir):
+def config_source(config_cache_dir: Path) -> DictConfigSource:
     source = DictConfigSource()
     source.add_property("cache-dir", str(config_cache_dir))
 
@@ -130,165 +194,158 @@ def config_source(config_cache_dir):
 
 
 @pytest.fixture
-def auth_config_source():
+def auth_config_source() -> DictConfigSource:
     source = DictConfigSource()
 
     return source
 
 
-@pytest.fixture
-def config(config_source, auth_config_source, mocker):
+@pytest.fixture(autouse=True)
+def config(
+    config_source: DictConfigSource,
+    auth_config_source: DictConfigSource,
+    mocker: MockerFixture,
+) -> Config:
     import keyring
 
     from keyring.backends.fail import Keyring
 
-    keyring.set_keyring(Keyring())
+    keyring.set_keyring(Keyring())  # type: ignore[no-untyped-call]
 
     c = Config()
     c.merge(config_source.config)
     c.set_config_source(config_source)
     c.set_auth_config_source(auth_config_source)
 
-    mocker.patch("poetry.factory.Factory.create_config", return_value=c)
+    mocker.patch("poetry.config.config.Config.create", return_value=c)
     mocker.patch("poetry.config.config.Config.set_config_source")
 
     return c
 
 
+@pytest.fixture
+def artifact_cache(config: Config) -> ArtifactCache:
+    return ArtifactCache(cache_dir=config.artifacts_cache_directory)
+
+
+@pytest.fixture()
+def config_dir(tmp_path: Path) -> Path:
+    path = tmp_path / "config"
+    path.mkdir()
+    return path
+
+
 @pytest.fixture(autouse=True)
-def mock_user_config_dir(mocker):
-    config_dir = tempfile.mkdtemp(prefix="poetry_config_")
+def mock_user_config_dir(mocker: MockerFixture, config_dir: Path) -> None:
     mocker.patch("poetry.locations.CONFIG_DIR", new=config_dir)
-    mocker.patch("poetry.factory.CONFIG_DIR", new=config_dir)
-    yield
-    shutil.rmtree(config_dir, ignore_errors=True)
+    mocker.patch("poetry.config.config.CONFIG_DIR", new=config_dir)
 
 
 @pytest.fixture(autouse=True)
-def download_mock(mocker):
+def download_mock(mocker: MockerFixture) -> None:
     # Patch download to not download anything but to just copy from fixtures
     mocker.patch("poetry.utils.helpers.download_file", new=mock_download)
-    mocker.patch("poetry.puzzle.provider.download_file", new=mock_download)
-    mocker.patch("poetry.repositories.pypi_repository.download_file", new=mock_download)
+    mocker.patch("poetry.packages.direct_origin.download_file", new=mock_download)
+    mocker.patch("poetry.repositories.http_repository.download_file", new=mock_download)
 
 
 @pytest.fixture(autouse=True)
-def pep517_metadata_mock(mocker):
-    @classmethod  # noqa
-    def _pep517_metadata(cls, path):
-        try:
+def pep517_metadata_mock(mocker: MockerFixture) -> None:
+    def get_pep517_metadata(path: Path) -> PackageInfo:
+        with suppress(PackageInfoError):
             return PackageInfo.from_setup_files(path)
-        except PackageInfoError:
-            pass
         return PackageInfo(name="demo", version="0.1.2")
 
     mocker.patch(
-        "poetry.inspection.info.PackageInfo._pep517_metadata",
-        _pep517_metadata,
+        "poetry.inspection.info.get_pep517_metadata",
+        get_pep517_metadata,
     )
 
 
 @pytest.fixture
-def environ():
-    original_environ = dict(os.environ)
-
-    yield
-
-    os.environ.clear()
-    os.environ.update(original_environ)
+def environ() -> Iterator[None]:
+    with isolated_environment():
+        yield
 
 
 @pytest.fixture(autouse=True)
-def git_mock(mocker):
+def isolate_environ() -> Iterator[None]:
+    """Ensure the environment is isolated from user configuration."""
+    with isolated_environment():
+        for var in os.environ:
+            if var.startswith("POETRY_") or var in {"PYTHONPATH", "VIRTUAL_ENV"}:
+                del os.environ[var]
+
+        yield
+
+
+@pytest.fixture(autouse=True)
+def git_mock(mocker: MockerFixture) -> None:
     # Patch git module to not actually clone projects
-    mocker.patch("poetry.core.vcs.git.Git.clone", new=mock_clone)
-    mocker.patch("poetry.core.vcs.git.Git.checkout", new=lambda *_: None)
-    p = mocker.patch("poetry.core.vcs.git.Git.rev_parse")
-    p.return_value = "9cf87a285a2d3fbb0b9fa621997b3acc3631ed24"
+    mocker.patch("poetry.vcs.git.Git.clone", new=mock_clone)
+    p = mocker.patch("poetry.vcs.git.Git.get_revision")
+    p.return_value = MOCK_DEFAULT_GIT_REVISION
 
 
 @pytest.fixture
-def http():
+def http() -> Iterator[type[httpretty.httpretty]]:
     httpretty.reset()
-    httpretty.enable(allow_net_connect=False)
-
-    yield httpretty
-
-    httpretty.activate()
-    httpretty.reset()
+    with httpretty.enabled(allow_net_connect=False):
+        yield httpretty
 
 
 @pytest.fixture
-def fixture_base():
+def project_root() -> Path:
+    return Path(__file__).parent.parent
+
+
+@pytest.fixture(scope="session")
+def fixture_base() -> Path:
     return Path(__file__).parent / "fixtures"
 
 
-@pytest.fixture
-def fixture_dir(fixture_base):
-    def _fixture_dir(name):
+@pytest.fixture(scope="session")
+def fixture_dir(fixture_base: Path) -> FixtureDirGetter:
+    def _fixture_dir(name: str) -> Path:
         return fixture_base / name
 
     return _fixture_dir
 
 
 @pytest.fixture
-def tmp_dir():
-    dir_ = tempfile.mkdtemp(prefix="poetry_")
+def tmp_venv(tmp_path: Path) -> Iterator[VirtualEnv]:
+    venv_path = tmp_path / "venv"
 
-    yield dir_
-
-    shutil.rmtree(dir_)
-
-
-@pytest.fixture
-def mocked_open_files(mocker):
-    files = []
-    original = Path.open
-
-    def mocked_open(self, *args, **kwargs):
-        if self.name in {"pyproject.toml"}:
-            return mocker.MagicMock()
-        return original(self, *args, **kwargs)
-
-    mocker.patch("pathlib.Path.open", mocked_open)
-
-    yield files
-
-
-@pytest.fixture
-def tmp_venv(tmp_dir):
-    venv_path = Path(tmp_dir) / "venv"
-
-    EnvManager.build_venv(str(venv_path))
+    EnvManager.build_venv(venv_path)
 
     venv = VirtualEnv(venv_path)
     yield venv
 
-    shutil.rmtree(str(venv.path))
+    shutil.rmtree(venv.path)
 
 
 @pytest.fixture
-def installed():
-    return Repository()
+def installed() -> Repository:
+    return Repository("installed")
 
 
 @pytest.fixture(scope="session")
-def current_env():
+def current_env() -> SystemEnv:
     return SystemEnv(Path(sys.executable))
 
 
 @pytest.fixture(scope="session")
-def current_python(current_env):
+def current_python(current_env: SystemEnv) -> tuple[int, int, int]:
     return current_env.version_info[:3]
 
 
 @pytest.fixture(scope="session")
-def default_python(current_python):
-    return "^{}".format(".".join(str(v) for v in current_python[:2]))
+def default_python(current_python: tuple[int, int, int]) -> str:
+    return "^" + ".".join(str(v) for v in current_python[:2])
 
 
 @pytest.fixture
-def repo(http):
+def repo(http: type[httpretty.httpretty]) -> TestRepository:
     http.register_uri(
         http.GET,
         re.compile("^https?://foo.bar/(.+?)$"),
@@ -297,28 +354,43 @@ def repo(http):
 
 
 @pytest.fixture
-def project_factory(tmp_dir, config, repo, installed, default_python):
-    workspace = Path(tmp_dir)
+def project_factory(
+    tmp_path: Path,
+    config: Config,
+    repo: TestRepository,
+    installed: Repository,
+    default_python: str,
+    load_required_fixtures: None,
+) -> ProjectFactory:
+    workspace = tmp_path
 
     def _factory(
-        name=None,
-        dependencies=None,
-        dev_dependencies=None,
-        pyproject_content=None,
-        poetry_lock_content=None,
-        install_deps=True,
-    ):
-        project_dir = workspace / "poetry-fixture-{}".format(name)
+        name: str | None = None,
+        dependencies: Mapping[str, str] | None = None,
+        dev_dependencies: Mapping[str, str] | None = None,
+        pyproject_content: str | None = None,
+        poetry_lock_content: str | None = None,
+        install_deps: bool = True,
+        source: Path | None = None,
+        locker_config: dict[str, Any] | None = None,
+        use_test_locker: bool = True,
+    ) -> Poetry:
+        project_dir = workspace / f"poetry-fixture-{name}"
         dependencies = dependencies or {}
         dev_dependencies = dev_dependencies or {}
 
-        if pyproject_content:
-            project_dir.mkdir(parents=True, exist_ok=True)
-            with project_dir.joinpath("pyproject.toml").open(
-                "w", encoding="utf-8"
-            ) as f:
-                f.write(pyproject_content)
+        if pyproject_content or source:
+            if source:
+                project_dir.parent.mkdir(parents=True, exist_ok=True)
+                shutil.copytree(source, project_dir)
+            else:
+                project_dir.mkdir(parents=True, exist_ok=True)
+
+            if pyproject_content:
+                with (project_dir / "pyproject.toml").open("w", encoding="utf-8") as f:
+                    f.write(pyproject_content)
         else:
+            assert name is not None
             layout("src")(
                 name,
                 "0.1.0",
@@ -335,15 +407,17 @@ def project_factory(tmp_dir, config, repo, installed, default_python):
 
         poetry = Factory().create_poetry(project_dir)
 
-        locker = TestLocker(
-            poetry.locker.lock.path, poetry.locker._local_config
-        )  # noqa
-        locker.write()
+        if use_test_locker:
+            locker = TestLocker(
+                poetry.locker.lock, locker_config or poetry.locker._local_config
+            )
+            locker.write()
 
-        poetry.set_locker(locker)
+            poetry.set_locker(locker)
+
         poetry.set_config(config)
 
-        pool = Pool()
+        pool = RepositoryPool()
         pool.add_repository(repo)
 
         poetry.set_pool(pool)
@@ -360,53 +434,54 @@ def project_factory(tmp_dir, config, repo, installed, default_python):
     return _factory
 
 
-@pytest.fixture
-def command_tester_factory(app, env):
-    def _tester(command, poetry=None, installer=None, executor=None, environment=None):
-        command = app.find(command)
-        tester = CommandTester(command)
-
-        # Setting the formatter from the application
-        # TODO: Find a better way to do this in Cleo
-        app_io = app.create_io()
-        formatter = app_io.output.formatter
-        tester.io.output.set_formatter(formatter)
-        tester.io.error_output.set_formatter(formatter)
-
-        if poetry:
-            app._poetry = poetry
-
-        poetry = app.poetry
-        command._pool = poetry.pool
-
-        if hasattr(command, "set_env"):
-            command.set_env(environment or env)
-
-        if hasattr(command, "set_installer"):
-            installer = installer or Installer(
-                tester.io,
-                env,
-                poetry.package,
-                poetry.locker,
-                poetry.pool,
-                poetry.config,
-                executor=executor
-                or TestExecutor(env, poetry.pool, poetry.config, tester.io),
-            )
-            installer.use_executor(True)
-            command.set_installer(installer)
-
-        return tester
-
-    return _tester
+@pytest.fixture(autouse=True)
+def set_simple_log_formatter() -> None:
+    """
+    This fixture removes any formatting added via IOFormatter.
+    """
+    for name in logging.Logger.manager.loggerDict:
+        for handler in logging.getLogger(name).handlers:
+            # replace formatter with simple formatter for testing
+            handler.setFormatter(logging.Formatter(fmt="%(message)s"))
 
 
 @pytest.fixture
-def do_lock(command_tester_factory, poetry):
-    command_tester_factory("lock").execute()
-    assert poetry.locker.lock.exists()
+def fixture_copier(fixture_base: Path, tmp_path: Path) -> FixtureCopier:
+    def _copy(relative_path: str, target: Path | None = None) -> Path:
+        path = fixture_base / relative_path
+        target = target or (tmp_path / relative_path)
+        target.parent.mkdir(parents=True, exist_ok=True)
+
+        if target.exists():
+            return target
+
+        if path.is_dir():
+            shutil.copytree(path, target)
+        else:
+            shutil.copyfile(path, target)
+
+        return target
+
+    return _copy
 
 
+def required_fixtures() -> list[str]:
+    return []
+
+
+@pytest.fixture(autouse=True)
+def load_required_fixtures(
+    required_fixtures: list[str], fixture_copier: FixtureCopier
+) -> None:
+    for fixture in required_fixtures:
+        fixture_copier(fixture)
+
+        
 @pytest.fixture
-def venv_flags_default():
-    return {"always-copy": False, "system-site-packages": False}
+def venv_flags_default() -> dict[str, bool]:
+    return {
+      "always-copy": False,
+      "system-site-packages": False,
+      "no-pip": False,
+      "no-setuptools": False,
+    }
