@@ -1,47 +1,66 @@
+from __future__ import annotations
+
 import os
 import shutil
+import subprocess
 import sys
 
 from pathlib import Path
+from typing import TYPE_CHECKING
+from typing import Any
 
 import pytest
 
 from cleo.testers.command_tester import CommandTester
+from packaging.utils import canonicalize_name
 
-from poetry.repositories import Pool
-from poetry.utils._compat import decode
-from tests.helpers import PoetryTestApplication
+from poetry.console.application import Application
+from poetry.console.commands.init import InitCommand
+from poetry.repositories import RepositoryPool
 from tests.helpers import get_package
 
 
-@pytest.fixture
-def source_dir(tmp_path) -> Path:
-    cwd = os.getcwd()
+if TYPE_CHECKING:
+    from collections.abc import Iterator
 
+    from _pytest.fixtures import FixtureRequest
+    from poetry.core.packages.package import Package
+    from pytest_mock import MockerFixture
+
+    from poetry.config.config import Config
+    from poetry.poetry import Poetry
+    from tests.helpers import PoetryTestApplication
+    from tests.helpers import TestRepository
+    from tests.types import FixtureDirGetter
+
+
+@pytest.fixture
+def source_dir(tmp_path: Path) -> Iterator[Path]:
+    cwd = Path.cwd()
     try:
-        os.chdir(str(tmp_path))
-        yield Path(tmp_path.as_posix())
+        os.chdir(tmp_path)
+        yield tmp_path
     finally:
         os.chdir(cwd)
 
 
 @pytest.fixture
-def patches(mocker, source_dir, repo):
+def patches(mocker: MockerFixture, source_dir: Path, repo: TestRepository) -> None:
     mocker.patch("pathlib.Path.cwd", return_value=source_dir)
     mocker.patch(
-        "poetry.console.commands.init.InitCommand._get_pool", return_value=Pool([repo])
+        "poetry.console.commands.init.InitCommand._get_pool",
+        return_value=RepositoryPool([repo]),
     )
 
 
 @pytest.fixture
-def tester(patches):
-    # we need a test application without poetry here.
-    app = PoetryTestApplication(None)
+def tester(patches: None) -> CommandTester:
+    app = Application()
     return CommandTester(app.find("init"))
 
 
 @pytest.fixture
-def init_basic_inputs():
+def init_basic_inputs() -> str:
     return "\n".join(
         [
             "my-package",  # Package name
@@ -58,7 +77,7 @@ def init_basic_inputs():
 
 
 @pytest.fixture()
-def init_basic_toml():
+def init_basic_toml() -> str:
     return """\
 [tool.poetry]
 name = "my-package"
@@ -74,13 +93,22 @@ python = "~2.7 || ^3.6"
 """
 
 
-def test_basic_interactive(tester, init_basic_inputs, init_basic_toml):
+def test_basic_interactive(
+    tester: CommandTester, init_basic_inputs: str, init_basic_toml: str
+) -> None:
     tester.execute(inputs=init_basic_inputs)
     assert init_basic_toml in tester.io.fetch_output()
 
 
-def test_noninteractive(app, mocker, poetry, repo, tmp_path):
+def test_noninteractive(
+    app: PoetryTestApplication,
+    mocker: MockerFixture,
+    poetry: Poetry,
+    repo: TestRepository,
+    tmp_path: Path,
+) -> None:
     command = app.find("init")
+    assert isinstance(command, InitCommand)
     command._pool = poetry.pool
 
     repo.add_package(get_package("pytest", "3.6.0"))
@@ -94,17 +122,20 @@ def test_noninteractive(app, mocker, poetry, repo, tmp_path):
 
     expected = "Using version ^3.6.0 for pytest\n"
     assert tester.io.fetch_output() == expected
-    assert "" == tester.io.fetch_error()
+    assert tester.io.fetch_error() == ""
 
     toml_content = (tmp_path / "pyproject.toml").read_text()
     assert 'name = "my-package"' in toml_content
     assert 'pytest = "^3.6.0"' in toml_content
 
 
-def test_interactive_with_dependencies(tester, repo):
+def test_interactive_with_dependencies(
+    tester: CommandTester, repo: TestRepository
+) -> None:
     repo.add_package(get_package("django-pendulum", "0.1.6-pre4"))
     repo.add_package(get_package("pendulum", "2.0.0"))
     repo.add_package(get_package("pytest", "3.6.0"))
+    repo.add_package(get_package("flask", "2.0.0"))
 
     inputs = [
         "my-package",  # Package name
@@ -117,6 +148,9 @@ def test_interactive_with_dependencies(tester, repo):
         "pendulu",  # Search for package
         "1",  # Second option is pendulum
         "",  # Do not set constraint
+        "Flask",
+        "0",
+        "",
         "",  # Stop searching for packages
         "",  # Interactive dev packages
         "pytest",  # Search for package
@@ -140,6 +174,7 @@ packages = [{include = "my_package"}]
 [tool.poetry.dependencies]
 python = "~2.7 || ^3.6"
 pendulum = "^2.0.0"
+flask = "^2.0.0"
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^3.6.0"
@@ -148,7 +183,51 @@ pytest = "^3.6.0"
     assert expected in tester.io.fetch_output()
 
 
-def test_empty_license(tester):
+# Regression test for https://github.com/python-poetry/poetry/issues/2355
+def test_interactive_with_dependencies_and_no_selection(
+    tester: CommandTester, repo: TestRepository
+) -> None:
+    repo.add_package(get_package("django-pendulum", "0.1.6-pre4"))
+    repo.add_package(get_package("pendulum", "2.0.0"))
+    repo.add_package(get_package("pytest", "3.6.0"))
+
+    inputs = [
+        "my-package",  # Package name
+        "1.2.3",  # Version
+        "This is a description",  # Description
+        "n",  # Author
+        "MIT",  # License
+        "~2.7 || ^3.6",  # Python
+        "",  # Interactive packages
+        "pendulu",  # Search for package
+        "",  # Do not select an option
+        "",  # Stop searching for packages
+        "",  # Interactive dev packages
+        "pytest",  # Search for package
+        "",  # Do not select an option
+        "",
+        "",
+        "\n",  # Generate
+    ]
+    tester.execute(inputs="\n".join(inputs))
+    expected = """\
+[tool.poetry]
+name = "my-package"
+version = "1.2.3"
+description = "This is a description"
+authors = ["Your Name <you@example.com>"]
+license = "MIT"
+readme = "README.md"
+packages = [{include = "my_package"}]
+
+[tool.poetry.dependencies]
+python = "~2.7 || ^3.6"
+"""
+
+    assert expected in tester.io.fetch_output()
+
+
+def test_empty_license(tester: CommandTester) -> None:
     inputs = [
         "my-package",  # Package name
         "1.2.3",  # Version
@@ -162,7 +241,8 @@ def test_empty_license(tester):
     ]
     tester.execute(inputs="\n".join(inputs))
 
-    expected = """\
+    python = ".".join(str(c) for c in sys.version_info[:2])
+    expected = f"""\
 [tool.poetry]
 name = "my-package"
 version = "1.2.3"
@@ -173,13 +253,13 @@ packages = [{{include = "my_package"}}]
 
 [tool.poetry.dependencies]
 python = "^{python}"
-""".format(
-        python=".".join(str(c) for c in sys.version_info[:2])
-    )
+"""
     assert expected in tester.io.fetch_output()
 
 
-def test_interactive_with_git_dependencies(tester, repo):
+def test_interactive_with_git_dependencies(
+    tester: CommandTester, repo: TestRepository
+) -> None:
     repo.add_package(get_package("pendulum", "2.0.0"))
     repo.add_package(get_package("pytest", "3.6.0"))
 
@@ -223,7 +303,56 @@ pytest = "^3.6.0"
     assert expected in tester.io.fetch_output()
 
 
-def test_interactive_with_git_dependencies_with_reference(tester, repo):
+_generate_choice_list_packages_params: list[list[Package]] = [
+    [
+        get_package("flask-blacklist", "1.0.0"),
+        get_package("Flask-Shelve", "1.0.0"),
+        get_package("flask-pwa", "1.0.0"),
+        get_package("Flask-test1", "1.0.0"),
+        get_package("Flask-test2", "1.0.0"),
+        get_package("Flask-test3", "1.0.0"),
+        get_package("Flask-test4", "1.0.0"),
+        get_package("Flask-test5", "1.0.0"),
+        get_package("Flask", "1.0.0"),
+        get_package("Flask-test6", "1.0.0"),
+        get_package("Flask-test7", "1.0.0"),
+    ],
+    [
+        get_package("flask-blacklist", "1.0.0"),
+        get_package("Flask-Shelve", "1.0.0"),
+        get_package("flask-pwa", "1.0.0"),
+        get_package("Flask-test1", "1.0.0"),
+        get_package("Flask", "1.0.0"),
+    ],
+]
+
+
+@pytest.fixture(params=_generate_choice_list_packages_params)
+def _generate_choice_list_packages(request: FixtureRequest) -> list[Package]:
+    packages: list[Package] = request.param
+    return packages
+
+
+@pytest.mark.parametrize("package_name", ["flask", "Flask", "flAsK"])
+def test_generate_choice_list(
+    tester: CommandTester,
+    package_name: str,
+    _generate_choice_list_packages: list[Package],
+) -> None:
+    init_command = tester.command
+    assert isinstance(init_command, InitCommand)
+
+    packages = _generate_choice_list_packages
+    choices = init_command._generate_choice_list(
+        packages, canonicalize_name(package_name)
+    )
+
+    assert choices[0] == "Flask"
+
+
+def test_interactive_with_git_dependencies_with_reference(
+    tester: CommandTester, repo: TestRepository
+) -> None:
     repo.add_package(get_package("pendulum", "2.0.0"))
     repo.add_package(get_package("pytest", "3.6.0"))
 
@@ -267,7 +396,9 @@ pytest = "^3.6.0"
     assert expected in tester.io.fetch_output()
 
 
-def test_interactive_with_git_dependencies_and_other_name(tester, repo):
+def test_interactive_with_git_dependencies_and_other_name(
+    tester: CommandTester, repo: TestRepository
+) -> None:
     repo.add_package(get_package("pendulum", "2.0.0"))
     repo.add_package(get_package("pytest", "3.6.0"))
 
@@ -311,7 +442,12 @@ pytest = "^3.6.0"
     assert expected in tester.io.fetch_output()
 
 
-def test_interactive_with_directory_dependency(tester, repo, source_dir, fixture_dir):
+def test_interactive_with_directory_dependency(
+    tester: CommandTester,
+    repo: TestRepository,
+    source_dir: Path,
+    fixture_dir: FixtureDirGetter,
+) -> None:
     repo.add_package(get_package("pendulum", "2.0.0"))
     repo.add_package(get_package("pytest", "3.6.0"))
 
@@ -358,8 +494,11 @@ pytest = "^3.6.0"
 
 
 def test_interactive_with_directory_dependency_and_other_name(
-    tester, repo, source_dir, fixture_dir
-):
+    tester: CommandTester,
+    repo: TestRepository,
+    source_dir: Path,
+    fixture_dir: FixtureDirGetter,
+) -> None:
     repo.add_package(get_package("pendulum", "2.0.0"))
     repo.add_package(get_package("pytest", "3.6.0"))
 
@@ -406,7 +545,12 @@ pytest = "^3.6.0"
     assert expected in tester.io.fetch_output()
 
 
-def test_interactive_with_file_dependency(tester, repo, source_dir, fixture_dir):
+def test_interactive_with_file_dependency(
+    tester: CommandTester,
+    repo: TestRepository,
+    source_dir: Path,
+    fixture_dir: FixtureDirGetter,
+) -> None:
     repo.add_package(get_package("pendulum", "2.0.0"))
     repo.add_package(get_package("pytest", "3.6.0"))
 
@@ -453,7 +597,53 @@ pytest = "^3.6.0"
     assert expected in tester.io.fetch_output()
 
 
-def test_python_option(tester):
+def test_interactive_with_wrong_dependency_inputs(
+    tester: CommandTester, repo: TestRepository
+) -> None:
+    inputs = [
+        "my-package",  # Package name
+        "1.2.3",  # Version
+        "This is a description",  # Description
+        "n",  # Author
+        "MIT",  # License
+        "^3.8",  # Python
+        "",  # Interactive packages
+        "foo 1.19.2",
+        "pendulum 2.0.0 foo",  # Package name and constraint (invalid)
+        "pendulum@^2.0.0",  # Package name and constraint (valid)
+        "",  # End package selection
+        "",  # Interactive dev packages
+        "pytest 3.6.0 foo",  # Dev package name and constraint (invalid)
+        "pytest 3.6.0",  # Dev package name and constraint (invalid)
+        "pytest@3.6.0",  # Dev package name and constraint (valid)
+        "",  # End package selection
+        "\n",  # Generate
+    ]
+    tester.execute(inputs="\n".join(inputs))
+
+    expected = """\
+[tool.poetry]
+name = "my-package"
+version = "1.2.3"
+description = "This is a description"
+authors = ["Your Name <you@example.com>"]
+license = "MIT"
+readme = "README.md"
+packages = [{include = "my_package"}]
+
+[tool.poetry.dependencies]
+python = "^3.8"
+foo = "1.19.2"
+pendulum = "^2.0.0"
+
+[tool.poetry.group.dev.dependencies]
+pytest = "3.6.0"
+"""
+
+    assert expected in tester.io.fetch_output()
+
+
+def test_python_option(tester: CommandTester) -> None:
     inputs = [
         "my-package",  # Package name
         "1.2.3",  # Version
@@ -483,7 +673,7 @@ python = "~2.7 || ^3.6"
     assert expected in tester.io.fetch_output()
 
 
-def test_predefined_dependency(tester, repo):
+def test_predefined_dependency(tester: CommandTester, repo: TestRepository) -> None:
     repo.add_package(get_package("pendulum", "2.0.0"))
 
     inputs = [
@@ -517,7 +707,9 @@ pendulum = "^2.0.0"
     assert expected in tester.io.fetch_output()
 
 
-def test_predefined_and_interactive_dependencies(tester, repo):
+def test_predefined_and_interactive_dependencies(
+    tester: CommandTester, repo: TestRepository
+) -> None:
     repo.add_package(get_package("pendulum", "2.0.0"))
     repo.add_package(get_package("pyramid", "1.10"))
 
@@ -558,7 +750,7 @@ python = "~2.7 || ^3.6"
     assert 'pyramid = "^1.10"' in output
 
 
-def test_predefined_dev_dependency(tester, repo):
+def test_predefined_dev_dependency(tester: CommandTester, repo: TestRepository) -> None:
     repo.add_package(get_package("pytest", "3.6.0"))
 
     inputs = [
@@ -595,7 +787,9 @@ pytest = "^3.6.0"
     assert expected in tester.io.fetch_output()
 
 
-def test_predefined_and_interactive_dev_dependencies(tester, repo):
+def test_predefined_and_interactive_dev_dependencies(
+    tester: CommandTester, repo: TestRepository
+) -> None:
     repo.add_package(get_package("pytest", "3.6.0"))
     repo.add_package(get_package("pytest-requests", "0.2.0"))
 
@@ -641,8 +835,57 @@ pytest-requests = "^0.2.0"
     assert 'pytest = "^3.6.0"' in output
 
 
-def test_add_package_with_extras_and_whitespace(tester):
-    result = tester.command._parse_requirements(["databases[postgresql, sqlite]"])
+def test_predefined_all_options(tester: CommandTester, repo: TestRepository) -> None:
+    repo.add_package(get_package("pendulum", "2.0.0"))
+    repo.add_package(get_package("pytest", "3.6.0"))
+
+    inputs = [
+        "1.2.3",  # Version
+        "",  # Author
+        "n",  # Interactive packages
+        "n",  # Interactive dev packages
+        "\n",  # Generate
+    ]
+
+    tester.execute(
+        (
+            "--name my-package "
+            "--description 'This is a description' "
+            "--author 'Foo Bar <foo@example.com>' "
+            "--python '^3.8' "
+            "--license MIT "
+            "--dependency pendulum "
+            "--dev-dependency pytest"
+        ),
+        inputs="\n".join(inputs),
+    )
+
+    expected = """\
+[tool.poetry]
+name = "my-package"
+version = "1.2.3"
+description = "This is a description"
+authors = ["Foo Bar <foo@example.com>"]
+license = "MIT"
+readme = "README.md"
+packages = [{include = "my_package"}]
+
+[tool.poetry.dependencies]
+python = "^3.8"
+pendulum = "^2.0.0"
+
+[tool.poetry.group.dev.dependencies]
+pytest = "^3.6.0"
+"""
+
+    output = tester.io.fetch_output()
+    assert expected in output
+
+
+def test_add_package_with_extras_and_whitespace(tester: CommandTester) -> None:
+    command = tester.command
+    assert isinstance(command, InitCommand)
+    result = command._parse_requirements(["databases[postgresql, sqlite]"])
 
     assert result[0]["name"] == "databases"
     assert len(result[0]["extras"]) == 2
@@ -651,37 +894,65 @@ def test_add_package_with_extras_and_whitespace(tester):
 
 
 def test_init_existing_pyproject_simple(
-    tester, source_dir, init_basic_inputs, init_basic_toml
-):
+    tester: CommandTester,
+    source_dir: Path,
+    init_basic_inputs: str,
+    init_basic_toml: str,
+) -> None:
     pyproject_file = source_dir / "pyproject.toml"
     existing_section = """
 [tool.black]
 line-length = 88
 """
-    pyproject_file.write_text(decode(existing_section))
+    pyproject_file.write_text(existing_section)
     tester.execute(inputs=init_basic_inputs)
-    assert (
-        "{}\n{}".format(existing_section, init_basic_toml) in pyproject_file.read_text()
-    )
+    assert f"{existing_section}\n{init_basic_toml}" in pyproject_file.read_text()
+
+
+@pytest.mark.parametrize("linesep", ["\n", "\r\n"])
+def test_init_existing_pyproject_consistent_linesep(
+    tester: CommandTester,
+    source_dir: Path,
+    init_basic_inputs: str,
+    init_basic_toml: str,
+    linesep: str,
+) -> None:
+    pyproject_file = source_dir / "pyproject.toml"
+    existing_section = """
+[tool.black]
+line-length = 88
+""".replace("\n", linesep)
+    with open(pyproject_file, "w", newline="") as f:
+        f.write(existing_section)
+    tester.execute(inputs=init_basic_inputs)
+    with open(pyproject_file, newline="") as f:
+        content = f.read()
+    init_basic_toml = init_basic_toml.replace("\n", linesep)
+    assert f"{existing_section}{linesep}{init_basic_toml}" in content
 
 
 def test_init_non_interactive_existing_pyproject_add_dependency(
-    tester, source_dir, init_basic_inputs, repo
-):
+    tester: CommandTester,
+    source_dir: Path,
+    init_basic_inputs: str,
+    repo: TestRepository,
+) -> None:
     pyproject_file = source_dir / "pyproject.toml"
     existing_section = """
 [tool.black]
 line-length = 88
 """
-    pyproject_file.write_text(decode(existing_section))
+    pyproject_file.write_text(existing_section)
 
     repo.add_package(get_package("foo", "1.19.2"))
 
     tester.execute(
-        "--author 'Your Name <you@example.com>' "
-        "--name 'my-package' "
-        "--python '^3.6' "
-        "--dependency foo",
+        (
+            "--author 'Your Name <you@example.com>' "
+            "--name 'my-package' "
+            "--python '^3.6' "
+            "--dependency foo"
+        ),
         interactive=False,
     )
 
@@ -698,22 +969,140 @@ packages = [{include = "my_package"}]
 python = "^3.6"
 foo = "^1.19.2"
 """
-    assert "{}\n{}".format(existing_section, expected) in pyproject_file.read_text()
+    assert f"{existing_section}\n{expected}" in pyproject_file.read_text()
 
 
 def test_init_existing_pyproject_with_build_system_fails(
-    tester, source_dir, init_basic_inputs
-):
+    tester: CommandTester, source_dir: Path, init_basic_inputs: str
+) -> None:
     pyproject_file = source_dir / "pyproject.toml"
     existing_section = """
 [build-system]
 requires = ["setuptools >= 40.6.0", "wheel"]
 build-backend = "setuptools.build_meta"
 """
-    pyproject_file.write_text(decode(existing_section))
+    pyproject_file.write_text(existing_section)
     tester.execute(inputs=init_basic_inputs)
     assert (
-        tester.io.fetch_output().strip()
+        tester.io.fetch_error().strip()
         == "A pyproject.toml file with a defined build-system already exists."
     )
-    assert "{}".format(existing_section) in pyproject_file.read_text()
+    assert existing_section in pyproject_file.read_text()
+
+
+@pytest.mark.parametrize(
+    "name",
+    [
+        None,
+        "",
+        "foo",
+        "   foo  ",
+        "foo==2.0",
+        "foo@2.0",
+        "  foo@2.0   ",
+        "foo 2.0",
+        "   foo 2.0  ",
+    ],
+)
+def test_validate_package_valid(name: str | None) -> None:
+    assert InitCommand._validate_package(name) == name
+
+
+@pytest.mark.parametrize(
+    "name", ["foo bar 2.0", "   foo bar 2.0   ", "foo bar foobar 2.0"]
+)
+def test_validate_package_invalid(name: str) -> None:
+    with pytest.raises(ValueError):
+        assert InitCommand._validate_package(name)
+
+
+@pytest.mark.parametrize(
+    "package_name, include",
+    (
+        ("mypackage", None),
+        ("my-package", "my_package"),
+        ("my.package", "my"),
+        ("my-awesome-package", "my_awesome_package"),
+        ("my.awesome.package", "my"),
+    ),
+)
+def test_package_include(
+    tester: CommandTester,
+    package_name: str,
+    include: str | None,
+) -> None:
+    tester.execute(
+        inputs="\n".join(
+            (
+                package_name,
+                "",  # Version
+                "",  # Description
+                "poetry",  # Author
+                "",  # License
+                "^3.10",  # Python
+                "n",  # Interactive packages
+                "n",  # Interactive dev packages
+                "\n",  # Generate
+            ),
+        ),
+    )
+
+    packages = "" if include is None else f'packages = [{{include = "{include}"}}]\n'
+
+    expected = (
+        f"[tool.poetry]\n"
+        f'name = "{package_name.replace(".", "-")}"\n'  # canonicalized
+        f'version = "0.1.0"\n'
+        f'description = ""\n'
+        f'authors = ["poetry"]\n'
+        f'readme = "README.md"\n'
+        f"{packages}"  # This line is optional. Thus no newline here.
+        f"\n"
+        f"[tool.poetry.dependencies]\n"
+        f'python = "^3.10"\n'
+    )
+    assert expected in tester.io.fetch_output()
+
+
+@pytest.mark.parametrize(
+    ["prefer_active", "python"],
+    [
+        (True, "1.1"),
+        (False, f"{sys.version_info[0]}.{sys.version_info[1]}"),
+    ],
+)
+def test_respect_prefer_active_on_init(
+    prefer_active: bool,
+    python: str,
+    config: Config,
+    mocker: MockerFixture,
+    tester: CommandTester,
+    source_dir: Path,
+) -> None:
+    from poetry.utils.env import GET_PYTHON_VERSION_ONELINER
+
+    orig_check_output = subprocess.check_output
+
+    def mock_check_output(cmd: str, *_: Any, **__: Any) -> str:
+        if GET_PYTHON_VERSION_ONELINER in cmd:
+            return "1.1.1"
+
+        result: str = orig_check_output(cmd, *_, **__)
+        return result
+
+    mocker.patch("subprocess.check_output", side_effect=mock_check_output)
+
+    config.config["virtualenvs"]["prefer-active-python"] = prefer_active
+    pyproject_file = source_dir / "pyproject.toml"
+
+    tester.execute(
+        "--author 'Your Name <you@example.com>' --name 'my-package'",
+        interactive=False,
+    )
+
+    expected = f"""\
+[tool.poetry.dependencies]
+python = "^{python}"
+"""
+
+    assert expected in pyproject_file.read_text()
