@@ -1,31 +1,35 @@
+from __future__ import annotations
+
 import itertools
 import json
 
 from pathlib import Path
+from typing import TYPE_CHECKING
+from typing import Any
 
 import pytest
 
+from cleo.io.buffered_io import BufferedIO
 from cleo.io.inputs.input import Input
-from cleo.io.io import IO
 from cleo.io.null_io import NullIO
-from cleo.io.outputs.buffered_output import BufferedOutput
 from cleo.io.outputs.output import Verbosity
-from deepdiff import DeepDiff
-
+from packaging.utils import canonicalize_name
+from poetry.core.packages.dependency_group import MAIN_GROUP
 from poetry.core.packages.dependency_group import DependencyGroup
 from poetry.core.packages.package import Package
 from poetry.core.packages.project_package import ProjectPackage
-from poetry.core.toml.file import TOMLFile
+
 from poetry.factory import Factory
-from poetry.installation import Installer as BaseInstaller
+from poetry.installation import Installer
 from poetry.installation.executor import Executor as BaseExecutor
-from poetry.installation.noop_installer import NoopInstaller
 from poetry.packages import Locker as BaseLocker
-from poetry.repositories import Pool
 from poetry.repositories import Repository
+from poetry.repositories import RepositoryPool
 from poetry.repositories.installed_repository import InstalledRepository
+from poetry.toml.file import TOMLFile
 from poetry.utils.env import MockEnv
 from poetry.utils.env import NullEnv
+from tests.helpers import MOCK_DEFAULT_GIT_REVISION
 from tests.helpers import get_dependency
 from tests.helpers import get_package
 from tests.repositories.test_legacy_repository import (
@@ -34,90 +38,99 @@ from tests.repositories.test_legacy_repository import (
 from tests.repositories.test_pypi_repository import MockRepository
 
 
+if TYPE_CHECKING:
+    from pytest_mock import MockerFixture
+
+    from poetry.installation.operations.operation import Operation
+    from poetry.utils.env import Env
+    from tests.conftest import Config
+    from tests.types import FixtureDirGetter
+
 RESERVED_PACKAGES = ("pip", "setuptools", "wheel")
 
 
-class Installer(BaseInstaller):
-    def _get_installer(self):
-        return NoopInstaller()
-
-
 class Executor(BaseExecutor):
-    def __init__(self, *args, **kwargs):
-        super(Executor, self).__init__(*args, **kwargs)
+    def __init__(self, *args: Any, **kwargs: Any) -> None:
+        super().__init__(*args, **kwargs)
 
-        self._installs = []
-        self._updates = []
-        self._uninstalls = []
+        self._installs: list[Package] = []
+        self._updates: list[Package] = []
+        self._uninstalls: list[Package] = []
 
     @property
-    def installations(self):
+    def installations(self) -> list[Package]:
         return self._installs
 
     @property
-    def updates(self):
+    def updates(self) -> list[Package]:
         return self._updates
 
     @property
-    def removals(self):
+    def removals(self) -> list[Package]:
         return self._uninstalls
 
-    def _do_execute_operation(self, operation):
-        super(Executor, self)._do_execute_operation(operation)
+    def _do_execute_operation(self, operation: Operation) -> int:
+        ret_val = super()._do_execute_operation(operation)
 
         if not operation.skipped:
-            getattr(self, "_{}s".format(operation.job_type)).append(operation.package)
+            getattr(self, f"_{operation.job_type}s").append(operation.package)
 
-    def _execute_install(self, operation):
+        return ret_val
+
+    def _execute_install(self, operation: Operation) -> int:
         return 0
 
-    def _execute_update(self, operation):
+    def _execute_update(self, operation: Operation) -> int:
         return 0
 
-    def _execute_uninstall(self, operation):
+    def _execute_uninstall(self, operation: Operation) -> int:
         return 0
 
 
 class CustomInstalledRepository(InstalledRepository):
     @classmethod
-    def load(cls, env):
+    def load(
+        cls, env: Env, with_dependencies: bool = False
+    ) -> CustomInstalledRepository:
         return cls()
 
 
 class Locker(BaseLocker):
-    def __init__(self):
-        self._lock = TOMLFile(Path.cwd().joinpath("poetry.lock"))
+    def __init__(self, lock_path: Path) -> None:
+        self._lock = lock_path / "poetry.lock"
         self._written_data = None
         self._locked = False
+        self._lock_data = None
         self._content_hash = self._get_content_hash()
 
     @property
-    def written_data(self):
+    def written_data(self) -> dict[str, Any]:
+        assert self._written_data is not None
         return self._written_data
 
-    def set_lock_path(self, lock):
-        self._lock = TOMLFile(Path(lock).joinpath("poetry.lock"))
+    def set_lock_path(self, lock: Path) -> Locker:
+        self._lock = lock / "poetry.lock"
 
         return self
 
-    def locked(self, is_locked=True):
+    def locked(self, is_locked: bool = True) -> Locker:
         self._locked = is_locked
 
         return self
 
-    def mock_lock_data(self, data):
+    def mock_lock_data(self, data: dict[str, Any]) -> None:
         self._lock_data = data
 
-    def is_locked(self):
+    def is_locked(self) -> bool:
         return self._locked
 
-    def is_fresh(self):
+    def is_fresh(self) -> bool:
         return True
 
-    def _get_content_hash(self):
+    def _get_content_hash(self) -> str:
         return "123456789"
 
-    def _write_lock_data(self, data):
+    def _write_lock_data(self, data: dict[str, Any]) -> None:
         for package in data["package"]:
             python_versions = str(package["python-versions"])
             package["python-versions"] = python_versions
@@ -127,7 +140,7 @@ class Locker(BaseLocker):
 
 
 @pytest.fixture()
-def package():
+def package() -> ProjectPackage:
     p = ProjectPackage("root", "1.0")
     p.root_dir = Path.cwd()
 
@@ -135,36 +148,43 @@ def package():
 
 
 @pytest.fixture()
-def repo():
-    return Repository()
+def repo() -> Repository:
+    return Repository("repo")
 
 
 @pytest.fixture()
-def pool(repo):
-    pool = Pool()
+def pool(repo: Repository) -> RepositoryPool:
+    pool = RepositoryPool()
     pool.add_repository(repo)
 
     return pool
 
 
 @pytest.fixture()
-def installed():
+def installed() -> CustomInstalledRepository:
     return CustomInstalledRepository()
 
 
 @pytest.fixture()
-def locker():
-    return Locker()
+def locker(project_root: Path) -> Locker:
+    return Locker(lock_path=project_root)
 
 
 @pytest.fixture()
-def env():
-    return NullEnv()
+def env(tmp_path: Path) -> NullEnv:
+    return NullEnv(path=tmp_path)
 
 
 @pytest.fixture()
-def installer(package, pool, locker, env, installed, config):
-    installer = Installer(
+def installer(
+    package: ProjectPackage,
+    pool: RepositoryPool,
+    locker: Locker,
+    env: NullEnv,
+    installed: CustomInstalledRepository,
+    config: Config,
+) -> Installer:
+    return Installer(
         NullIO(),
         env,
         package,
@@ -174,25 +194,26 @@ def installer(package, pool, locker, env, installed, config):
         installed=installed,
         executor=Executor(env, pool, config, NullIO()),
     )
-    installer.use_executor(True)
-
-    return installer
 
 
-def fixture(name):
-    file = TOMLFile(Path(__file__).parent / "fixtures" / "{}.test".format(name))
+def fixture(name: str) -> dict[str, Any]:
+    file = TOMLFile(Path(__file__).parent / "fixtures" / f"{name}.test")
+    content: dict[str, Any] = file.read()
 
-    return json.loads(json.dumps(file.read()))
+    return content
 
 
-def test_run_no_dependencies(installer, locker):
-    installer.run()
+def test_run_no_dependencies(installer: Installer, locker: Locker) -> None:
+    result = installer.run()
+    assert result == 0
+
     expected = fixture("no-dependencies")
-
     assert locker.written_data == expected
 
 
-def test_run_with_dependencies(installer, locker, repo, package):
+def test_run_with_dependencies(
+    installer: Installer, locker: Locker, repo: Repository, package: ProjectPackage
+) -> None:
     package_a = get_package("A", "1.0")
     package_b = get_package("B", "1.1")
     repo.add_package(package_a)
@@ -201,15 +222,20 @@ def test_run_with_dependencies(installer, locker, repo, package):
     package.add_dependency(Factory.create_dependency("A", "~1.0"))
     package.add_dependency(Factory.create_dependency("B", "^1.0"))
 
-    installer.run()
-    expected = fixture("with-dependencies")
+    result = installer.run()
+    assert result == 0
 
+    expected = fixture("with-dependencies")
     assert locker.written_data == expected
 
 
 def test_run_update_after_removing_dependencies(
-    installer, locker, repo, package, installed
-):
+    installer: Installer,
+    locker: Locker,
+    repo: Repository,
+    package: ProjectPackage,
+    installed: CustomInstalledRepository,
+) -> None:
     locker.locked(True)
     locker.mock_lock_data(
         {
@@ -217,7 +243,6 @@ def test_run_update_after_removing_dependencies(
                 {
                     "name": "A",
                     "version": "1.0",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -226,7 +251,6 @@ def test_run_update_after_removing_dependencies(
                 {
                     "name": "B",
                     "version": "1.1",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -235,7 +259,6 @@ def test_run_update_after_removing_dependencies(
                 {
                     "name": "C",
                     "version": "1.2",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -246,7 +269,7 @@ def test_run_update_after_removing_dependencies(
                 "python-versions": "*",
                 "platform": "*",
                 "content-hash": "123456789",
-                "hashes": {"A": [], "B": [], "C": []},
+                "files": {"A": [], "B": [], "C": []},
             },
         }
     )
@@ -265,19 +288,25 @@ def test_run_update_after_removing_dependencies(
     package.add_dependency(Factory.create_dependency("B", "~1.1"))
 
     installer.update(True)
-    installer.run()
-    expected = fixture("with-dependencies")
+    result = installer.run()
+    assert result == 0
 
+    expected = fixture("with-dependencies")
     assert locker.written_data == expected
 
-    assert 0 == installer.executor.installations_count
-    assert 0 == installer.executor.updates_count
-    assert 1 == installer.executor.removals_count
+    assert installer.executor.installations_count == 0
+    assert installer.executor.updates_count == 0
+    assert installer.executor.removals_count == 1
 
 
 def _configure_run_install_dev(
-    locker, repo, package, installed, with_optional_group=False
-):
+    locker: Locker,
+    repo: Repository,
+    package: ProjectPackage,
+    installed: CustomInstalledRepository,
+    with_optional_group: bool = False,
+    with_packages_installed: bool = False,
+) -> None:
     """
     Perform common test setup for `test_run_install_*dev*()` methods.
     """
@@ -288,7 +317,6 @@ def _configure_run_install_dev(
                 {
                     "name": "A",
                     "version": "1.0",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -297,7 +325,6 @@ def _configure_run_install_dev(
                 {
                     "name": "B",
                     "version": "1.1",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -306,7 +333,6 @@ def _configure_run_install_dev(
                 {
                     "name": "C",
                     "version": "1.2",
-                    "category": "dev",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -317,7 +343,7 @@ def _configure_run_install_dev(
                 "python-versions": "*",
                 "platform": "*",
                 "content-hash": "123456789",
-                "hashes": {"A": [], "B": [], "C": []},
+                "files": {"A": [], "B": [], "C": []},
             },
         }
     )
@@ -328,9 +354,10 @@ def _configure_run_install_dev(
     repo.add_package(package_b)
     repo.add_package(package_c)
 
-    installed.add_package(package_a)
-    installed.add_package(package_b)
-    installed.add_package(package_c)
+    if with_packages_installed:
+        installed.add_package(package_a)
+        installed.add_package(package_b)
+        installed.add_package(package_c)
 
     package.add_dependency(Factory.create_dependency("A", "~1.0"))
     package.add_dependency(Factory.create_dependency("B", "~1.1"))
@@ -340,45 +367,61 @@ def _configure_run_install_dev(
     package.add_dependency_group(group)
 
 
-def test_run_install_no_group(installer, locker, repo, package, installed):
-    _configure_run_install_dev(locker, repo, package, installed)
-
-    installer.without_groups(["dev"])
-    installer.run()
-
-    assert 0 == installer.executor.installations_count
-    assert 0 == installer.executor.updates_count
-    assert 0 == installer.executor.removals_count
-
-
-def test_run_install_group_only(installer, locker, repo, package, installed):
-    _configure_run_install_dev(locker, repo, package, installed)
-
-    installer.only_groups(["dev"])
-    installer.run()
-
-    assert 0 == installer.executor.installations_count
-    assert 0 == installer.executor.updates_count
-    assert 0 == installer.executor.removals_count
-
-
-def test_run_install_with_optional_group_not_selected(
-    installer, locker, repo, package, installed
-):
+@pytest.mark.parametrize(
+    ("groups", "installs", "updates", "removals", "with_packages_installed"),
+    [
+        (None, 2, 0, 0, False),
+        (None, 0, 0, 1, True),
+        ([], 0, 0, 0, False),
+        ([], 0, 0, 3, True),
+        (["dev"], 1, 0, 0, False),
+        (["dev"], 0, 0, 2, True),
+        ([MAIN_GROUP], 2, 0, 0, False),
+        ([MAIN_GROUP], 0, 0, 1, True),
+        ([MAIN_GROUP, "dev"], 3, 0, 0, False),
+        ([MAIN_GROUP, "dev"], 0, 0, 0, True),
+    ],
+)
+def test_run_install_with_dependency_groups(
+    groups: list[str] | None,
+    installs: int,
+    updates: int,
+    removals: int,
+    with_packages_installed: bool,
+    installer: Installer,
+    locker: Locker,
+    repo: Repository,
+    package: ProjectPackage,
+    installed: CustomInstalledRepository,
+) -> None:
     _configure_run_install_dev(
-        locker, repo, package, installed, with_optional_group=True
+        locker,
+        repo,
+        package,
+        installed,
+        with_optional_group=True,
+        with_packages_installed=with_packages_installed,
     )
 
-    installer.run()
+    if groups is not None:
+        installer.only_groups(groups)
 
-    assert 0 == installer.executor.installations_count
-    assert 0 == installer.executor.updates_count
-    assert 0 == installer.executor.removals_count
+    installer.requires_synchronization(True)
+    result = installer.run()
+    assert result == 0
+
+    assert installer.executor.installations_count == installs
+    assert installer.executor.updates_count == updates
+    assert installer.executor.removals_count == removals
 
 
 def test_run_install_does_not_remove_locked_packages_if_installed_but_not_required(
-    installer, locker, repo, package, installed
-):
+    installer: Installer,
+    locker: Locker,
+    repo: Repository,
+    package: ProjectPackage,
+    installed: CustomInstalledRepository,
+) -> None:
     package_a = get_package("a", "1.0")
     package_b = get_package("b", "1.1")
     package_c = get_package("c", "1.2")
@@ -392,7 +435,9 @@ def test_run_install_does_not_remove_locked_packages_if_installed_but_not_requir
 
     installed.add_package(package)  # Root package never removed.
 
-    package.add_dependency(Factory.create_dependency(package_a.name, package_a.version))
+    package.add_dependency(
+        Factory.create_dependency(package_a.name, str(package_a.version))
+    )
 
     locker.locked(True)
     locker.mock_lock_data(
@@ -401,7 +446,6 @@ def test_run_install_does_not_remove_locked_packages_if_installed_but_not_requir
                 {
                     "name": package_a.name,
                     "version": package_a.version.text,
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -410,7 +454,6 @@ def test_run_install_does_not_remove_locked_packages_if_installed_but_not_requir
                 {
                     "name": package_b.name,
                     "version": package_b.version.text,
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -419,7 +462,6 @@ def test_run_install_does_not_remove_locked_packages_if_installed_but_not_requir
                 {
                     "name": package_c.name,
                     "version": package_c.version.text,
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -430,21 +472,26 @@ def test_run_install_does_not_remove_locked_packages_if_installed_but_not_requir
                 "python-versions": "*",
                 "platform": "*",
                 "content-hash": "123456789",
-                "hashes": {package_a.name: [], package_b.name: [], package_c.name: []},
+                "files": {package_a.name: [], package_b.name: [], package_c.name: []},
             },
         }
     )
 
-    installer.run()
+    result = installer.run()
+    assert result == 0
 
-    assert 0 == installer.executor.installations_count
-    assert 0 == installer.executor.updates_count
-    assert 0 == installer.executor.removals_count
+    assert installer.executor.installations_count == 0
+    assert installer.executor.updates_count == 0
+    assert installer.executor.removals_count == 0
 
 
-def test_run_install_removes_locked_packages_if_installed_and_synchronization_is_required(
-    installer, locker, repo, package, installed
-):
+def test_run_install_removes_locked_packages_if_installed_and_synchronization_is_required(  # noqa: E501
+    installer: Installer,
+    locker: Locker,
+    repo: Repository,
+    package: ProjectPackage,
+    installed: CustomInstalledRepository,
+) -> None:
     package_a = get_package("a", "1.0")
     package_b = get_package("b", "1.1")
     package_c = get_package("c", "1.2")
@@ -458,7 +505,9 @@ def test_run_install_removes_locked_packages_if_installed_and_synchronization_is
 
     installed.add_package(package)  # Root package never removed.
 
-    package.add_dependency(Factory.create_dependency(package_a.name, package_a.version))
+    package.add_dependency(
+        Factory.create_dependency(package_a.name, str(package_a.version))
+    )
 
     locker.locked(True)
     locker.mock_lock_data(
@@ -467,7 +516,6 @@ def test_run_install_removes_locked_packages_if_installed_and_synchronization_is
                 {
                     "name": package_a.name,
                     "version": package_a.version.text,
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -476,7 +524,6 @@ def test_run_install_removes_locked_packages_if_installed_and_synchronization_is
                 {
                     "name": package_b.name,
                     "version": package_b.version.text,
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -485,7 +532,6 @@ def test_run_install_removes_locked_packages_if_installed_and_synchronization_is
                 {
                     "name": package_c.name,
                     "version": package_c.version.text,
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -496,7 +542,7 @@ def test_run_install_removes_locked_packages_if_installed_and_synchronization_is
                 "python-versions": "*",
                 "platform": "*",
                 "content-hash": "123456789",
-                "hashes": {package_a.name: [], package_b.name: [], package_c.name: []},
+                "files": {package_a.name: [], package_b.name: [], package_c.name: []},
             },
         }
     )
@@ -504,14 +550,18 @@ def test_run_install_removes_locked_packages_if_installed_and_synchronization_is
     installer.requires_synchronization(True)
     installer.run()
 
-    assert 0 == installer.executor.installations_count
-    assert 0 == installer.executor.updates_count
-    assert 2 == installer.executor.removals_count
+    assert installer.executor.installations_count == 0
+    assert installer.executor.updates_count == 0
+    assert installer.executor.removals_count == 2
 
 
 def test_run_install_removes_no_longer_locked_packages_if_installed(
-    installer, locker, repo, package, installed
-):
+    installer: Installer,
+    locker: Locker,
+    repo: Repository,
+    package: ProjectPackage,
+    installed: CustomInstalledRepository,
+) -> None:
     package_a = get_package("a", "1.0")
     package_b = get_package("b", "1.1")
     package_c = get_package("c", "1.2")
@@ -525,7 +575,9 @@ def test_run_install_removes_no_longer_locked_packages_if_installed(
 
     installed.add_package(package)  # Root package never removed.
 
-    package.add_dependency(Factory.create_dependency(package_a.name, package_a.version))
+    package.add_dependency(
+        Factory.create_dependency(package_a.name, str(package_a.version))
+    )
 
     locker.locked(True)
     locker.mock_lock_data(
@@ -534,7 +586,6 @@ def test_run_install_removes_no_longer_locked_packages_if_installed(
                 {
                     "name": package_a.name,
                     "version": package_a.version.text,
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -543,7 +594,6 @@ def test_run_install_removes_no_longer_locked_packages_if_installed(
                 {
                     "name": package_b.name,
                     "version": package_b.version.text,
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -552,7 +602,6 @@ def test_run_install_removes_no_longer_locked_packages_if_installed(
                 {
                     "name": package_c.name,
                     "version": package_c.version.text,
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -563,49 +612,37 @@ def test_run_install_removes_no_longer_locked_packages_if_installed(
                 "python-versions": "*",
                 "platform": "*",
                 "content-hash": "123456789",
-                "hashes": {package_a.name: [], package_b.name: [], package_c.name: []},
+                "files": {package_a.name: [], package_b.name: [], package_c.name: []},
             },
         }
     )
 
     installer.update(True)
-    installer.run()
+    result = installer.run()
+    assert result == 0
 
-    assert 0 == installer.executor.installations_count
-    assert 0 == installer.executor.updates_count
-    assert 2 == installer.executor.removals_count
-
-
-def test_run_install_with_optional_group_selected(
-    installer, locker, repo, package, installed
-):
-    _configure_run_install_dev(
-        locker, repo, package, installed, with_optional_group=True
-    )
-
-    installer.with_groups(["dev"])
-    installer.run()
-
-    assert 0 == installer.executor.installations_count
-    assert 0 == installer.executor.updates_count
-    assert 0 == installer.executor.removals_count
+    assert installer.executor.installations_count == 0
+    assert installer.executor.updates_count == 0
+    assert installer.executor.removals_count == 2
 
 
 @pytest.mark.parametrize(
     "managed_reserved_package_names",
-    [
-        i
-        for i in itertools.chain(
-            [tuple()],
-            itertools.permutations(RESERVED_PACKAGES, 1),
-            itertools.permutations(RESERVED_PACKAGES, 2),
-            [RESERVED_PACKAGES],
-        )
-    ],
+    itertools.chain(
+        [()],
+        itertools.permutations(RESERVED_PACKAGES, 1),
+        itertools.permutations(RESERVED_PACKAGES, 2),
+        [RESERVED_PACKAGES],
+    ),
 )
 def test_run_install_with_synchronization(
-    managed_reserved_package_names, installer, locker, repo, package, installed
-):
+    managed_reserved_package_names: tuple[str, ...],
+    installer: Installer,
+    locker: Locker,
+    repo: Repository,
+    package: ProjectPackage,
+    installed: CustomInstalledRepository,
+) -> None:
     package_a = get_package("a", "1.0")
     package_b = get_package("b", "1.1")
     package_c = get_package("c", "1.2")
@@ -633,7 +670,9 @@ def test_run_install_with_synchronization(
 
     installed.add_package(package)  # Root package never removed.
 
-    package.add_dependency(Factory.create_dependency(package_a.name, package_a.version))
+    package.add_dependency(
+        Factory.create_dependency(package_a.name, str(package_a.version))
+    )
 
     locker.locked(True)
     locker.mock_lock_data(
@@ -642,7 +681,6 @@ def test_run_install_with_synchronization(
                 {
                     "name": pkg.name,
                     "version": pkg.version,
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -654,16 +692,17 @@ def test_run_install_with_synchronization(
                 "python-versions": "*",
                 "platform": "*",
                 "content-hash": "123456789",
-                "hashes": {pkg.name: [] for pkg in locked_packages},
+                "files": {pkg.name: [] for pkg in locked_packages},
             },
         }
     )
 
     installer.requires_synchronization(True)
-    installer.run()
+    result = installer.run()
+    assert result == 0
 
-    assert 0 == installer.executor.installations_count
-    assert 0 == installer.executor.updates_count
+    assert installer.executor.installations_count == 0
+    assert installer.executor.updates_count == 0
     assert 2 + len(managed_reserved_packages) == installer.executor.removals_count
 
     expected_removals = {
@@ -672,10 +711,13 @@ def test_run_install_with_synchronization(
         *managed_reserved_package_names,
     }
 
-    assert expected_removals == set(r.name for r in installer.executor.removals)
+    assert isinstance(installer.executor, Executor)
+    assert {r.name for r in installer.executor.removals} == expected_removals
 
 
-def test_run_whitelist_add(installer, locker, repo, package):
+def test_run_whitelist_add(
+    installer: Installer, locker: Locker, repo: Repository, package: ProjectPackage
+) -> None:
     locker.locked(True)
     locker.mock_lock_data(
         {
@@ -683,7 +725,6 @@ def test_run_whitelist_add(installer, locker, repo, package):
                 {
                     "name": "A",
                     "version": "1.0",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -694,7 +735,7 @@ def test_run_whitelist_add(installer, locker, repo, package):
                 "python-versions": "*",
                 "platform": "*",
                 "content-hash": "123456789",
-                "hashes": {"A": []},
+                "files": {"A": []},
             },
         }
     )
@@ -711,13 +752,20 @@ def test_run_whitelist_add(installer, locker, repo, package):
     installer.update(True)
     installer.whitelist(["B"])
 
-    installer.run()
-    expected = fixture("with-dependencies")
+    result = installer.run()
+    assert result == 0
 
+    expected = fixture("with-dependencies")
     assert locker.written_data == expected
 
 
-def test_run_whitelist_remove(installer, locker, repo, package, installed):
+def test_run_whitelist_remove(
+    installer: Installer,
+    locker: Locker,
+    repo: Repository,
+    package: ProjectPackage,
+    installed: CustomInstalledRepository,
+) -> None:
     locker.locked(True)
     locker.mock_lock_data(
         {
@@ -725,7 +773,6 @@ def test_run_whitelist_remove(installer, locker, repo, package, installed):
                 {
                     "name": "A",
                     "version": "1.0",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -734,7 +781,6 @@ def test_run_whitelist_remove(installer, locker, repo, package, installed):
                 {
                     "name": "B",
                     "version": "1.1",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -745,7 +791,7 @@ def test_run_whitelist_remove(installer, locker, repo, package, installed):
                 "python-versions": "*",
                 "platform": "*",
                 "content-hash": "123456789",
-                "hashes": {"A": [], "B": []},
+                "files": {"A": [], "B": []},
             },
         }
     )
@@ -760,16 +806,19 @@ def test_run_whitelist_remove(installer, locker, repo, package, installed):
     installer.update(True)
     installer.whitelist(["B"])
 
-    installer.run()
+    result = installer.run()
+    assert result == 0
+
     expected = fixture("remove")
-
     assert locker.written_data == expected
-    assert 1 == installer.executor.installations_count
-    assert 0 == installer.executor.updates_count
-    assert 1 == installer.executor.removals_count
+    assert installer.executor.installations_count == 1
+    assert installer.executor.updates_count == 0
+    assert installer.executor.removals_count == 1
 
 
-def test_add_with_sub_dependencies(installer, locker, repo, package):
+def test_add_with_sub_dependencies(
+    installer: Installer, locker: Locker, repo: Repository, package: ProjectPackage
+) -> None:
     package_a = get_package("A", "1.0")
     package_b = get_package("B", "1.1")
     package_c = get_package("C", "1.2")
@@ -785,13 +834,16 @@ def test_add_with_sub_dependencies(installer, locker, repo, package):
     package_a.add_dependency(Factory.create_dependency("D", "^1.0"))
     package_b.add_dependency(Factory.create_dependency("C", "~1.2"))
 
-    installer.run()
-    expected = fixture("with-sub-dependencies")
+    result = installer.run()
+    assert result == 0
 
+    expected = fixture("with-sub-dependencies")
     assert locker.written_data == expected
 
 
-def test_run_with_python_versions(installer, locker, repo, package):
+def test_run_with_python_versions(
+    installer: Installer, locker: Locker, repo: Repository, package: ProjectPackage
+) -> None:
     package.python_versions = "~2.7 || ^3.4"
 
     package_a = get_package("A", "1.0")
@@ -810,15 +862,16 @@ def test_run_with_python_versions(installer, locker, repo, package):
     package.add_dependency(Factory.create_dependency("B", "^1.0"))
     package.add_dependency(Factory.create_dependency("C", "^1.0"))
 
-    installer.run()
-    expected = fixture("with-python-versions")
+    result = installer.run()
+    assert result == 0
 
+    expected = fixture("with-python-versions")
     assert locker.written_data == expected
 
 
 def test_run_with_optional_and_python_restricted_dependencies(
-    installer, locker, repo, package
-):
+    installer: Installer, locker: Locker, repo: Repository, package: ProjectPackage
+) -> None:
     package.python_versions = "~2.7 || ^3.4"
 
     package_a = get_package("A", "1.0")
@@ -834,7 +887,7 @@ def test_run_with_optional_and_python_restricted_dependencies(
     repo.add_package(package_c13)
     repo.add_package(package_d)
 
-    package.extras = {"foo": [get_dependency("A", "~1.0")]}
+    package.extras = {canonicalize_name("foo"): [get_dependency("A", "~1.0")]}
     package.add_dependency(
         Factory.create_dependency("A", {"version": "~1.0", "optional": True})
     )
@@ -845,22 +898,28 @@ def test_run_with_optional_and_python_restricted_dependencies(
         Factory.create_dependency("C", {"version": "^1.0", "python": "~2.7 || ^3.4"})
     )
 
-    installer.run()
-    expected = fixture("with-optional-dependencies")
+    result = installer.run()
+    assert result == 0
 
+    expected = fixture("with-optional-dependencies")
     assert locker.written_data == expected
 
     # We should only have 2 installs:
     # C,D since python version is not compatible
     # with B's python constraint and A is optional
-    assert 2 == installer.executor.installations_count
-    assert "d" == installer.executor.installations[0].name
-    assert "c" == installer.executor.installations[1].name
+    assert isinstance(installer.executor, Executor)
+    assert installer.executor.installations_count == 2
+    assert installer.executor.installations[0].name == "d"
+    assert installer.executor.installations[1].name == "c"
 
 
 def test_run_with_optional_and_platform_restricted_dependencies(
-    installer, locker, repo, package, mocker
-):
+    installer: Installer,
+    locker: Locker,
+    repo: Repository,
+    package: ProjectPackage,
+    mocker: MockerFixture,
+) -> None:
     mocker.patch("sys.platform", "darwin")
 
     package_a = get_package("A", "1.0")
@@ -876,7 +935,7 @@ def test_run_with_optional_and_platform_restricted_dependencies(
     repo.add_package(package_c13)
     repo.add_package(package_d)
 
-    package.extras = {"foo": [get_dependency("A", "~1.0")]}
+    package.extras = {canonicalize_name("foo"): [get_dependency("A", "~1.0")]}
     package.add_dependency(
         Factory.create_dependency("A", {"version": "~1.0", "optional": True})
     )
@@ -887,25 +946,29 @@ def test_run_with_optional_and_platform_restricted_dependencies(
         Factory.create_dependency("C", {"version": "^1.0", "platform": "darwin"})
     )
 
-    installer.run()
-    expected = fixture("with-platform-dependencies")
+    result = installer.run()
+    assert result == 0
 
+    expected = fixture("with-platform-dependencies")
     assert locker.written_data == expected
 
     # We should only have 2 installs:
     # C,D since the mocked python version is not compatible
     # with B's python constraint and A is optional
-    assert 2 == installer.executor.installations_count
-    assert "d" == installer.executor.installations[0].name
-    assert "c" == installer.executor.installations[1].name
+    assert isinstance(installer.executor, Executor)
+    assert installer.executor.installations_count == 2
+    assert installer.executor.installations[0].name == "d"
+    assert installer.executor.installations[1].name == "c"
 
 
-def test_run_with_dependencies_extras(installer, locker, repo, package):
+def test_run_with_dependencies_extras(
+    installer: Installer, locker: Locker, repo: Repository, package: ProjectPackage
+) -> None:
     package_a = get_package("A", "1.0")
     package_b = get_package("B", "1.0")
     package_c = get_package("C", "1.0")
 
-    package_b.extras = {"foo": [get_dependency("C", "^1.0")]}
+    package_b.extras = {canonicalize_name("foo"): [get_dependency("C", "^1.0")]}
     package_b.add_dependency(
         Factory.create_dependency("C", {"version": "^1.0", "optional": True})
     )
@@ -919,13 +982,16 @@ def test_run_with_dependencies_extras(installer, locker, repo, package):
         Factory.create_dependency("B", {"version": "^1.0", "extras": ["foo"]})
     )
 
-    installer.run()
-    expected = fixture("with-dependencies-extras")
+    result = installer.run()
+    assert result == 0
 
+    expected = fixture("with-dependencies-extras")
     assert locker.written_data == expected
 
 
-def test_run_with_dependencies_nested_extras(installer, locker, repo, package):
+def test_run_with_dependencies_nested_extras(
+    installer: Installer, locker: Locker, repo: Repository, package: ProjectPackage
+) -> None:
     package_a = get_package("A", "1.0")
     package_b = get_package("B", "1.0")
     package_c = get_package("C", "1.0")
@@ -936,11 +1002,11 @@ def test_run_with_dependencies_nested_extras(installer, locker, repo, package):
     )
     dependency_a = Factory.create_dependency("A", {"version": "^1.0", "extras": ["B"]})
 
-    package_b.extras = {"C": [dependency_c]}
+    package_b.extras = {canonicalize_name("c"): [dependency_c]}
     package_b.add_dependency(dependency_c)
 
     package_a.add_dependency(dependency_b)
-    package_a.extras = {"B": [dependency_b]}
+    package_a.extras = {canonicalize_name("b"): [dependency_b]}
 
     repo.add_package(package_a)
     repo.add_package(package_b)
@@ -948,14 +1014,17 @@ def test_run_with_dependencies_nested_extras(installer, locker, repo, package):
 
     package.add_dependency(dependency_a)
 
-    installer.run()
-    expected = fixture("with-dependencies-nested-extras")
+    result = installer.run()
+    assert result == 0
 
+    expected = fixture("with-dependencies-nested-extras")
     assert locker.written_data == expected
 
 
-def test_run_does_not_install_extras_if_not_requested(installer, locker, repo, package):
-    package.extras["foo"] = [get_dependency("D")]
+def test_run_does_not_install_extras_if_not_requested(
+    installer: Installer, locker: Locker, repo: Repository, package: ProjectPackage
+) -> None:
+    package.extras[canonicalize_name("foo")] = [get_dependency("D")]
     package_a = get_package("A", "1.0")
     package_b = get_package("B", "1.0")
     package_c = get_package("C", "1.0")
@@ -973,18 +1042,21 @@ def test_run_does_not_install_extras_if_not_requested(installer, locker, repo, p
         Factory.create_dependency("D", {"version": "^1.0", "optional": True})
     )
 
-    installer.run()
-    expected = fixture("extras")
+    result = installer.run()
+    assert result == 0
 
+    expected = fixture("extras")
     # Extras are pinned in lock
     assert locker.written_data == expected
 
     # But should not be installed
-    assert 3 == installer.executor.installations_count  # A, B, C
+    assert installer.executor.installations_count == 3  # A, B, C
 
 
-def test_run_installs_extras_if_requested(installer, locker, repo, package):
-    package.extras["foo"] = [get_dependency("D")]
+def test_run_installs_extras_if_requested(
+    installer: Installer, locker: Locker, repo: Repository, package: ProjectPackage
+) -> None:
+    package.extras[canonicalize_name("foo")] = [get_dependency("D")]
     package_a = get_package("A", "1.0")
     package_b = get_package("B", "1.0")
     package_c = get_package("C", "1.0")
@@ -1003,18 +1075,21 @@ def test_run_installs_extras_if_requested(installer, locker, repo, package):
     )
 
     installer.extras(["foo"])
-    installer.run()
-    expected = fixture("extras")
+    result = installer.run()
+    assert result == 0
 
     # Extras are pinned in lock
+    expected = fixture("extras")
     assert locker.written_data == expected
 
     # But should not be installed
-    assert 4 == installer.executor.installations_count  # A, B, C, D
+    assert installer.executor.installations_count == 4  # A, B, C, D
 
 
-def test_run_installs_extras_with_deps_if_requested(installer, locker, repo, package):
-    package.extras["foo"] = [get_dependency("C")]
+def test_run_installs_extras_with_deps_if_requested(
+    installer: Installer, locker: Locker, repo: Repository, package: ProjectPackage
+) -> None:
+    package.extras[canonicalize_name("foo")] = [get_dependency("C")]
     package_a = get_package("A", "1.0")
     package_b = get_package("B", "1.0")
     package_c = get_package("C", "1.0")
@@ -1034,22 +1109,24 @@ def test_run_installs_extras_with_deps_if_requested(installer, locker, repo, pac
     package_c.add_dependency(Factory.create_dependency("D", "^1.0"))
 
     installer.extras(["foo"])
-    installer.run()
+    result = installer.run()
+    assert result == 0
+
     expected = fixture("extras-with-dependencies")
 
     # Extras are pinned in lock
     assert locker.written_data == expected
 
     # But should not be installed
-    assert 4 == installer.executor.installations_count  # A, B, C, D
+    assert installer.executor.installations_count == 4  # A, B, C, D
 
 
 def test_run_installs_extras_with_deps_if_requested_locked(
-    installer, locker, repo, package
-):
+    installer: Installer, locker: Locker, repo: Repository, package: ProjectPackage
+) -> None:
     locker.locked(True)
     locker.mock_lock_data(fixture("extras-with-dependencies"))
-    package.extras["foo"] = [get_dependency("C")]
+    package.extras[canonicalize_name("foo")] = [get_dependency("C")]
     package_a = get_package("A", "1.0")
     package_b = get_package("B", "1.0")
     package_c = get_package("C", "1.0")
@@ -1069,81 +1146,136 @@ def test_run_installs_extras_with_deps_if_requested_locked(
     package_c.add_dependency(Factory.create_dependency("D", "^1.0"))
 
     installer.extras(["foo"])
-    installer.run()
+    result = installer.run()
+    assert result == 0
 
-    # But should not be installed
-    assert 4 == installer.executor.installations_count  # A, B, C, D
+    assert installer.executor.installations_count == 4  # A, B, C, D
 
 
-def test_installer_with_pypi_repository(package, locker, installed, config):
-    pool = Pool()
+def test_installer_with_pypi_repository(
+    package: ProjectPackage,
+    locker: Locker,
+    installed: CustomInstalledRepository,
+    config: Config,
+    env: NullEnv,
+) -> None:
+    pool = RepositoryPool()
     pool.add_repository(MockRepository())
 
     installer = Installer(
-        NullIO(), NullEnv(), package, locker, pool, config, installed=installed
+        NullIO(), env, package, locker, pool, config, installed=installed
     )
 
+    package.python_versions = ">=3.7"
     package.add_dependency(Factory.create_dependency("pytest", "^3.5", groups=["dev"]))
-    installer.run()
+    result = installer.run()
+    assert result == 0
 
     expected = fixture("with-pypi-repository")
-    assert not DeepDiff(expected, locker.written_data, ignore_order=True)
+
+    assert expected == locker.written_data
 
 
-def test_run_installs_with_local_file(installer, locker, repo, package, fixture_dir):
+def test_run_installs_with_local_file(
+    installer: Installer,
+    locker: Locker,
+    repo: Repository,
+    package: ProjectPackage,
+    fixture_dir: FixtureDirGetter,
+) -> None:
+    root_dir = Path(__file__).parent.parent.parent
+    package.root_dir = root_dir
+    locker.set_lock_path(root_dir)
     file_path = fixture_dir("distributions/demo-0.1.0-py2.py3-none-any.whl")
-    package.add_dependency(Factory.create_dependency("demo", {"file": str(file_path)}))
-
-    repo.add_package(get_package("pendulum", "1.4.4"))
-
-    installer.run()
-
-    expected = fixture("with-file-dependency")
-
-    assert locker.written_data == expected
-    assert 2 == installer.executor.installations_count
-
-
-def test_run_installs_wheel_with_no_requires_dist(
-    installer, locker, repo, package, fixture_dir
-):
-    file_path = fixture_dir(
-        "wheel_with_no_requires_dist/demo-0.1.0-py2.py3-none-any.whl"
-    )
-    package.add_dependency(Factory.create_dependency("demo", {"file": str(file_path)}))
-
-    installer.run()
-
-    expected = fixture("with-wheel-dependency-no-requires-dist")
-
-    assert locker.written_data == expected
-
-    assert 1 == installer.executor.installations_count
-
-
-def test_run_installs_with_local_poetry_directory_and_extras(
-    installer, locker, repo, package, tmpdir, fixture_dir
-):
-    file_path = fixture_dir("project_with_extras")
     package.add_dependency(
         Factory.create_dependency(
-            "project-with-extras", {"path": str(file_path), "extras": ["extras_a"]}
+            "demo", {"file": str(file_path.relative_to(root_dir))}, root_dir=root_dir
         )
     )
 
     repo.add_package(get_package("pendulum", "1.4.4"))
 
-    installer.run()
+    result = installer.run()
+    assert result == 0
+
+    expected = fixture("with-file-dependency")
+
+    assert locker.written_data == expected
+    assert installer.executor.installations_count == 2
+
+
+def test_run_installs_wheel_with_no_requires_dist(
+    installer: Installer,
+    locker: Locker,
+    repo: Repository,
+    package: ProjectPackage,
+    fixture_dir: FixtureDirGetter,
+) -> None:
+    root_dir = Path(__file__).parent.parent.parent
+    package.root_dir = root_dir
+    locker.set_lock_path(root_dir)
+    file_path = fixture_dir(
+        "wheel_with_no_requires_dist/demo-0.1.0-py2.py3-none-any.whl"
+    )
+    package.add_dependency(
+        Factory.create_dependency(
+            "demo", {"file": str(file_path.relative_to(root_dir))}, root_dir=root_dir
+        )
+    )
+
+    result = installer.run()
+    assert result == 0
+
+    expected = fixture("with-wheel-dependency-no-requires-dist")
+
+    assert locker.written_data == expected
+
+    assert installer.executor.installations_count == 1
+
+
+def test_run_installs_with_local_poetry_directory_and_extras(
+    installer: Installer,
+    locker: Locker,
+    repo: Repository,
+    package: ProjectPackage,
+    tmpdir: Path,
+    fixture_dir: FixtureDirGetter,
+) -> None:
+    root_dir = Path(__file__).parent.parent.parent
+    package.root_dir = root_dir
+    locker.set_lock_path(root_dir)
+    file_path = fixture_dir("project_with_extras")
+    package.add_dependency(
+        Factory.create_dependency(
+            "project-with-extras",
+            {"path": str(file_path.relative_to(root_dir)), "extras": ["extras_a"]},
+            root_dir=root_dir,
+        )
+    )
+
+    repo.add_package(get_package("pendulum", "1.4.4"))
+
+    result = installer.run()
+    assert result == 0
 
     expected = fixture("with-directory-dependency-poetry")
     assert locker.written_data == expected
 
-    assert 2 == installer.executor.installations_count
+    assert installer.executor.installations_count == 2
 
 
-def test_run_installs_with_local_poetry_directory_transitive(
-    installer, locker, repo, package, tmpdir, fixture_dir
-):
+@pytest.mark.parametrize("skip_directory", [True, False])
+def test_run_installs_with_local_poetry_directory_and_skip_directory_flag(
+    installer: Installer,
+    locker: Locker,
+    repo: Repository,
+    package: ProjectPackage,
+    fixture_dir: FixtureDirGetter,
+    skip_directory: bool,
+) -> None:
+    """When we set Installer.skip_directory(True) no path dependencies should
+    be installed (including transitive dependencies).
+    """
     root_dir = fixture_dir("directory")
     package.root_dir = root_dir
     locker.set_lock_path(root_dir)
@@ -1159,18 +1291,36 @@ def test_run_installs_with_local_poetry_directory_transitive(
     repo.add_package(get_package("pendulum", "1.4.4"))
     repo.add_package(get_package("cachy", "0.2.0"))
 
-    installer.run()
+    installer.skip_directory(skip_directory)
+
+    result = installer.run()
+    assert result == 0
 
     expected = fixture("with-directory-dependency-poetry-transitive")
 
     assert locker.written_data == expected
 
-    assert 6 == installer.executor.installations_count
+    assert isinstance(installer.executor, Executor)
+    directory_installs = [
+        p.name for p in installer.executor.installations if p.source_type == "directory"
+    ]
+
+    if skip_directory:
+        assert not directory_installs, directory_installs
+        assert installer.executor.installations_count == 2
+    else:
+        assert directory_installs, directory_installs
+        assert installer.executor.installations_count == 6
 
 
 def test_run_installs_with_local_poetry_file_transitive(
-    installer, locker, repo, package, tmpdir, fixture_dir
-):
+    installer: Installer,
+    locker: Locker,
+    repo: Repository,
+    package: ProjectPackage,
+    tmpdir: str,
+    fixture_dir: FixtureDirGetter,
+) -> None:
     root_dir = fixture_dir("directory")
     package.root_dir = root_dir
     locker.set_lock_path(root_dir)
@@ -1188,35 +1338,51 @@ def test_run_installs_with_local_poetry_file_transitive(
     repo.add_package(get_package("pendulum", "1.4.4"))
     repo.add_package(get_package("cachy", "0.2.0"))
 
-    installer.run()
+    result = installer.run()
+    assert result == 0
 
     expected = fixture("with-file-dependency-transitive")
 
     assert locker.written_data == expected
 
-    assert 4 == installer.executor.installations_count
+    assert installer.executor.installations_count == 4
 
 
 def test_run_installs_with_local_setuptools_directory(
-    installer, locker, repo, package, tmpdir, fixture_dir
-):
+    installer: Installer,
+    locker: Locker,
+    repo: Repository,
+    package: ProjectPackage,
+    tmpdir: Path,
+    fixture_dir: FixtureDirGetter,
+) -> None:
+    root_dir = Path(__file__).parent.parent.parent
+    package.root_dir = root_dir
+    locker.set_lock_path(root_dir)
     file_path = fixture_dir("project_with_setup/")
     package.add_dependency(
-        Factory.create_dependency("project-with-setup", {"path": str(file_path)})
+        Factory.create_dependency(
+            "project-with-setup",
+            {"path": str(file_path.relative_to(root_dir))},
+            root_dir=root_dir,
+        )
     )
 
     repo.add_package(get_package("pendulum", "1.4.4"))
     repo.add_package(get_package("cachy", "0.2.0"))
 
-    installer.run()
+    result = installer.run()
+    assert result == 0
 
     expected = fixture("with-directory-dependency-setuptools")
 
     assert locker.written_data == expected
-    assert 3 == installer.executor.installations_count
+    assert installer.executor.installations_count == 3
 
 
-def test_run_with_prereleases(installer, locker, repo, package):
+def test_run_with_prereleases(
+    installer: Installer, locker: Locker, repo: Repository, package: ProjectPackage
+) -> None:
     locker.locked(True)
     locker.mock_lock_data(
         {
@@ -1224,7 +1390,6 @@ def test_run_with_prereleases(installer, locker, repo, package):
                 {
                     "name": "A",
                     "version": "1.0a2",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -1235,7 +1400,7 @@ def test_run_with_prereleases(installer, locker, repo, package):
                 "python-versions": "*",
                 "platform": "*",
                 "content-hash": "123456789",
-                "hashes": {"A": []},
+                "files": {"A": []},
             },
         }
     )
@@ -1252,13 +1417,16 @@ def test_run_with_prereleases(installer, locker, repo, package):
     installer.update(True)
     installer.whitelist({"B": "^1.1"})
 
-    installer.run()
+    result = installer.run()
+    assert result == 0
+
     expected = fixture("with-prereleases")
-
     assert locker.written_data == expected
 
 
-def test_run_changes_category_if_needed(installer, locker, repo, package):
+def test_run_update_all_with_lock(
+    installer: Installer, locker: Locker, repo: Repository, package: ProjectPackage
+) -> None:
     locker.locked(True)
     locker.mock_lock_data(
         {
@@ -1266,7 +1434,6 @@ def test_run_changes_category_if_needed(installer, locker, repo, package):
                 {
                     "name": "A",
                     "version": "1.0",
-                    "category": "dev",
                     "optional": True,
                     "platform": "*",
                     "python-versions": "*",
@@ -1277,52 +1444,7 @@ def test_run_changes_category_if_needed(installer, locker, repo, package):
                 "python-versions": "*",
                 "platform": "*",
                 "content-hash": "123456789",
-                "hashes": {"A": []},
-            },
-        }
-    )
-    package_a = get_package("A", "1.0")
-    package_b = get_package("B", "1.1")
-    package_b.add_dependency(Factory.create_dependency("A", "^1.0"))
-    repo.add_package(package_a)
-    repo.add_package(package_b)
-
-    package.add_dependency(
-        Factory.create_dependency(
-            "A", {"version": "^1.0", "optional": True}, groups=["dev"]
-        )
-    )
-    package.add_dependency(Factory.create_dependency("B", "^1.1"))
-
-    installer.update(True)
-    installer.whitelist(["B"])
-
-    installer.run()
-    expected = fixture("with-category-change")
-
-    assert locker.written_data == expected
-
-
-def test_run_update_all_with_lock(installer, locker, repo, package):
-    locker.locked(True)
-    locker.mock_lock_data(
-        {
-            "package": [
-                {
-                    "name": "A",
-                    "version": "1.0",
-                    "category": "dev",
-                    "optional": True,
-                    "platform": "*",
-                    "python-versions": "*",
-                    "checksum": [],
-                }
-            ],
-            "metadata": {
-                "python-versions": "*",
-                "platform": "*",
-                "content-hash": "123456789",
-                "hashes": {"A": []},
+                "files": {"A": []},
             },
         }
     )
@@ -1334,13 +1456,16 @@ def test_run_update_all_with_lock(installer, locker, repo, package):
 
     installer.update(True)
 
-    installer.run()
-    expected = fixture("update-with-lock")
+    result = installer.run()
+    assert result == 0
 
+    expected = fixture("update-with-lock")
     assert locker.written_data == expected
 
 
-def test_run_update_with_locked_extras(installer, locker, repo, package):
+def test_run_update_with_locked_extras(
+    installer: Installer, locker: Locker, repo: Repository, package: ProjectPackage
+) -> None:
     locker.locked(True)
     locker.mock_lock_data(
         {
@@ -1348,7 +1473,6 @@ def test_run_update_with_locked_extras(installer, locker, repo, package):
                 {
                     "name": "A",
                     "version": "1.0",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -1358,7 +1482,6 @@ def test_run_update_with_locked_extras(installer, locker, repo, package):
                 {
                     "name": "B",
                     "version": "1.0",
-                    "category": "dev",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -1367,7 +1490,6 @@ def test_run_update_with_locked_extras(installer, locker, repo, package):
                 {
                     "name": "C",
                     "version": "1.1",
-                    "category": "dev",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -1379,14 +1501,14 @@ def test_run_update_with_locked_extras(installer, locker, repo, package):
                 "python-versions": "*",
                 "platform": "*",
                 "content-hash": "123456789",
-                "hashes": {"A": [], "B": [], "C": []},
+                "files": {"A": [], "B": [], "C": []},
             },
         }
     )
     package_a = get_package("A", "1.0")
-    package_a.extras["foo"] = [get_dependency("B")]
+    package_a.extras[canonicalize_name("foo")] = [get_dependency("B")]
     b_dependency = get_dependency("B", "^1.0", optional=True)
-    b_dependency.in_extras.append("foo")
+    b_dependency.in_extras.append(canonicalize_name("foo"))
     c_dependency = get_dependency("C", "^1.0")
     c_dependency.python_versions = "~2.7"
     package_a.add_dependency(b_dependency)
@@ -1405,15 +1527,16 @@ def test_run_update_with_locked_extras(installer, locker, repo, package):
     installer.update(True)
     installer.whitelist("D")
 
-    installer.run()
-    expected = fixture("update-with-locked-extras")
+    result = installer.run()
+    assert result == 0
 
+    expected = fixture("update-with-locked-extras")
     assert locker.written_data == expected
 
 
 def test_run_install_duplicate_dependencies_different_constraints(
-    installer, locker, repo, package
-):
+    installer: Installer, locker: Locker, repo: Repository, package: ProjectPackage
+) -> None:
     package.add_dependency(Factory.create_dependency("A", "*"))
 
     package_a = get_package("A", "1.0")
@@ -1438,25 +1561,27 @@ def test_run_install_duplicate_dependencies_different_constraints(
     repo.add_package(package_c12)
     repo.add_package(package_c15)
 
-    installer.run()
+    result = installer.run()
+    assert result == 0
 
     expected = fixture("with-duplicate-dependencies")
 
     assert locker.written_data == expected
 
+    assert isinstance(installer.executor, Executor)
     installs = installer.executor.installations
-    assert 3 == installer.executor.installations_count
+    assert installer.executor.installations_count == 3
     assert installs[0] == package_c12
     assert installs[1] == package_b10
     assert installs[2] == package_a
 
-    assert 0 == installer.executor.updates_count
-    assert 0 == installer.executor.removals_count
+    assert installer.executor.updates_count == 0
+    assert installer.executor.removals_count == 0
 
 
 def test_run_install_duplicate_dependencies_different_constraints_with_lock(
-    installer, locker, repo, package
-):
+    installer: Installer, locker: Locker, repo: Repository, package: ProjectPackage
+) -> None:
     locker.locked(True)
     locker.mock_lock_data(
         {
@@ -1464,7 +1589,6 @@ def test_run_install_duplicate_dependencies_different_constraints_with_lock(
                 {
                     "name": "A",
                     "version": "1.0",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -1479,7 +1603,6 @@ def test_run_install_duplicate_dependencies_different_constraints_with_lock(
                 {
                     "name": "B",
                     "version": "1.0",
-                    "category": "dev",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -1490,7 +1613,6 @@ def test_run_install_duplicate_dependencies_different_constraints_with_lock(
                 {
                     "name": "B",
                     "version": "2.0",
-                    "category": "dev",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -1501,7 +1623,6 @@ def test_run_install_duplicate_dependencies_different_constraints_with_lock(
                 {
                     "name": "C",
                     "version": "1.2",
-                    "category": "dev",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -1510,7 +1631,6 @@ def test_run_install_duplicate_dependencies_different_constraints_with_lock(
                 {
                     "name": "C",
                     "version": "1.5",
-                    "category": "dev",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -1521,7 +1641,7 @@ def test_run_install_duplicate_dependencies_different_constraints_with_lock(
                 "python-versions": "*",
                 "platform": "*",
                 "content-hash": "123456789",
-                "hashes": {"A": [], "B": [], "C": []},
+                "files": {"A": [], "B": [], "C": []},
             },
         }
     )
@@ -1550,20 +1670,25 @@ def test_run_install_duplicate_dependencies_different_constraints_with_lock(
     repo.add_package(package_c15)
 
     installer.update(True)
-    installer.run()
+    result = installer.run()
+    assert result == 0
 
     expected = fixture("with-duplicate-dependencies")
 
     assert locker.written_data == expected
 
-    assert 3 == installer.executor.installations_count
-    assert 0 == installer.executor.updates_count
-    assert 0 == installer.executor.removals_count
+    assert installer.executor.installations_count == 3
+    assert installer.executor.updates_count == 0
+    assert installer.executor.removals_count == 0
 
 
 def test_run_update_uninstalls_after_removal_transient_dependency(
-    installer, locker, repo, package, installed
-):
+    installer: Installer,
+    locker: Locker,
+    repo: Repository,
+    package: ProjectPackage,
+    installed: CustomInstalledRepository,
+) -> None:
     locker.locked(True)
     locker.mock_lock_data(
         {
@@ -1571,7 +1696,6 @@ def test_run_update_uninstalls_after_removal_transient_dependency(
                 {
                     "name": "A",
                     "version": "1.0",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -1581,7 +1705,6 @@ def test_run_update_uninstalls_after_removal_transient_dependency(
                 {
                     "name": "B",
                     "version": "1.0",
-                    "category": "dev",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -1592,7 +1715,7 @@ def test_run_update_uninstalls_after_removal_transient_dependency(
                 "python-versions": "*",
                 "platform": "*",
                 "content-hash": "123456789",
-                "hashes": {"A": [], "B": []},
+                "files": {"A": [], "B": []},
             },
         }
     )
@@ -1612,16 +1735,21 @@ def test_run_update_uninstalls_after_removal_transient_dependency(
     installed.add_package(get_package("B", "1.0"))
 
     installer.update(True)
-    installer.run()
+    result = installer.run()
+    assert result == 0
 
-    assert 0 == installer.executor.installations_count
-    assert 0 == installer.executor.updates_count
-    assert 1 == installer.executor.removals_count
+    assert installer.executor.installations_count == 0
+    assert installer.executor.updates_count == 0
+    assert installer.executor.removals_count == 1
 
 
 def test_run_install_duplicate_dependencies_different_constraints_with_lock_update(
-    installer, locker, repo, package, installed
-):
+    installer: Installer,
+    locker: Locker,
+    repo: Repository,
+    package: ProjectPackage,
+    installed: CustomInstalledRepository,
+) -> None:
     locker.locked(True)
     locker.mock_lock_data(
         {
@@ -1629,7 +1757,6 @@ def test_run_install_duplicate_dependencies_different_constraints_with_lock_upda
                 {
                     "name": "A",
                     "version": "1.0",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -1644,7 +1771,6 @@ def test_run_install_duplicate_dependencies_different_constraints_with_lock_upda
                 {
                     "name": "B",
                     "version": "1.0",
-                    "category": "dev",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -1655,7 +1781,6 @@ def test_run_install_duplicate_dependencies_different_constraints_with_lock_upda
                 {
                     "name": "B",
                     "version": "2.0",
-                    "category": "dev",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -1666,7 +1791,6 @@ def test_run_install_duplicate_dependencies_different_constraints_with_lock_upda
                 {
                     "name": "C",
                     "version": "1.2",
-                    "category": "dev",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -1675,7 +1799,6 @@ def test_run_install_duplicate_dependencies_different_constraints_with_lock_upda
                 {
                     "name": "C",
                     "version": "1.5",
-                    "category": "dev",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -1686,7 +1809,7 @@ def test_run_install_duplicate_dependencies_different_constraints_with_lock_upda
                 "python-versions": "*",
                 "platform": "*",
                 "content-hash": "123456789",
-                "hashes": {"A": [], "B": [], "C": []},
+                "files": {"A": [], "B": [], "C": []},
             },
         }
     )
@@ -1713,23 +1836,28 @@ def test_run_install_duplicate_dependencies_different_constraints_with_lock_upda
 
     installer.update(True)
     installer.whitelist(["A"])
-    installer.run()
+    result = installer.run()
+    assert result == 0
 
     expected = fixture("with-duplicate-dependencies-update")
 
     assert locker.written_data == expected
 
-    assert 2 == installer.executor.installations_count
-    assert 1 == installer.executor.updates_count
-    assert 0 == installer.executor.removals_count
+    assert installer.executor.installations_count == 2
+    assert installer.executor.updates_count == 1
+    assert installer.executor.removals_count == 0
 
 
 @pytest.mark.skip(
     "This is not working at the moment due to limitations in the resolver"
 )
-def test_installer_test_solver_finds_compatible_package_for_dependency_python_not_fully_compatible_with_package_python(
-    installer, locker, repo, package, installed
-):
+def test_installer_test_solver_finds_compatible_package_for_dependency_python_not_fully_compatible_with_package_python(  # noqa: E501
+    installer: Installer,
+    locker: Locker,
+    repo: Repository,
+    package: ProjectPackage,
+    installed: CustomInstalledRepository,
+) -> None:
     package.python_versions = "~2.7 || ^3.4"
     package.add_dependency(
         Factory.create_dependency("A", {"version": "^1.0", "python": "^3.5"})
@@ -1744,16 +1872,24 @@ def test_installer_test_solver_finds_compatible_package_for_dependency_python_no
     repo.add_package(package_a100)
     repo.add_package(package_a101)
 
-    installer.run()
+    result = installer.run()
+    assert result == 0
 
     expected = fixture("with-conditional-dependency")
     assert locker.written_data == expected
-    assert 1 == installer.executor.installations_count
+    assert installer.executor.installations_count == 1
 
 
-def test_installer_required_extras_should_not_be_removed_when_updating_single_dependency(
-    installer, locker, repo, package, installed, env, pool, config
-):
+def test_installer_required_extras_should_not_be_removed_when_updating_single_dependency(  # noqa: E501
+    installer: Installer,
+    locker: Locker,
+    repo: Repository,
+    package: ProjectPackage,
+    installed: CustomInstalledRepository,
+    env: NullEnv,
+    pool: RepositoryPool,
+    config: Config,
+) -> None:
     package.add_dependency(Factory.create_dependency("A", {"version": "^1.0"}))
 
     package_a = get_package("A", "1.0.0")
@@ -1765,7 +1901,7 @@ def test_installer_required_extras_should_not_be_removed_when_updating_single_de
     package_b.add_dependency(
         Factory.create_dependency("C", {"version": "^1.0", "optional": True})
     )
-    package_b.extras = {"foo": [get_dependency("C")]}
+    package_b.extras = {canonicalize_name("foo"): [get_dependency("C")]}
 
     package_c = get_package("C", "1.0.0")
     package_d = get_package("D", "1.0.0")
@@ -1776,11 +1912,12 @@ def test_installer_required_extras_should_not_be_removed_when_updating_single_de
     repo.add_package(package_d)
 
     installer.update(True)
-    installer.run()
+    result = installer.run()
+    assert result == 0
 
-    assert 3 == installer.executor.installations_count
-    assert 0 == installer.executor.updates_count
-    assert 0 == installer.executor.removals_count
+    assert installer.executor.installations_count == 3
+    assert installer.executor.updates_count == 0
+    assert installer.executor.removals_count == 0
 
     package.add_dependency(Factory.create_dependency("D", "^1.0"))
     locker.locked(True)
@@ -1800,23 +1937,28 @@ def test_installer_required_extras_should_not_be_removed_when_updating_single_de
         installed=installed,
         executor=Executor(env, pool, config, NullIO()),
     )
-    installer.use_executor()
-
     installer.update(True)
     installer.whitelist(["D"])
-    installer.run()
+    result = installer.run()
+    assert result == 0
 
-    assert 1 == installer.executor.installations_count
-    assert 0 == installer.executor.updates_count
-    assert 0 == installer.executor.removals_count
+    assert installer.executor.installations_count == 1
+    assert installer.executor.updates_count == 0
+    assert installer.executor.removals_count == 0
 
 
-def test_installer_required_extras_should_not_be_removed_when_updating_single_dependency_pypi_repository(
-    locker, repo, package, installed, env, mocker, config
-):
+def test_installer_required_extras_should_not_be_removed_when_updating_single_dependency_pypi_repository(  # noqa: E501
+    locker: Locker,
+    repo: Repository,
+    package: ProjectPackage,
+    installed: CustomInstalledRepository,
+    env: NullEnv,
+    mocker: MockerFixture,
+    config: Config,
+) -> None:
     mocker.patch("sys.platform", "darwin")
 
-    pool = Pool()
+    pool = RepositoryPool()
     pool.add_repository(MockRepository())
 
     installer = Installer(
@@ -1829,22 +1971,23 @@ def test_installer_required_extras_should_not_be_removed_when_updating_single_de
         installed=installed,
         executor=Executor(env, pool, config, NullIO()),
     )
-    installer.use_executor()
 
     package.add_dependency(Factory.create_dependency("poetry", {"version": "^0.12.0"}))
 
     installer.update(True)
-    installer.run()
+    result = installer.run()
+    assert result == 0
 
-    assert 3 == installer.executor.installations_count
-    assert 0 == installer.executor.updates_count
-    assert 0 == installer.executor.removals_count
+    assert installer.executor.installations_count == 3
+    assert installer.executor.updates_count == 0
+    assert installer.executor.removals_count == 0
 
     package.add_dependency(Factory.create_dependency("pytest", "^3.5"))
 
     locker.locked(True)
     locker.mock_lock_data(locker.written_data)
 
+    assert isinstance(installer.executor, Executor)
     for pkg in installer.executor.installations:
         installed.add_package(pkg)
 
@@ -1858,21 +2001,25 @@ def test_installer_required_extras_should_not_be_removed_when_updating_single_de
         installed=installed,
         executor=Executor(env, pool, config, NullIO()),
     )
-    installer.use_executor()
-
     installer.update(True)
     installer.whitelist(["pytest"])
-    installer.run()
+    result = installer.run()
+    assert result == 0
 
-    assert 7 == installer.executor.installations_count
-    assert 0 == installer.executor.updates_count
-    assert 0 == installer.executor.removals_count
+    assert installer.executor.installations_count == 7
+    assert installer.executor.updates_count == 0
+    assert installer.executor.removals_count == 0
 
 
 def test_installer_required_extras_should_be_installed(
-    locker, repo, package, installed, env, config
-):
-    pool = Pool()
+    locker: Locker,
+    repo: Repository,
+    package: ProjectPackage,
+    installed: CustomInstalledRepository,
+    env: NullEnv,
+    config: Config,
+) -> None:
+    pool = RepositoryPool()
     pool.add_repository(MockRepository())
 
     installer = Installer(
@@ -1885,8 +2032,6 @@ def test_installer_required_extras_should_be_installed(
         installed=installed,
         executor=Executor(env, pool, config, NullIO()),
     )
-    installer.use_executor()
-
     package.add_dependency(
         Factory.create_dependency(
             "cachecontrol", {"version": "^0.12.5", "extras": ["filecache"]}
@@ -1894,11 +2039,12 @@ def test_installer_required_extras_should_be_installed(
     )
 
     installer.update(True)
-    installer.run()
+    result = installer.run()
+    assert result == 0
 
-    assert 2 == installer.executor.installations_count
-    assert 0 == installer.executor.updates_count
-    assert 0 == installer.executor.removals_count
+    assert installer.executor.installations_count == 2
+    assert installer.executor.updates_count == 0
+    assert installer.executor.removals_count == 0
 
     locker.locked(True)
     locker.mock_lock_data(locker.written_data)
@@ -1913,19 +2059,18 @@ def test_installer_required_extras_should_be_installed(
         installed=installed,
         executor=Executor(env, pool, config, NullIO()),
     )
-    installer.use_executor()
-
     installer.update(True)
-    installer.run()
+    result = installer.run()
+    assert result == 0
 
-    assert 2 == installer.executor.installations_count
-    assert 0 == installer.executor.updates_count
-    assert 0 == installer.executor.removals_count
+    assert installer.executor.installations_count == 2
+    assert installer.executor.updates_count == 0
+    assert installer.executor.removals_count == 0
 
 
 def test_update_multiple_times_with_split_dependencies_is_idempotent(
-    installer, locker, repo, package
-):
+    installer: Installer, locker: Locker, repo: Repository, package: ProjectPackage
+) -> None:
     locker.locked(True)
     locker.mock_lock_data(
         {
@@ -1933,7 +2078,6 @@ def test_update_multiple_times_with_split_dependencies_is_idempotent(
                 {
                     "name": "A",
                     "version": "1.0",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -1943,7 +2087,6 @@ def test_update_multiple_times_with_split_dependencies_is_idempotent(
                 {
                     "name": "B",
                     "version": "1.0.1",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": ">=2.7,!=3.0.*,!=3.1.*,!=3.2.*,!=3.3.*",
@@ -1955,7 +2098,7 @@ def test_update_multiple_times_with_split_dependencies_is_idempotent(
                 "python-versions": "*",
                 "platform": "*",
                 "content-hash": "123456789",
-                "hashes": {"A": [], "B": []},
+                "files": {"A": [], "B": []},
             },
         }
     )
@@ -1983,34 +2126,41 @@ def test_update_multiple_times_with_split_dependencies_is_idempotent(
     expected = fixture("with-multiple-updates")
 
     installer.update(True)
-    installer.run()
+    result = installer.run()
+    assert result == 0
 
-    assert expected == locker.written_data
-
-    locker.mock_lock_data(locker.written_data)
-
-    installer.update(True)
-    installer.run()
-
-    assert expected == locker.written_data
+    assert locker.written_data == expected
 
     locker.mock_lock_data(locker.written_data)
 
     installer.update(True)
-    installer.run()
+    result = installer.run()
+    assert result == 0
 
-    assert expected == locker.written_data
+    assert locker.written_data == expected
+
+    locker.mock_lock_data(locker.written_data)
+
+    installer.update(True)
+    result = installer.run()
+    assert result == 0
+
+    assert locker.written_data == expected
 
 
 def test_installer_can_install_dependencies_from_forced_source(
-    locker, package, installed, env, config
-):
+    locker: Locker,
+    package: ProjectPackage,
+    installed: CustomInstalledRepository,
+    env: NullEnv,
+    config: Config,
+) -> None:
     package.python_versions = "^3.7"
     package.add_dependency(
         Factory.create_dependency("tomlkit", {"version": "^0.5", "source": "legacy"})
     )
 
-    pool = Pool()
+    pool = RepositoryPool()
     pool.add_repository(MockLegacyRepository())
     pool.add_repository(MockRepository())
 
@@ -2024,34 +2174,87 @@ def test_installer_can_install_dependencies_from_forced_source(
         installed=installed,
         executor=Executor(env, pool, config, NullIO()),
     )
-    installer.use_executor()
-
     installer.update(True)
-    installer.run()
+    result = installer.run()
+    assert result == 0
 
-    assert 1 == installer.executor.installations_count
-    assert 0 == installer.executor.updates_count
-    assert 0 == installer.executor.removals_count
+    assert installer.executor.installations_count == 1
+    assert installer.executor.updates_count == 0
+    assert installer.executor.removals_count == 0
 
 
-def test_run_installs_with_url_file(installer, locker, repo, package):
+def test_run_installs_with_url_file(
+    installer: Installer, locker: Locker, repo: Repository, package: ProjectPackage
+) -> None:
     url = "https://python-poetry.org/distributions/demo-0.1.0-py2.py3-none-any.whl"
     package.add_dependency(Factory.create_dependency("demo", {"url": url}))
 
     repo.add_package(get_package("pendulum", "1.4.4"))
 
-    installer.run()
+    result = installer.run()
+    assert result == 0
 
     expected = fixture("with-url-dependency")
 
     assert locker.written_data == expected
 
-    assert 2 == installer.executor.installations_count
+    assert installer.executor.installations_count == 2
+
+
+@pytest.mark.parametrize("env_platform", ["linux", "win32"])
+def test_run_installs_with_same_version_url_files(
+    pool: RepositoryPool,
+    locker: Locker,
+    installed: CustomInstalledRepository,
+    config: Config,
+    repo: Repository,
+    package: ProjectPackage,
+    env_platform: str,
+) -> None:
+    urls = {
+        "linux": "https://python-poetry.org/distributions/demo-0.1.0.tar.gz",
+        "win32": (
+            "https://python-poetry.org/distributions/demo-0.1.0-py2.py3-none-any.whl"
+        ),
+    }
+    for platform, url in urls.items():
+        package.add_dependency(
+            Factory.create_dependency(
+                "demo",
+                {"url": url, "markers": f"sys_platform == '{platform}'"},
+            )
+        )
+    repo.add_package(get_package("pendulum", "1.4.4"))
+
+    installer = Installer(
+        NullIO(),
+        MockEnv(platform=env_platform),
+        package,
+        locker,
+        pool,
+        config,
+        installed=installed,
+        executor=Executor(
+            MockEnv(platform=env_platform),
+            pool,
+            config,
+            NullIO(),
+        ),
+    )
+    result = installer.run()
+    assert result == 0
+
+    expected = fixture("with-same-version-url-dependencies")
+    assert locker.written_data == expected
+    assert isinstance(installer.executor, Executor)
+    assert installer.executor.installations_count == 2
+    demo_package = next(p for p in installer.executor.installations if p.name == "demo")
+    assert demo_package.source_url == urls[env_platform]
 
 
 def test_installer_uses_prereleases_if_they_are_compatible(
-    installer, locker, package, repo
-):
+    installer: Installer, locker: Locker, package: ProjectPackage, repo: Repository
+) -> None:
     package.python_versions = "~2.7 || ^3.4"
     package.add_dependency(
         Factory.create_dependency(
@@ -2064,9 +2267,9 @@ def test_installer_uses_prereleases_if_they_are_compatible(
 
     repo.add_package(package_b)
 
-    installer.run()
+    result = installer.run()
+    assert result == 0
 
-    del installer.installer.installs[:]
     locker.locked(True)
     locker.mock_lock_data(locker.written_data)
 
@@ -2074,15 +2277,20 @@ def test_installer_uses_prereleases_if_they_are_compatible(
 
     installer.whitelist(["b"])
     installer.update(True)
-    installer.run()
+    result = installer.run()
+    assert result == 0
 
-    assert 2 == installer.executor.installations_count
+    assert installer.executor.installations_count == 2
 
 
 def test_installer_can_handle_old_lock_files(
-    installer, locker, package, repo, installed, config
-):
-    pool = Pool()
+    locker: Locker,
+    package: ProjectPackage,
+    repo: Repository,
+    installed: CustomInstalledRepository,
+    config: Config,
+) -> None:
+    pool = RepositoryPool()
     pool.add_repository(MockRepository())
 
     package.add_dependency(Factory.create_dependency("pytest", "^3.5", groups=["dev"]))
@@ -2100,11 +2308,10 @@ def test_installer_can_handle_old_lock_files(
         installed=installed,
         executor=Executor(MockEnv(), pool, config, NullIO()),
     )
-    installer.use_executor()
+    result = installer.run()
+    assert result == 0
 
-    installer.run()
-
-    assert 6 == installer.executor.installations_count
+    assert installer.executor.installations_count == 6
 
     installer = Installer(
         NullIO(),
@@ -2121,12 +2328,11 @@ def test_installer_can_handle_old_lock_files(
             NullIO(),
         ),
     )
-    installer.use_executor()
-
-    installer.run()
+    result = installer.run()
+    assert result == 0
 
     # funcsigs will be added
-    assert 7 == installer.executor.installations_count
+    assert installer.executor.installations_count == 7
 
     installer = Installer(
         NullIO(),
@@ -2143,42 +2349,71 @@ def test_installer_can_handle_old_lock_files(
             NullIO(),
         ),
     )
-    installer.use_executor()
-
-    installer.run()
+    result = installer.run()
+    assert result == 0
 
     # colorama will be added
-    assert 8 == installer.executor.installations_count
+    assert installer.executor.installations_count == 8
+
+
+def test_installer_does_not_write_lock_file_when_installation_fails(
+    installer: Installer,
+    locker: Locker,
+    repo: Repository,
+    package: ProjectPackage,
+    mocker: MockerFixture,
+) -> None:
+    repo.add_package(get_package("A", "1.0"))
+    package.add_dependency(Factory.create_dependency("A", "~1.0"))
+
+    locker.locked(False)
+
+    mocker.patch("poetry.installation.installer.Installer._execute", return_value=1)
+    result = installer.run()
+    assert result == 1  # error
+
+    assert locker._lock_data is None
+
+    assert installer.executor.installations_count == 0
+    assert installer.executor.updates_count == 0
+    assert installer.executor.removals_count == 0
 
 
 @pytest.mark.parametrize("quiet", [True, False])
-def test_run_with_dependencies_quiet(installer, locker, repo, package, quiet):
+def test_run_with_dependencies_quiet(
+    installer: Installer,
+    locker: Locker,
+    repo: Repository,
+    package: ProjectPackage,
+    quiet: bool,
+) -> None:
     package_a = get_package("A", "1.0")
     package_b = get_package("B", "1.1")
     repo.add_package(package_a)
     repo.add_package(package_b)
 
-    installer._io = IO(Input(), BufferedOutput(), BufferedOutput())
+    installer._io = BufferedIO(Input())
     installer._io.set_verbosity(Verbosity.QUIET if quiet else Verbosity.NORMAL)
 
     package.add_dependency(Factory.create_dependency("A", "~1.0"))
     package.add_dependency(Factory.create_dependency("B", "^1.0"))
 
-    installer.run()
-    expected = fixture("with-dependencies")
+    result = installer.run()
+    assert result == 0
 
+    expected = fixture("with-dependencies")
     assert locker.written_data == expected
 
-    installer._io.output._buffer.seek(0)
+    output = installer._io.fetch_output()
     if quiet:
-        assert installer._io.output._buffer.read() == ""
+        assert output == ""
     else:
-        assert installer._io.output._buffer.read() != ""
+        assert output != ""
 
 
 def test_installer_should_use_the_locked_version_of_git_dependencies(
-    installer, locker, package, repo
-):
+    installer: Installer, locker: Locker, package: ProjectPackage, repo: Repository
+) -> None:
     locker.locked(True)
     locker.mock_lock_data(
         {
@@ -2186,7 +2421,6 @@ def test_installer_should_use_the_locked_version_of_git_dependencies(
                 {
                     "name": "demo",
                     "version": "0.1.1",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -2202,7 +2436,6 @@ def test_installer_should_use_the_locked_version_of_git_dependencies(
                 {
                     "name": "pendulum",
                     "version": "1.4.4",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -2214,7 +2447,7 @@ def test_installer_should_use_the_locked_version_of_git_dependencies(
                 "python-versions": "*",
                 "platform": "*",
                 "content-hash": "123456789",
-                "hashes": {"demo": [], "pendulum": []},
+                "files": {"demo": [], "pendulum": []},
             },
         }
     )
@@ -2227,8 +2460,10 @@ def test_installer_should_use_the_locked_version_of_git_dependencies(
 
     repo.add_package(get_package("pendulum", "1.4.4"))
 
-    installer.run()
+    result = installer.run()
+    assert result == 0
 
+    assert isinstance(installer.executor, Executor)
     assert installer.executor.installations[-1] == Package(
         "demo",
         "0.1.1",
@@ -2236,4 +2471,191 @@ def test_installer_should_use_the_locked_version_of_git_dependencies(
         source_url="https://github.com/demo/demo.git",
         source_reference="master",
         source_resolved_reference="123456",
+    )
+
+
+@pytest.mark.parametrize("is_locked", [False, True])
+def test_installer_should_use_the_locked_version_of_git_dependencies_with_extras(
+    installer: Installer,
+    locker: Locker,
+    package: ProjectPackage,
+    repo: Repository,
+    is_locked: bool,
+) -> None:
+    if is_locked:
+        locker.locked(True)
+        locker.mock_lock_data(fixture("with-vcs-dependency-with-extras"))
+        expected_reference = "123456"
+    else:
+        expected_reference = MOCK_DEFAULT_GIT_REVISION
+
+    package.add_dependency(
+        Factory.create_dependency(
+            "demo",
+            {
+                "git": "https://github.com/demo/demo.git",
+                "branch": "master",
+                "extras": ["foo"],
+            },
+        )
+    )
+
+    repo.add_package(get_package("pendulum", "1.4.4"))
+    repo.add_package(get_package("cleo", "1.0.0"))
+
+    result = installer.run()
+    assert result == 0
+
+    assert isinstance(installer.executor, Executor)
+    assert len(installer.executor.installations) == 3
+    assert installer.executor.installations[-1] == Package(
+        "demo",
+        "0.1.2",
+        source_type="git",
+        source_url="https://github.com/demo/demo.git",
+        source_reference="master",
+        source_resolved_reference=expected_reference,
+    )
+
+
+@pytest.mark.parametrize("is_locked", [False, True])
+def test_installer_should_use_the_locked_version_of_git_dependencies_without_reference(
+    installer: Installer,
+    locker: Locker,
+    package: ProjectPackage,
+    repo: Repository,
+    is_locked: bool,
+) -> None:
+    """
+    If there is no explicit reference (branch or tag or rev) in pyproject.toml,
+    HEAD is used.
+    """
+    if is_locked:
+        locker.locked(True)
+        locker.mock_lock_data(fixture("with-vcs-dependency-without-ref"))
+        expected_reference = "123456"
+    else:
+        expected_reference = MOCK_DEFAULT_GIT_REVISION
+
+    package.add_dependency(
+        Factory.create_dependency("demo", {"git": "https://github.com/demo/demo.git"})
+    )
+
+    repo.add_package(get_package("pendulum", "1.4.4"))
+
+    result = installer.run()
+    assert result == 0
+
+    assert isinstance(installer.executor, Executor)
+    assert len(installer.executor.installations) == 2
+    assert installer.executor.installations[-1] == Package(
+        "demo",
+        "0.1.2",
+        source_type="git",
+        source_url="https://github.com/demo/demo.git",
+        source_reference="HEAD",
+        source_resolved_reference=expected_reference,
+    )
+
+
+# https://github.com/python-poetry/poetry/issues/6710
+@pytest.mark.parametrize("env_platform", ["darwin", "linux"])
+def test_installer_distinguishes_locked_packages_by_source(
+    pool: RepositoryPool,
+    locker: Locker,
+    installed: CustomInstalledRepository,
+    config: Config,
+    repo: Repository,
+    package: ProjectPackage,
+    env_platform: str,
+) -> None:
+    # Require 1.11.0+cpu from pytorch for most platforms, but specify 1.11.0 and pypi on
+    # darwin.
+    package.add_dependency(
+        Factory.create_dependency(
+            "torch",
+            {
+                "version": "1.11.0+cpu",
+                "markers": "sys_platform != 'darwin'",
+                "source": "pytorch",
+            },
+        )
+    )
+    package.add_dependency(
+        Factory.create_dependency(
+            "torch",
+            {
+                "version": "1.11.0",
+                "markers": "sys_platform == 'darwin'",
+                "source": "pypi",
+            },
+        )
+    )
+
+    # Locking finds both the pypi and the pytorch packages.
+    locker.locked(True)
+    locker.mock_lock_data(
+        {
+            "package": [
+                {
+                    "name": "torch",
+                    "version": "1.11.0",
+                    "optional": False,
+                    "files": [],
+                    "python-versions": "*",
+                },
+                {
+                    "name": "torch",
+                    "version": "1.11.0+cpu",
+                    "optional": False,
+                    "files": [],
+                    "python-versions": "*",
+                    "source": {
+                        "type": "legacy",
+                        "url": "https://download.pytorch.org/whl",
+                        "reference": "pytorch",
+                    },
+                },
+            ],
+            "metadata": {
+                "python-versions": "*",
+                "platform": "*",
+                "content-hash": "123456789",
+            },
+        }
+    )
+    installer = Installer(
+        NullIO(),
+        MockEnv(platform=env_platform),
+        package,
+        locker,
+        pool,
+        config,
+        installed=installed,
+        executor=Executor(
+            MockEnv(platform=env_platform),
+            pool,
+            config,
+            NullIO(),
+        ),
+    )
+    result = installer.run()
+    assert result == 0
+
+    # Results of installation are consistent with the platform requirements.
+    version = "1.11.0" if env_platform == "darwin" else "1.11.0+cpu"
+    source_type = None if env_platform == "darwin" else "legacy"
+    source_url = (
+        None if env_platform == "darwin" else "https://download.pytorch.org/whl"
+    )
+    source_reference = None if env_platform == "darwin" else "pytorch"
+
+    assert isinstance(installer.executor, Executor)
+    assert len(installer.executor.installations) == 1
+    assert installer.executor.installations[0] == Package(
+        "torch",
+        version,
+        source_type=source_type,
+        source_url=source_url,
+        source_reference=source_reference,
     )
