@@ -96,8 +96,8 @@ def is_decorated():
     if WINDOWS:
         return (
             os.getenv("ANSICON") is not None
-            or "ON" == os.getenv("ConEmuANSI")
-            or "xterm" == os.getenv("Term")
+            or os.getenv("ConEmuANSI") == "ON"
+            or os.getenv("Term") == "xterm"
         )
 
     if not hasattr(sys.stdout, "fileno"):
@@ -213,7 +213,7 @@ def _get_win_folder_with_ctypes(csidl_name):
 
 if WINDOWS:
     try:
-        from ctypes import windll  # noqa
+        from ctypes import windll  # noqa: F401
 
         _get_win_folder = _get_win_folder_with_ctypes
     except ImportError:
@@ -273,7 +273,8 @@ class PoetryInstallationError(RuntimeError):
 class VirtualEnvironment:
     def __init__(self, path: Path) -> None:
         self._path = path
-        # str is required for compatibility with subprocess run on CPython <= 3.7 on Windows
+        # str is required for compatibility with subprocess run on CPython <= 3.7 on
+        # Windows
         self._python = str(
             self._path.joinpath("Scripts/python.exe" if WINDOWS else "bin/python")
         )
@@ -313,7 +314,8 @@ class VirtualEnvironment:
 
         env = cls(target)
 
-        # we do this here to ensure that outdated system default pip does not trigger older bugs
+        # we do this here to ensure that outdated system default pip does not trigger
+        # older bugs
         env.pip("install", "--disable-pip-version-check", "--upgrade", "pip")
 
         return env
@@ -480,7 +482,8 @@ class Installer:
             mx = self.VERSION_REGEX.match(x)
 
             if mx is None:
-                # the version is not semver, perhaps scm or file, we assume upgrade is supported
+                # the version is not semver, perhaps scm or file, we assume upgrade is
+                # supported
                 return True
 
             vx = tuple(int(p) for p in mx.groups()[:3]) + (mx.group(5),)
@@ -490,8 +493,11 @@ class Installer:
             self._write(
                 colorize(
                     "warning",
-                    f"You are installing {version}. When using the current installer, this version does not support "
-                    f"updating using the 'self update' command. Please use 1.1.7 or later.",
+                    (
+                        f"You are installing {version}. When using the current"
+                        " installer, this version does not support updating using the"
+                        " 'self update' command. Please use 1.1.7 or later."
+                    ),
                 )
             )
             if not self._accept_all:
@@ -580,7 +586,7 @@ class Installer:
         try:
             self._install_comment(version, "Creating environment")
             yield VirtualEnvironment.make(env_path)
-        except Exception as e:  # noqa
+        except Exception as e:
             if env_path.exists():
                 self._install_comment(
                     version, "An error occurred. Removing partial environment."
@@ -871,7 +877,7 @@ def main():
     try:
         return installer.run()
     except PoetryInstallationError as e:
-        installer._write(colorize("error", "Poetry installation failed."))  # noqa
+        installer._write(colorize("error", "Poetry installation failed."))
 
         if e.log is not None:
             import traceback
@@ -882,8 +888,11 @@ def main():
                 dir=str(Path.cwd()),
                 text=True,
             )
-            installer._write(colorize("error", f"See {path} for error logs."))  # noqa
-            text = f"{e.log}\nTraceback:\n\n{''.join(traceback.format_tb(e.__traceback__))}"
+            installer._write(colorize("error", f"See {path} for error logs."))
+            text = (
+                f"{e.log}\n"
+                f"Traceback:\n\n{''.join(traceback.format_tb(e.__traceback__))}"
+            )
             Path(path).write_text(text)
 
         return e.return_code
@@ -893,8 +902,11 @@ if __name__ == "__main__":
     sys.stdout.write(
         colorize(
             "warning",
-            "The canonical source for Poetry's installation script is now https://install.python-poetry.org. "
-            "Please update your usage to reflect this.\n",
+            (
+                "The canonical source for Poetry's installation script is now"
+                " https://install.python-poetry.org. Please update your usage to"
+                " reflect this.\n"
+            ),
         )
     )
     sys.exit(main())
