@@ -1,16 +1,36 @@
+from __future__ import annotations
+
+from typing import TYPE_CHECKING
+
 import pytest
 
+from poetry.core.packages.dependency_group import MAIN_GROUP
 from poetry.core.packages.dependency_group import DependencyGroup
+
 from poetry.factory import Factory
+from poetry.utils._compat import tomllib
+from tests.helpers import MOCK_DEFAULT_GIT_REVISION
+from tests.helpers import TestLocker
 from tests.helpers import get_package
 
 
+if TYPE_CHECKING:
+    from cleo.testers.command_tester import CommandTester
+
+    from poetry.poetry import Poetry
+    from poetry.repositories import Repository
+    from tests.helpers import TestRepository
+    from tests.types import CommandTesterFactory
+
+
 @pytest.fixture
-def tester(command_tester_factory):
+def tester(command_tester_factory: CommandTesterFactory) -> CommandTester:
     return command_tester_factory("show")
 
 
-def test_show_basic_with_installed_packages(tester, poetry, installed):
+def test_show_basic_with_installed_packages(
+    tester: CommandTester, poetry: Poetry, installed: Repository
+) -> None:
     poetry.package.add_dependency(Factory.create_dependency("cachy", "^0.1.0"))
     poetry.package.add_dependency(Factory.create_dependency("pendulum", "^2.0.0"))
     poetry.package.add_dependency(
@@ -25,12 +45,12 @@ def test_show_basic_with_installed_packages(tester, poetry, installed):
 
     pytest_373 = get_package("pytest", "3.7.3")
     pytest_373.description = "Pytest package"
-    pytest_373.category = "dev"
 
     installed.add_package(cachy_010)
     installed.add_package(pendulum_200)
     installed.add_package(pytest_373)
 
+    assert isinstance(poetry.locker, TestLocker)
     poetry.locker.mock_lock_data(
         {
             "package": [
@@ -38,7 +58,6 @@ def test_show_basic_with_installed_packages(tester, poetry, installed):
                     "name": "cachy",
                     "version": "0.1.0",
                     "description": "Cachy package",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -48,7 +67,6 @@ def test_show_basic_with_installed_packages(tester, poetry, installed):
                     "name": "pendulum",
                     "version": "2.0.0",
                     "description": "Pendulum package",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -58,7 +76,6 @@ def test_show_basic_with_installed_packages(tester, poetry, installed):
                     "name": "pytest",
                     "version": "3.7.3",
                     "description": "Pytest package",
-                    "category": "dev",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -69,7 +86,7 @@ def test_show_basic_with_installed_packages(tester, poetry, installed):
                 "python-versions": "*",
                 "platform": "*",
                 "content-hash": "123456789",
-                "hashes": {"cachy": [], "pendulum": [], "pytest": []},
+                "files": {"cachy": [], "pendulum": [], "pytest": []},
             },
         }
     )
@@ -82,17 +99,35 @@ pendulum 2.0.0 Pendulum package
 pytest   3.7.3 Pytest package
 """
 
-    assert expected == tester.io.fetch_output()
+    assert tester.io.fetch_output() == expected
 
 
-def test_show_basic_with_installed_packages_single(tester, poetry, installed):
+def _configure_project_with_groups(poetry: Poetry, installed: Repository) -> None:
     poetry.package.add_dependency(Factory.create_dependency("cachy", "^0.1.0"))
+
+    poetry.package.add_dependency_group(DependencyGroup(name="time", optional=True))
+    poetry.package.add_dependency(
+        Factory.create_dependency("pendulum", "^2.0.0", groups=["time"])
+    )
+
+    poetry.package.add_dependency(
+        Factory.create_dependency("pytest", "^3.7.3", groups=["test"])
+    )
 
     cachy_010 = get_package("cachy", "0.1.0")
     cachy_010.description = "Cachy package"
 
-    installed.add_package(cachy_010)
+    pendulum_200 = get_package("pendulum", "2.0.0")
+    pendulum_200.description = "Pendulum package"
 
+    pytest_373 = get_package("pytest", "3.7.3")
+    pytest_373.description = "Pytest package"
+
+    installed.add_package(cachy_010)
+    installed.add_package(pendulum_200)
+    installed.add_package(pytest_373)
+
+    assert isinstance(poetry.locker, TestLocker)
     poetry.locker.mock_lock_data(
         {
             "package": [
@@ -100,7 +135,24 @@ def test_show_basic_with_installed_packages_single(tester, poetry, installed):
                     "name": "cachy",
                     "version": "0.1.0",
                     "description": "Cachy package",
-                    "category": "main",
+                    "optional": False,
+                    "platform": "*",
+                    "python-versions": "*",
+                    "checksum": [],
+                },
+                {
+                    "name": "pendulum",
+                    "version": "2.0.0",
+                    "description": "Pendulum package",
+                    "optional": False,
+                    "platform": "*",
+                    "python-versions": "*",
+                    "checksum": [],
+                },
+                {
+                    "name": "pytest",
+                    "version": "3.7.3",
+                    "description": "Pytest package",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -111,7 +163,132 @@ def test_show_basic_with_installed_packages_single(tester, poetry, installed):
                 "python-versions": "*",
                 "platform": "*",
                 "content-hash": "123456789",
-                "hashes": {"cachy": []},
+                "files": {"cachy": [], "pendulum": [], "pytest": []},
+            },
+        }
+    )
+
+
+@pytest.mark.parametrize(
+    ("options", "expected"),
+    [
+        (
+            "",
+            """\
+cachy  0.1.0 Cachy package
+pytest 3.7.3 Pytest package
+""",
+        ),
+        (
+            "--with time",
+            """\
+cachy    0.1.0 Cachy package
+pendulum 2.0.0 Pendulum package
+pytest   3.7.3 Pytest package
+""",
+        ),
+        (
+            "--without test",
+            """\
+cachy 0.1.0 Cachy package
+""",
+        ),
+        (
+            f"--without {MAIN_GROUP}",
+            """\
+pytest 3.7.3 Pytest package
+""",
+        ),
+        (
+            f"--only {MAIN_GROUP}",
+            """\
+cachy 0.1.0 Cachy package
+""",
+        ),
+        (
+            "--no-dev",
+            """\
+cachy 0.1.0 Cachy package
+""",
+        ),
+        (
+            "--with time --without test",
+            """\
+cachy    0.1.0 Cachy package
+pendulum 2.0.0 Pendulum package
+""",
+        ),
+        (
+            f"--with time --without {MAIN_GROUP},test",
+            """\
+pendulum 2.0.0 Pendulum package
+""",
+        ),
+        (
+            "--only time",
+            """\
+pendulum 2.0.0 Pendulum package
+""",
+        ),
+        (
+            "--only time --with test",
+            """\
+pendulum 2.0.0 Pendulum package
+""",
+        ),
+        (
+            "--with time",
+            """\
+cachy    0.1.0 Cachy package
+pendulum 2.0.0 Pendulum package
+pytest   3.7.3 Pytest package
+""",
+        ),
+    ],
+)
+def test_show_basic_with_group_options(
+    options: str,
+    expected: str,
+    tester: CommandTester,
+    poetry: Poetry,
+    installed: Repository,
+) -> None:
+    _configure_project_with_groups(poetry, installed)
+
+    tester.execute(options)
+
+    assert tester.io.fetch_output() == expected
+
+
+def test_show_basic_with_installed_packages_single(
+    tester: CommandTester, poetry: Poetry, installed: Repository
+) -> None:
+    poetry.package.add_dependency(Factory.create_dependency("cachy", "^0.1.0"))
+
+    cachy_010 = get_package("cachy", "0.1.0")
+    cachy_010.description = "Cachy package"
+
+    installed.add_package(cachy_010)
+
+    assert isinstance(poetry.locker, TestLocker)
+    poetry.locker.mock_lock_data(
+        {
+            "package": [
+                {
+                    "name": "cachy",
+                    "version": "0.1.0",
+                    "description": "Cachy package",
+                    "optional": False,
+                    "platform": "*",
+                    "python-versions": "*",
+                    "checksum": [],
+                },
+            ],
+            "metadata": {
+                "python-versions": "*",
+                "platform": "*",
+                "content-hash": "123456789",
+                "files": {"cachy": []},
             },
         }
     )
@@ -125,37 +302,24 @@ def test_show_basic_with_installed_packages_single(tester, poetry, installed):
     ] == [line.strip() for line in tester.io.fetch_output().splitlines()]
 
 
-def test_show_basic_with_not_installed_packages_non_decorated(
-    tester, poetry, installed
-):
-    poetry.package.add_dependency(Factory.create_dependency("cachy", "^0.1.0"))
-    poetry.package.add_dependency(Factory.create_dependency("pendulum", "^2.0.0"))
+def test_show_basic_with_installed_packages_single_canonicalized(
+    tester: CommandTester, poetry: Poetry, installed: Repository
+) -> None:
+    poetry.package.add_dependency(Factory.create_dependency("foo-bar", "^0.1.0"))
 
-    cachy_010 = get_package("cachy", "0.1.0")
-    cachy_010.description = "Cachy package"
+    foo_bar = get_package("foo-bar", "0.1.0")
+    foo_bar.description = "Foobar package"
 
-    pendulum_200 = get_package("pendulum", "2.0.0")
-    pendulum_200.description = "Pendulum package"
+    installed.add_package(foo_bar)
 
-    installed.add_package(cachy_010)
+    assert isinstance(poetry.locker, TestLocker)
     poetry.locker.mock_lock_data(
         {
             "package": [
                 {
-                    "name": "cachy",
+                    "name": "foo-bar",
                     "version": "0.1.0",
-                    "description": "Cachy package",
-                    "category": "main",
-                    "optional": False,
-                    "platform": "*",
-                    "python-versions": "*",
-                    "checksum": [],
-                },
-                {
-                    "name": "pendulum",
-                    "version": "2.0.0",
-                    "description": "Pendulum package",
-                    "category": "main",
+                    "description": "Foobar package",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -166,7 +330,62 @@ def test_show_basic_with_not_installed_packages_non_decorated(
                 "python-versions": "*",
                 "platform": "*",
                 "content-hash": "123456789",
-                "hashes": {"cachy": [], "pendulum": []},
+                "files": {"foo-bar": []},
+            },
+        }
+    )
+
+    tester.execute("Foo_Bar")
+
+    assert [
+        "name         : foo-bar",
+        "version      : 0.1.0",
+        "description  : Foobar package",
+    ] == [line.strip() for line in tester.io.fetch_output().splitlines()]
+
+
+def test_show_basic_with_not_installed_packages_non_decorated(
+    tester: CommandTester, poetry: Poetry, installed: Repository
+) -> None:
+    poetry.package.add_dependency(Factory.create_dependency("cachy", "^0.1.0"))
+    poetry.package.add_dependency(Factory.create_dependency("pendulum", "^2.0.0"))
+
+    cachy_010 = get_package("cachy", "0.1.0")
+    cachy_010.description = "Cachy package"
+
+    pendulum_200 = get_package("pendulum", "2.0.0")
+    pendulum_200.description = "Pendulum package"
+
+    installed.add_package(cachy_010)
+
+    assert isinstance(poetry.locker, TestLocker)
+    poetry.locker.mock_lock_data(
+        {
+            "package": [
+                {
+                    "name": "cachy",
+                    "version": "0.1.0",
+                    "description": "Cachy package",
+                    "optional": False,
+                    "platform": "*",
+                    "python-versions": "*",
+                    "checksum": [],
+                },
+                {
+                    "name": "pendulum",
+                    "version": "2.0.0",
+                    "description": "Pendulum package",
+                    "optional": False,
+                    "platform": "*",
+                    "python-versions": "*",
+                    "checksum": [],
+                },
+            ],
+            "metadata": {
+                "python-versions": "*",
+                "platform": "*",
+                "content-hash": "123456789",
+                "files": {"cachy": [], "pendulum": []},
             },
         }
     )
@@ -178,10 +397,12 @@ cachy        0.1.0 Cachy package
 pendulum (!) 2.0.0 Pendulum package
 """
 
-    assert expected == tester.io.fetch_output()
+    assert tester.io.fetch_output() == expected
 
 
-def test_show_basic_with_not_installed_packages_decorated(tester, poetry, installed):
+def test_show_basic_with_not_installed_packages_decorated(
+    tester: CommandTester, poetry: Poetry, installed: Repository
+) -> None:
     poetry.package.add_dependency(Factory.create_dependency("cachy", "^0.1.0"))
     poetry.package.add_dependency(Factory.create_dependency("pendulum", "^2.0.0"))
 
@@ -192,6 +413,8 @@ def test_show_basic_with_not_installed_packages_decorated(tester, poetry, instal
     pendulum_200.description = "Pendulum package"
 
     installed.add_package(cachy_010)
+
+    assert isinstance(poetry.locker, TestLocker)
     poetry.locker.mock_lock_data(
         {
             "package": [
@@ -199,7 +422,6 @@ def test_show_basic_with_not_installed_packages_decorated(tester, poetry, instal
                     "name": "cachy",
                     "version": "0.1.0",
                     "description": "Cachy package",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -209,7 +431,6 @@ def test_show_basic_with_not_installed_packages_decorated(tester, poetry, instal
                     "name": "pendulum",
                     "version": "2.0.0",
                     "description": "Pendulum package",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -220,7 +441,7 @@ def test_show_basic_with_not_installed_packages_decorated(tester, poetry, instal
                 "python-versions": "*",
                 "platform": "*",
                 "content-hash": "123456789",
-                "hashes": {"cachy": [], "pendulum": []},
+                "files": {"cachy": [], "pendulum": []},
             },
         }
     )
@@ -232,10 +453,15 @@ def test_show_basic_with_not_installed_packages_decorated(tester, poetry, instal
 \033[31mpendulum\033[39m \033[39;1m2.0.0\033[39;22m Pendulum package
 """
 
-    assert expected == tester.io.fetch_output()
+    assert tester.io.fetch_output() == expected
 
 
-def test_show_latest_non_decorated(tester, poetry, installed, repo):
+def test_show_latest_non_decorated(
+    tester: CommandTester,
+    poetry: Poetry,
+    installed: Repository,
+    repo: TestRepository,
+) -> None:
     poetry.package.add_dependency(Factory.create_dependency("cachy", "^0.1.0"))
     poetry.package.add_dependency(Factory.create_dependency("pendulum", "^2.0.0"))
 
@@ -257,6 +483,7 @@ def test_show_latest_non_decorated(tester, poetry, installed, repo):
     repo.add_package(pendulum_200)
     repo.add_package(pendulum_201)
 
+    assert isinstance(poetry.locker, TestLocker)
     poetry.locker.mock_lock_data(
         {
             "package": [
@@ -264,7 +491,6 @@ def test_show_latest_non_decorated(tester, poetry, installed, repo):
                     "name": "cachy",
                     "version": "0.1.0",
                     "description": "Cachy package",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -274,7 +500,6 @@ def test_show_latest_non_decorated(tester, poetry, installed, repo):
                     "name": "pendulum",
                     "version": "2.0.0",
                     "description": "Pendulum package",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -285,7 +510,7 @@ def test_show_latest_non_decorated(tester, poetry, installed, repo):
                 "python-versions": "*",
                 "platform": "*",
                 "content-hash": "123456789",
-                "hashes": {"cachy": [], "pendulum": []},
+                "files": {"cachy": [], "pendulum": []},
             },
         }
     )
@@ -297,10 +522,15 @@ cachy    0.1.0 0.2.0 Cachy package
 pendulum 2.0.0 2.0.1 Pendulum package
 """
 
-    assert expected == tester.io.fetch_output()
+    assert tester.io.fetch_output() == expected
 
 
-def test_show_latest_decorated(tester, poetry, installed, repo):
+def test_show_latest_decorated(
+    tester: CommandTester,
+    poetry: Poetry,
+    installed: Repository,
+    repo: TestRepository,
+) -> None:
     poetry.package.add_dependency(Factory.create_dependency("cachy", "^0.1.0"))
     poetry.package.add_dependency(Factory.create_dependency("pendulum", "^2.0.0"))
 
@@ -322,6 +552,7 @@ def test_show_latest_decorated(tester, poetry, installed, repo):
     repo.add_package(pendulum_200)
     repo.add_package(pendulum_201)
 
+    assert isinstance(poetry.locker, TestLocker)
     poetry.locker.mock_lock_data(
         {
             "package": [
@@ -329,7 +560,6 @@ def test_show_latest_decorated(tester, poetry, installed, repo):
                     "name": "cachy",
                     "version": "0.1.0",
                     "description": "Cachy package",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -339,7 +569,6 @@ def test_show_latest_decorated(tester, poetry, installed, repo):
                     "name": "pendulum",
                     "version": "2.0.0",
                     "description": "Pendulum package",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -350,7 +579,7 @@ def test_show_latest_decorated(tester, poetry, installed, repo):
                 "python-versions": "*",
                 "platform": "*",
                 "content-hash": "123456789",
-                "hashes": {"cachy": [], "pendulum": []},
+                "files": {"cachy": [], "pendulum": []},
             },
         }
     )
@@ -358,14 +587,21 @@ def test_show_latest_decorated(tester, poetry, installed, repo):
     tester.execute("--latest", decorated=True)
 
     expected = """\
-\033[36mcachy   \033[39m \033[39;1m0.1.0\033[39;22m \033[33m0.2.0\033[39m Cachy package
-\033[36mpendulum\033[39m \033[39;1m2.0.0\033[39;22m \033[31m2.0.1\033[39m Pendulum package
+\033[36mcachy   \033[39m \033[39;1m0.1.0\033[39;22m\
+ \033[33m0.2.0\033[39m Cachy package
+\033[36mpendulum\033[39m \033[39;1m2.0.0\033[39;22m\
+ \033[31m2.0.1\033[39m Pendulum package
 """
 
-    assert expected == tester.io.fetch_output()
+    assert tester.io.fetch_output() == expected
 
 
-def test_show_outdated(tester, poetry, installed, repo):
+def test_show_outdated(
+    tester: CommandTester,
+    poetry: Poetry,
+    installed: Repository,
+    repo: TestRepository,
+) -> None:
     poetry.package.add_dependency(Factory.create_dependency("cachy", "^0.1.0"))
     poetry.package.add_dependency(Factory.create_dependency("pendulum", "^2.0.0"))
 
@@ -384,6 +620,7 @@ def test_show_outdated(tester, poetry, installed, repo):
     repo.add_package(cachy_020)
     repo.add_package(pendulum_200)
 
+    assert isinstance(poetry.locker, TestLocker)
     poetry.locker.mock_lock_data(
         {
             "package": [
@@ -391,7 +628,6 @@ def test_show_outdated(tester, poetry, installed, repo):
                     "name": "cachy",
                     "version": "0.1.0",
                     "description": "Cachy package",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -401,7 +637,6 @@ def test_show_outdated(tester, poetry, installed, repo):
                     "name": "pendulum",
                     "version": "2.0.0",
                     "description": "Pendulum package",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -412,7 +647,7 @@ def test_show_outdated(tester, poetry, installed, repo):
                 "python-versions": "*",
                 "platform": "*",
                 "content-hash": "123456789",
-                "hashes": {"cachy": [], "pendulum": []},
+                "files": {"cachy": [], "pendulum": []},
             },
         }
     )
@@ -423,16 +658,22 @@ def test_show_outdated(tester, poetry, installed, repo):
 cachy 0.1.0 0.2.0 Cachy package
 """
 
-    assert expected == tester.io.fetch_output()
+    assert tester.io.fetch_output() == expected
 
 
-def test_show_outdated_with_only_up_to_date_packages(tester, poetry, installed, repo):
+def test_show_outdated_with_only_up_to_date_packages(
+    tester: CommandTester,
+    poetry: Poetry,
+    installed: Repository,
+    repo: TestRepository,
+) -> None:
     cachy_020 = get_package("cachy", "0.2.0")
     cachy_020.description = "Cachy package"
 
     installed.add_package(cachy_020)
     repo.add_package(cachy_020)
 
+    assert isinstance(poetry.locker, TestLocker)
     poetry.locker.mock_lock_data(
         {
             "package": [
@@ -440,7 +681,6 @@ def test_show_outdated_with_only_up_to_date_packages(tester, poetry, installed, 
                     "name": "cachy",
                     "version": "0.2.0",
                     "description": "Cachy package",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -451,7 +691,7 @@ def test_show_outdated_with_only_up_to_date_packages(tester, poetry, installed, 
                 "python-versions": "*",
                 "platform": "*",
                 "content-hash": "123456789",
-                "hashes": {"cachy": []},
+                "files": {"cachy": []},
             },
         }
     )
@@ -460,10 +700,15 @@ def test_show_outdated_with_only_up_to_date_packages(tester, poetry, installed, 
 
     expected = ""
 
-    assert expected == tester.io.fetch_output()
+    assert tester.io.fetch_output() == expected
 
 
-def test_show_outdated_has_prerelease_but_not_allowed(tester, poetry, installed, repo):
+def test_show_outdated_has_prerelease_but_not_allowed(
+    tester: CommandTester,
+    poetry: Poetry,
+    installed: Repository,
+    repo: TestRepository,
+) -> None:
     poetry.package.add_dependency(Factory.create_dependency("cachy", "^0.1.0"))
     poetry.package.add_dependency(Factory.create_dependency("pendulum", "^2.0.0"))
 
@@ -487,6 +732,7 @@ def test_show_outdated_has_prerelease_but_not_allowed(tester, poetry, installed,
     repo.add_package(cachy_020)
     repo.add_package(pendulum_200)
 
+    assert isinstance(poetry.locker, TestLocker)
     poetry.locker.mock_lock_data(
         {
             "package": [
@@ -494,7 +740,6 @@ def test_show_outdated_has_prerelease_but_not_allowed(tester, poetry, installed,
                     "name": "cachy",
                     "version": "0.1.0",
                     "description": "Cachy package",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -504,7 +749,6 @@ def test_show_outdated_has_prerelease_but_not_allowed(tester, poetry, installed,
                     "name": "pendulum",
                     "version": "2.0.0",
                     "description": "Pendulum package",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -515,7 +759,7 @@ def test_show_outdated_has_prerelease_but_not_allowed(tester, poetry, installed,
                 "python-versions": "*",
                 "platform": "*",
                 "content-hash": "123456789",
-                "hashes": {"cachy": [], "pendulum": []},
+                "files": {"cachy": [], "pendulum": []},
             },
         }
     )
@@ -526,13 +770,18 @@ def test_show_outdated_has_prerelease_but_not_allowed(tester, poetry, installed,
 cachy 0.1.0 0.2.0 Cachy package
 """
 
-    assert expected == tester.io.fetch_output()
+    assert tester.io.fetch_output() == expected
 
 
-def test_show_outdated_has_prerelease_and_allowed(tester, poetry, installed, repo):
+def test_show_outdated_has_prerelease_and_allowed(
+    tester: CommandTester,
+    poetry: Poetry,
+    installed: Repository,
+    repo: TestRepository,
+) -> None:
     poetry.package.add_dependency(
         Factory.create_dependency(
-            "cachy", {"version": "^0.1.0", "allow-prereleases": True}
+            "cachy", {"version": ">=0.0.1", "allow-prereleases": True}
         )
     )
     poetry.package.add_dependency(Factory.create_dependency("pendulum", "^2.0.0"))
@@ -557,6 +806,7 @@ def test_show_outdated_has_prerelease_and_allowed(tester, poetry, installed, rep
     repo.add_package(cachy_020)
     repo.add_package(pendulum_200)
 
+    assert isinstance(poetry.locker, TestLocker)
     poetry.locker.mock_lock_data(
         {
             "package": [
@@ -564,7 +814,6 @@ def test_show_outdated_has_prerelease_and_allowed(tester, poetry, installed, rep
                     "name": "cachy",
                     "version": "0.1.0.dev1",
                     "description": "Cachy package",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -574,7 +823,6 @@ def test_show_outdated_has_prerelease_and_allowed(tester, poetry, installed, rep
                     "name": "pendulum",
                     "version": "2.0.0",
                     "description": "Pendulum package",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -585,7 +833,7 @@ def test_show_outdated_has_prerelease_and_allowed(tester, poetry, installed, rep
                 "python-versions": "*",
                 "platform": "*",
                 "content-hash": "123456789",
-                "hashes": {"cachy": [], "pendulum": []},
+                "files": {"cachy": [], "pendulum": []},
             },
         }
     )
@@ -596,10 +844,15 @@ def test_show_outdated_has_prerelease_and_allowed(tester, poetry, installed, rep
 cachy 0.1.0.dev1 0.3.0.dev123 Cachy package
 """
 
-    assert expected == tester.io.fetch_output()
+    assert tester.io.fetch_output() == expected
 
 
-def test_show_outdated_formatting(tester, poetry, installed, repo):
+def test_show_outdated_formatting(
+    tester: CommandTester,
+    poetry: Poetry,
+    installed: Repository,
+    repo: TestRepository,
+) -> None:
     poetry.package.add_dependency(Factory.create_dependency("cachy", "^0.1.0"))
     poetry.package.add_dependency(Factory.create_dependency("pendulum", "^2.0.0"))
 
@@ -621,6 +874,7 @@ def test_show_outdated_formatting(tester, poetry, installed, repo):
     repo.add_package(pendulum_200)
     repo.add_package(pendulum_201)
 
+    assert isinstance(poetry.locker, TestLocker)
     poetry.locker.mock_lock_data(
         {
             "package": [
@@ -628,7 +882,6 @@ def test_show_outdated_formatting(tester, poetry, installed, repo):
                     "name": "cachy",
                     "version": "0.1.0",
                     "description": "Cachy package",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -638,7 +891,6 @@ def test_show_outdated_formatting(tester, poetry, installed, repo):
                     "name": "pendulum",
                     "version": "2.0.0",
                     "description": "Pendulum package",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -649,7 +901,7 @@ def test_show_outdated_formatting(tester, poetry, installed, repo):
                 "python-versions": "*",
                 "platform": "*",
                 "content-hash": "123456789",
-                "hashes": {"cachy": [], "pendulum": []},
+                "files": {"cachy": [], "pendulum": []},
             },
         }
     )
@@ -661,11 +913,24 @@ cachy    0.1.0 0.2.0 Cachy package
 pendulum 2.0.0 2.0.1 Pendulum package
 """
 
-    assert expected == tester.io.fetch_output()
+    assert tester.io.fetch_output() == expected
 
 
-@pytest.mark.parametrize("project_directory", ["project_with_local_dependencies"])
-def test_show_outdated_local_dependencies(tester, poetry, installed, repo):
+@pytest.mark.parametrize(
+    ("project_directory", "required_fixtures"),
+    [
+        (
+            "project_with_local_dependencies",
+            ["distributions/demo-0.1.0-py2.py3-none-any.whl", "project_with_setup"],
+        ),
+    ],
+)
+def test_show_outdated_local_dependencies(
+    tester: CommandTester,
+    poetry: Poetry,
+    installed: Repository,
+    repo: TestRepository,
+) -> None:
     cachy_010 = get_package("cachy", "0.1.0")
     cachy_010.description = "Cachy package"
     cachy_020 = get_package("cachy", "0.2.0")
@@ -679,18 +944,19 @@ def test_show_outdated_local_dependencies(tester, poetry, installed, repo):
     demo_010 = get_package("demo", "0.1.0")
     demo_010.description = ""
 
-    my_package_012 = get_package("project-with-setup", "0.1.2")
-    my_package_012.description = "Demo project."
+    my_package_011 = get_package("project-with-setup", "0.1.1")
+    my_package_011.description = "Demo project."
 
     installed.add_package(cachy_020)
     installed.add_package(pendulum_200)
     installed.add_package(demo_010)
-    installed.add_package(my_package_012)
+    installed.add_package(my_package_011)
 
     repo.add_package(cachy_020)
     repo.add_package(cachy_030)
     repo.add_package(pendulum_200)
 
+    assert isinstance(poetry.locker, TestLocker)
     poetry.locker.mock_lock_data(
         {
             "package": [
@@ -698,7 +964,6 @@ def test_show_outdated_local_dependencies(tester, poetry, installed, repo):
                     "name": "cachy",
                     "version": "0.2.0",
                     "description": "Cachy package",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -708,7 +973,6 @@ def test_show_outdated_local_dependencies(tester, poetry, installed, repo):
                     "name": "pendulum",
                     "version": "2.0.0",
                     "description": "Pendulum package",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -718,7 +982,6 @@ def test_show_outdated_local_dependencies(tester, poetry, installed, repo):
                     "name": "demo",
                     "version": "0.1.0",
                     "description": "Demo package",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -733,13 +996,12 @@ def test_show_outdated_local_dependencies(tester, poetry, installed, repo):
                     "name": "project-with-setup",
                     "version": "0.1.1",
                     "description": "Demo project.",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
                     "checksum": [],
                     "dependencies": {
-                        "pendulum": "pendulum>=1.4.4",
+                        "pendulum": ">=1.4.4",
                         "cachy": {"version": ">=0.2.0", "extras": ["msgpack"]},
                     },
                     "source": {
@@ -753,7 +1015,7 @@ def test_show_outdated_local_dependencies(tester, poetry, installed, repo):
                 "python-versions": "*",
                 "platform": "*",
                 "content-hash": "123456789",
-                "hashes": {
+                "files": {
                     "cachy": [],
                     "pendulum": [],
                     "demo": [],
@@ -769,13 +1031,19 @@ def test_show_outdated_local_dependencies(tester, poetry, installed, repo):
 cachy              0.2.0                       0.3.0
 project-with-setup 0.1.1 ../project_with_setup 0.1.2 ../project_with_setup
 """
-    assert expected.rstrip() == "\n".join(
-        line.rstrip() for line in tester.io.fetch_output().splitlines()
+    assert (
+        "\n".join(line.rstrip() for line in tester.io.fetch_output().splitlines())
+        == expected.rstrip()
     )
 
 
 @pytest.mark.parametrize("project_directory", ["project_with_git_dev_dependency"])
-def test_show_outdated_git_dev_dependency(tester, poetry, installed, repo):
+def test_show_outdated_git_dev_dependency(
+    tester: CommandTester,
+    poetry: Poetry,
+    installed: Repository,
+    repo: TestRepository,
+) -> None:
     cachy_010 = get_package("cachy", "0.1.0")
     cachy_010.description = "Cachy package"
     cachy_020 = get_package("cachy", "0.2.0")
@@ -800,6 +1068,7 @@ def test_show_outdated_git_dev_dependency(tester, poetry, installed, repo):
     repo.add_package(pendulum_200)
     repo.add_package(pytest)
 
+    assert isinstance(poetry.locker, TestLocker)
     poetry.locker.mock_lock_data(
         {
             "package": [
@@ -807,7 +1076,6 @@ def test_show_outdated_git_dev_dependency(tester, poetry, installed, repo):
                     "name": "cachy",
                     "version": "0.1.0",
                     "description": "Cachy package",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -817,7 +1085,6 @@ def test_show_outdated_git_dev_dependency(tester, poetry, installed, repo):
                     "name": "pendulum",
                     "version": "2.0.0",
                     "description": "Pendulum package",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -827,15 +1094,14 @@ def test_show_outdated_git_dev_dependency(tester, poetry, installed, repo):
                     "name": "demo",
                     "version": "0.1.1",
                     "description": "Demo package",
-                    "category": "dev",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
                     "checksum": [],
                     "source": {
                         "type": "git",
-                        "reference": "9cf87a285a2d3fbb0b9fa621997b3acc3631ed24",
-                        "resolved_reference": "9cf87a285a2d3fbb0b9fa621997b3acc3631ed24",
+                        "reference": MOCK_DEFAULT_GIT_REVISION,
+                        "resolved_reference": MOCK_DEFAULT_GIT_REVISION,
                         "url": "https://github.com/demo/demo.git",
                     },
                 },
@@ -843,7 +1109,6 @@ def test_show_outdated_git_dev_dependency(tester, poetry, installed, repo):
                     "name": "pytest",
                     "version": "3.4.3",
                     "description": "Pytest",
-                    "category": "dev",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -854,7 +1119,7 @@ def test_show_outdated_git_dev_dependency(tester, poetry, installed, repo):
                 "python-versions": "*",
                 "platform": "*",
                 "content-hash": "123456789",
-                "hashes": {"cachy": [], "pendulum": [], "demo": [], "pytest": []},
+                "files": {"cachy": [], "pendulum": [], "demo": [], "pytest": []},
             },
         }
     )
@@ -866,11 +1131,16 @@ cachy 0.1.0         0.2.0         Cachy package
 demo  0.1.1 9cf87a2 0.1.2 9cf87a2 Demo package
 """
 
-    assert expected == tester.io.fetch_output()
+    assert tester.io.fetch_output() == expected
 
 
 @pytest.mark.parametrize("project_directory", ["project_with_git_dev_dependency"])
-def test_show_outdated_no_dev_git_dev_dependency(tester, poetry, installed, repo):
+def test_show_outdated_no_dev_git_dev_dependency(
+    tester: CommandTester,
+    poetry: Poetry,
+    installed: Repository,
+    repo: TestRepository,
+) -> None:
     cachy_010 = get_package("cachy", "0.1.0")
     cachy_010.description = "Cachy package"
     cachy_020 = get_package("cachy", "0.2.0")
@@ -895,6 +1165,7 @@ def test_show_outdated_no_dev_git_dev_dependency(tester, poetry, installed, repo
     repo.add_package(pendulum_200)
     repo.add_package(pytest)
 
+    assert isinstance(poetry.locker, TestLocker)
     poetry.locker.mock_lock_data(
         {
             "package": [
@@ -902,7 +1173,6 @@ def test_show_outdated_no_dev_git_dev_dependency(tester, poetry, installed, repo
                     "name": "cachy",
                     "version": "0.1.0",
                     "description": "Cachy package",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -912,7 +1182,6 @@ def test_show_outdated_no_dev_git_dev_dependency(tester, poetry, installed, repo
                     "name": "pendulum",
                     "version": "2.0.0",
                     "description": "Pendulum package",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -922,14 +1191,13 @@ def test_show_outdated_no_dev_git_dev_dependency(tester, poetry, installed, repo
                     "name": "demo",
                     "version": "0.1.1",
                     "description": "Demo package",
-                    "category": "dev",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
                     "checksum": [],
                     "source": {
                         "type": "git",
-                        "reference": "9cf87a285a2d3fbb0b9fa621997b3acc3631ed24",
+                        "reference": MOCK_DEFAULT_GIT_REVISION,
                         "url": "https://github.com/demo/pyproject-demo.git",
                     },
                 },
@@ -937,7 +1205,6 @@ def test_show_outdated_no_dev_git_dev_dependency(tester, poetry, installed, repo
                     "name": "pytest",
                     "version": "3.4.3",
                     "description": "Pytest",
-                    "category": "dev",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -948,7 +1215,7 @@ def test_show_outdated_no_dev_git_dev_dependency(tester, poetry, installed, repo
                 "python-versions": "*",
                 "platform": "*",
                 "content-hash": "123456789",
-                "hashes": {"cachy": [], "pendulum": [], "demo": [], "pytest": []},
+                "files": {"cachy": [], "pendulum": [], "demo": [], "pytest": []},
             },
         }
     )
@@ -959,10 +1226,15 @@ def test_show_outdated_no_dev_git_dev_dependency(tester, poetry, installed, repo
 cachy 0.1.0 0.2.0 Cachy package
 """
 
-    assert expected == tester.io.fetch_output()
+    assert tester.io.fetch_output() == expected
 
 
-def test_show_hides_incompatible_package(tester, poetry, installed, repo):
+def test_show_hides_incompatible_package(
+    tester: CommandTester,
+    poetry: Poetry,
+    installed: Repository,
+    repo: TestRepository,
+) -> None:
     poetry.package.add_dependency(
         Factory.create_dependency("cachy", {"version": "^0.1.0", "python": "< 2.0"})
     )
@@ -976,6 +1248,7 @@ def test_show_hides_incompatible_package(tester, poetry, installed, repo):
 
     installed.add_package(pendulum_200)
 
+    assert isinstance(poetry.locker, TestLocker)
     poetry.locker.mock_lock_data(
         {
             "package": [
@@ -983,7 +1256,6 @@ def test_show_hides_incompatible_package(tester, poetry, installed, repo):
                     "name": "cachy",
                     "version": "0.1.0",
                     "description": "Cachy package",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -993,7 +1265,6 @@ def test_show_hides_incompatible_package(tester, poetry, installed, repo):
                     "name": "pendulum",
                     "version": "2.0.0",
                     "description": "Pendulum package",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -1004,7 +1275,7 @@ def test_show_hides_incompatible_package(tester, poetry, installed, repo):
                 "python-versions": "*",
                 "platform": "*",
                 "content-hash": "123456789",
-                "hashes": {"cachy": [], "pendulum": []},
+                "files": {"cachy": [], "pendulum": []},
             },
         }
     )
@@ -1015,10 +1286,15 @@ def test_show_hides_incompatible_package(tester, poetry, installed, repo):
 pendulum 2.0.0 Pendulum package
 """
 
-    assert expected == tester.io.fetch_output()
+    assert tester.io.fetch_output() == expected
 
 
-def test_show_all_shows_incompatible_package(tester, poetry, installed, repo):
+def test_show_all_shows_incompatible_package(
+    tester: CommandTester,
+    poetry: Poetry,
+    installed: Repository,
+    repo: TestRepository,
+) -> None:
     cachy_010 = get_package("cachy", "0.1.0")
     cachy_010.description = "Cachy package"
 
@@ -1027,6 +1303,7 @@ def test_show_all_shows_incompatible_package(tester, poetry, installed, repo):
 
     installed.add_package(pendulum_200)
 
+    assert isinstance(poetry.locker, TestLocker)
     poetry.locker.mock_lock_data(
         {
             "package": [
@@ -1034,7 +1311,6 @@ def test_show_all_shows_incompatible_package(tester, poetry, installed, repo):
                     "name": "cachy",
                     "version": "0.1.0",
                     "description": "Cachy package",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -1045,7 +1321,6 @@ def test_show_all_shows_incompatible_package(tester, poetry, installed, repo):
                     "name": "pendulum",
                     "version": "2.0.0",
                     "description": "Pendulum package",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -1056,7 +1331,7 @@ def test_show_all_shows_incompatible_package(tester, poetry, installed, repo):
                 "python-versions": "*",
                 "platform": "*",
                 "content-hash": "123456789",
-                "hashes": {"cachy": [], "pendulum": []},
+                "files": {"cachy": [], "pendulum": []},
             },
         }
     )
@@ -1068,10 +1343,111 @@ cachy     0.1.0 Cachy package
 pendulum  2.0.0 Pendulum package
 """
 
-    assert expected == tester.io.fetch_output()
+    assert tester.io.fetch_output() == expected
 
 
-def test_show_non_dev_with_basic_installed_packages(tester, poetry, installed):
+def test_show_hides_incompatible_package_with_duplicate(
+    tester: CommandTester,
+    poetry: Poetry,
+    installed: Repository,
+    repo: TestRepository,
+) -> None:
+    poetry.package.add_dependency(
+        Factory.create_dependency("cachy", {"version": "0.1.0", "platform": "linux"})
+    )
+    poetry.package.add_dependency(
+        Factory.create_dependency("cachy", {"version": "0.1.1", "platform": "darwin"})
+    )
+
+    assert isinstance(poetry.locker, TestLocker)
+    poetry.locker.mock_lock_data(
+        {
+            "package": [
+                {
+                    "name": "cachy",
+                    "version": "0.1.0",
+                    "description": "Cachy package",
+                    "optional": False,
+                    "platform": "*",
+                    "python-versions": "*",
+                    "files": [],
+                },
+                {
+                    "name": "cachy",
+                    "version": "0.1.1",
+                    "description": "Cachy package",
+                    "optional": False,
+                    "platform": "*",
+                    "python-versions": "*",
+                    "files": [],
+                },
+            ],
+            "metadata": {"content-hash": "123456789"},
+        }
+    )
+
+    tester.execute()
+
+    expected = """\
+cachy (!) 0.1.1 Cachy package
+"""
+
+    assert tester.io.fetch_output() == expected
+
+
+def test_show_all_shows_all_duplicates(
+    tester: CommandTester,
+    poetry: Poetry,
+    installed: Repository,
+    repo: TestRepository,
+) -> None:
+    poetry.package.add_dependency(
+        Factory.create_dependency("cachy", {"version": "0.1.0", "platform": "linux"})
+    )
+    poetry.package.add_dependency(
+        Factory.create_dependency("cachy", {"version": "0.1.1", "platform": "darwin"})
+    )
+
+    assert isinstance(poetry.locker, TestLocker)
+    poetry.locker.mock_lock_data(
+        {
+            "package": [
+                {
+                    "name": "cachy",
+                    "version": "0.1.0",
+                    "description": "Cachy package",
+                    "optional": False,
+                    "platform": "*",
+                    "python-versions": "*",
+                    "files": [],
+                },
+                {
+                    "name": "cachy",
+                    "version": "0.1.1",
+                    "description": "Cachy package",
+                    "optional": False,
+                    "platform": "*",
+                    "python-versions": "*",
+                    "files": [],
+                },
+            ],
+            "metadata": {"content-hash": "123456789"},
+        }
+    )
+
+    tester.execute("--all")
+
+    expected = """\
+cachy     0.1.0 Cachy package
+cachy (!) 0.1.1 Cachy package
+"""
+
+    assert tester.io.fetch_output() == expected
+
+
+def test_show_non_dev_with_basic_installed_packages(
+    tester: CommandTester, poetry: Poetry, installed: Repository
+) -> None:
     poetry.package.add_dependency(Factory.create_dependency("cachy", "^0.1.0"))
     poetry.package.add_dependency(Factory.create_dependency("pendulum", "^2.0.0"))
     poetry.package.add_dependency(
@@ -1086,12 +1462,12 @@ def test_show_non_dev_with_basic_installed_packages(tester, poetry, installed):
 
     pytest_373 = get_package("pytest", "3.7.3")
     pytest_373.description = "Pytest package"
-    pytest_373.category = "dev"
 
     installed.add_package(cachy_010)
     installed.add_package(pendulum_200)
     installed.add_package(pytest_373)
 
+    assert isinstance(poetry.locker, TestLocker)
     poetry.locker.mock_lock_data(
         {
             "package": [
@@ -1099,7 +1475,6 @@ def test_show_non_dev_with_basic_installed_packages(tester, poetry, installed):
                     "name": "cachy",
                     "version": "0.1.0",
                     "description": "Cachy package",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -1109,7 +1484,6 @@ def test_show_non_dev_with_basic_installed_packages(tester, poetry, installed):
                     "name": "pendulum",
                     "version": "2.0.0",
                     "description": "Pendulum package",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -1119,7 +1493,6 @@ def test_show_non_dev_with_basic_installed_packages(tester, poetry, installed):
                     "name": "pytest",
                     "version": "3.7.3",
                     "description": "Pytest package",
-                    "category": "dev",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -1130,7 +1503,7 @@ def test_show_non_dev_with_basic_installed_packages(tester, poetry, installed):
                 "python-versions": "*",
                 "platform": "*",
                 "content-hash": "123456789",
-                "hashes": {"cachy": [], "pendulum": [], "pytest": []},
+                "files": {"cachy": [], "pendulum": [], "pytest": []},
             },
         }
     )
@@ -1142,10 +1515,12 @@ cachy    0.1.0 Cachy package
 pendulum 2.0.0 Pendulum package
 """
 
-    assert expected == tester.io.fetch_output()
+    assert tester.io.fetch_output() == expected
 
 
-def test_show_with_group_only(tester, poetry, installed):
+def test_show_with_group_only(
+    tester: CommandTester, poetry: Poetry, installed: Repository
+) -> None:
     poetry.package.add_dependency(Factory.create_dependency("cachy", "^0.1.0"))
     poetry.package.add_dependency(Factory.create_dependency("pendulum", "^2.0.0"))
     poetry.package.add_dependency(
@@ -1160,12 +1535,12 @@ def test_show_with_group_only(tester, poetry, installed):
 
     pytest_373 = get_package("pytest", "3.7.3")
     pytest_373.description = "Pytest package"
-    pytest_373.category = "dev"
 
     installed.add_package(cachy_010)
     installed.add_package(pendulum_200)
     installed.add_package(pytest_373)
 
+    assert isinstance(poetry.locker, TestLocker)
     poetry.locker.mock_lock_data(
         {
             "package": [
@@ -1173,7 +1548,6 @@ def test_show_with_group_only(tester, poetry, installed):
                     "name": "cachy",
                     "version": "0.1.0",
                     "description": "Cachy package",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -1183,7 +1557,6 @@ def test_show_with_group_only(tester, poetry, installed):
                     "name": "pendulum",
                     "version": "2.0.0",
                     "description": "Pendulum package",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -1193,7 +1566,6 @@ def test_show_with_group_only(tester, poetry, installed):
                     "name": "pytest",
                     "version": "3.7.3",
                     "description": "Pytest package",
-                    "category": "dev",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -1204,7 +1576,7 @@ def test_show_with_group_only(tester, poetry, installed):
                 "python-versions": "*",
                 "platform": "*",
                 "content-hash": "123456789",
-                "hashes": {"cachy": [], "pendulum": [], "pytest": []},
+                "files": {"cachy": [], "pendulum": [], "pytest": []},
             },
         }
     )
@@ -1215,10 +1587,12 @@ def test_show_with_group_only(tester, poetry, installed):
 pytest 3.7.3 Pytest package
 """
 
-    assert expected == tester.io.fetch_output()
+    assert tester.io.fetch_output() == expected
 
 
-def test_show_with_optional_group(tester, poetry, installed):
+def test_show_with_optional_group(
+    tester: CommandTester, poetry: Poetry, installed: Repository
+) -> None:
     poetry.package.add_dependency(Factory.create_dependency("cachy", "^0.1.0"))
     poetry.package.add_dependency(Factory.create_dependency("pendulum", "^2.0.0"))
     group = DependencyGroup("dev", optional=True)
@@ -1233,12 +1607,12 @@ def test_show_with_optional_group(tester, poetry, installed):
 
     pytest_373 = get_package("pytest", "3.7.3")
     pytest_373.description = "Pytest package"
-    pytest_373.category = "dev"
 
     installed.add_package(cachy_010)
     installed.add_package(pendulum_200)
     installed.add_package(pytest_373)
 
+    assert isinstance(poetry.locker, TestLocker)
     poetry.locker.mock_lock_data(
         {
             "package": [
@@ -1246,7 +1620,6 @@ def test_show_with_optional_group(tester, poetry, installed):
                     "name": "cachy",
                     "version": "0.1.0",
                     "description": "Cachy package",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -1256,7 +1629,6 @@ def test_show_with_optional_group(tester, poetry, installed):
                     "name": "pendulum",
                     "version": "2.0.0",
                     "description": "Pendulum package",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -1266,7 +1638,6 @@ def test_show_with_optional_group(tester, poetry, installed):
                     "name": "pytest",
                     "version": "3.7.3",
                     "description": "Pytest package",
-                    "category": "dev",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -1277,7 +1648,7 @@ def test_show_with_optional_group(tester, poetry, installed):
                 "python-versions": "*",
                 "platform": "*",
                 "content-hash": "123456789",
-                "hashes": {"cachy": [], "pendulum": [], "pytest": []},
+                "files": {"cachy": [], "pendulum": [], "pytest": []},
             },
         }
     )
@@ -1289,7 +1660,7 @@ cachy    0.1.0 Cachy package
 pendulum 2.0.0 Pendulum package
 """
 
-    assert expected == tester.io.fetch_output()
+    assert tester.io.fetch_output() == expected
 
     tester.execute("--with dev")
 
@@ -1299,10 +1670,12 @@ pendulum 2.0.0 Pendulum package
 pytest   3.7.3 Pytest package
 """
 
-    assert expected == tester.io.fetch_output()
+    assert tester.io.fetch_output() == expected
 
 
-def test_show_tree(tester, poetry, installed):
+def test_show_tree(
+    tester: CommandTester, poetry: Poetry, installed: Repository
+) -> None:
     poetry.package.add_dependency(Factory.create_dependency("cachy", "^0.2.0"))
 
     cachy2 = get_package("cachy", "0.2.0")
@@ -1310,6 +1683,7 @@ def test_show_tree(tester, poetry, installed):
 
     installed.add_package(cachy2)
 
+    assert isinstance(poetry.locker, TestLocker)
     poetry.locker.mock_lock_data(
         {
             "package": [
@@ -1317,7 +1691,6 @@ def test_show_tree(tester, poetry, installed):
                     "name": "cachy",
                     "version": "0.2.0",
                     "description": "",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -1328,7 +1701,6 @@ def test_show_tree(tester, poetry, installed):
                     "name": "msgpack-python",
                     "version": "0.5.1",
                     "description": "",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -1339,7 +1711,7 @@ def test_show_tree(tester, poetry, installed):
                 "python-versions": "*",
                 "platform": "*",
                 "content-hash": "123456789",
-                "hashes": {"cachy": [], "msgpack-python": []},
+                "files": {"cachy": [], "msgpack-python": []},
             },
         }
     )
@@ -1351,10 +1723,12 @@ cachy 0.2.0
 `-- msgpack-python >=0.5 <0.6
 """
 
-    assert expected == tester.io.fetch_output()
+    assert tester.io.fetch_output() == expected
 
 
-def test_show_tree_no_dev(tester, poetry, installed):
+def test_show_tree_no_dev(
+    tester: CommandTester, poetry: Poetry, installed: Repository
+) -> None:
     poetry.package.add_dependency(Factory.create_dependency("cachy", "^0.2.0"))
     poetry.package.add_dependency(
         Factory.create_dependency("pytest", "^6.1.0", groups=["dev"])
@@ -1367,6 +1741,7 @@ def test_show_tree_no_dev(tester, poetry, installed):
     pytest = get_package("pytest", "6.1.1")
     installed.add_package(pytest)
 
+    assert isinstance(poetry.locker, TestLocker)
     poetry.locker.mock_lock_data(
         {
             "package": [
@@ -1374,7 +1749,6 @@ def test_show_tree_no_dev(tester, poetry, installed):
                     "name": "cachy",
                     "version": "0.2.0",
                     "description": "",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -1385,7 +1759,6 @@ def test_show_tree_no_dev(tester, poetry, installed):
                     "name": "msgpack-python",
                     "version": "0.5.1",
                     "description": "",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -1395,7 +1768,6 @@ def test_show_tree_no_dev(tester, poetry, installed):
                     "name": "pytest",
                     "version": "6.1.1",
                     "description": "",
-                    "category": "dev",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -1406,7 +1778,7 @@ def test_show_tree_no_dev(tester, poetry, installed):
                 "python-versions": "*",
                 "platform": "*",
                 "content-hash": "123456789",
-                "hashes": {"cachy": [], "msgpack-python": [], "pytest": []},
+                "files": {"cachy": [], "msgpack-python": [], "pytest": []},
             },
         }
     )
@@ -1418,10 +1790,131 @@ cachy 0.2.0
 └── msgpack-python >=0.5 <0.6
 """
 
-    assert expected == tester.io.fetch_output()
+    assert tester.io.fetch_output() == expected
 
 
-def test_show_required_by_deps(tester, poetry, installed):
+def test_show_tree_why_package(
+    tester: CommandTester, poetry: Poetry, installed: Repository
+) -> None:
+    poetry.package.add_dependency(Factory.create_dependency("a", "=0.0.1"))
+
+    a = get_package("a", "0.0.1")
+    installed.add_package(a)
+    a.add_dependency(Factory.create_dependency("b", "=0.0.1"))
+
+    b = get_package("b", "0.0.1")
+    a.add_dependency(Factory.create_dependency("c", "=0.0.1"))
+    installed.add_package(b)
+
+    c = get_package("c", "0.0.1")
+    installed.add_package(c)
+
+    assert isinstance(poetry.locker, TestLocker)
+    poetry.locker.mock_lock_data(
+        {
+            "package": [
+                {
+                    "name": "a",
+                    "version": "0.0.1",
+                    "dependencies": {"b": "=0.0.1"},
+                    "python-versions": "*",
+                    "optional": False,
+                },
+                {
+                    "name": "b",
+                    "version": "0.0.1",
+                    "dependencies": {"c": "=0.0.1"},
+                    "python-versions": "*",
+                    "optional": False,
+                },
+                {
+                    "name": "c",
+                    "version": "0.0.1",
+                    "python-versions": "*",
+                    "optional": False,
+                },
+            ],
+            "metadata": {
+                "python-versions": "*",
+                "platform": "*",
+                "content-hash": "123456789",
+                "files": {"a": [], "b": [], "c": []},
+            },
+        }
+    )
+
+    tester.execute("--tree --why b")
+
+    expected = """\
+a 0.0.1
+└── b =0.0.1
+    └── c =0.0.1 \n"""
+
+    assert tester.io.fetch_output() == expected
+
+
+def test_show_tree_why(
+    tester: CommandTester, poetry: Poetry, installed: Repository
+) -> None:
+    poetry.package.add_dependency(Factory.create_dependency("a", "=0.0.1"))
+
+    a = get_package("a", "0.0.1")
+    installed.add_package(a)
+    a.add_dependency(Factory.create_dependency("b", "=0.0.1"))
+
+    b = get_package("b", "0.0.1")
+    a.add_dependency(Factory.create_dependency("c", "=0.0.1"))
+    installed.add_package(b)
+
+    c = get_package("c", "0.0.1")
+    installed.add_package(c)
+
+    assert isinstance(poetry.locker, TestLocker)
+    poetry.locker.mock_lock_data(
+        {
+            "package": [
+                {
+                    "name": "a",
+                    "version": "0.0.1",
+                    "dependencies": {"b": "=0.0.1"},
+                    "python-versions": "*",
+                    "optional": False,
+                },
+                {
+                    "name": "b",
+                    "version": "0.0.1",
+                    "dependencies": {"c": "=0.0.1"},
+                    "python-versions": "*",
+                    "optional": False,
+                },
+                {
+                    "name": "c",
+                    "version": "0.0.1",
+                    "python-versions": "*",
+                    "optional": False,
+                },
+            ],
+            "metadata": {
+                "python-versions": "*",
+                "platform": "*",
+                "content-hash": "123456789",
+                "files": {"a": [], "b": [], "c": []},
+            },
+        }
+    )
+
+    tester.execute("--why")
+
+    # this has to be on a single line due to the padding whitespace, which gets stripped
+    # by pre-commit.
+    expected = """a 0.0.1        \nb 0.0.1 from a \nc 0.0.1 from b \n"""
+
+    assert tester.io.fetch_output() == expected
+
+
+def test_show_required_by_deps(
+    tester: CommandTester, poetry: Poetry, installed: Repository
+) -> None:
     poetry.package.add_dependency(Factory.create_dependency("cachy", "^0.2.0"))
     poetry.package.add_dependency(Factory.create_dependency("pendulum", "2.0.0"))
 
@@ -1434,6 +1927,7 @@ def test_show_required_by_deps(tester, poetry, installed):
     installed.add_package(cachy2)
     installed.add_package(pendulum)
 
+    assert isinstance(poetry.locker, TestLocker)
     poetry.locker.mock_lock_data(
         {
             "package": [
@@ -1441,7 +1935,6 @@ def test_show_required_by_deps(tester, poetry, installed):
                     "name": "cachy",
                     "version": "0.2.0",
                     "description": "",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -1452,7 +1945,6 @@ def test_show_required_by_deps(tester, poetry, installed):
                     "name": "pendulum",
                     "version": "2.0.0",
                     "description": "Pendulum package",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -1463,7 +1955,6 @@ def test_show_required_by_deps(tester, poetry, installed):
                     "name": "msgpack-python",
                     "version": "0.5.1",
                     "description": "",
-                    "category": "main",
                     "optional": False,
                     "platform": "*",
                     "python-versions": "*",
@@ -1474,7 +1965,7 @@ def test_show_required_by_deps(tester, poetry, installed):
                 "python-versions": "*",
                 "platform": "*",
                 "content-hash": "123456789",
-                "hashes": {"cachy": [], "pendulum": [], "msgpack-python": []},
+                "files": {"cachy": [], "pendulum": [], "msgpack-python": []},
             },
         }
     )
@@ -1494,3 +1985,284 @@ required by
 """.splitlines()
     actual = [line.rstrip() for line in tester.io.fetch_output().splitlines()]
     assert actual == expected
+
+
+def test_show_errors_without_lock_file(tester: CommandTester, poetry: Poetry) -> None:
+    assert not poetry.locker.lock.exists()
+
+    tester.execute()
+
+    expected = "Error: poetry.lock not found. Run `poetry lock` to create it.\n"
+    assert tester.io.fetch_error() == expected
+    assert tester.status_code == 1
+
+
+def test_show_dependency_installed_from_git_in_dev(
+    tester: CommandTester,
+    poetry: Poetry,
+    installed: Repository,
+    repo: TestRepository,
+) -> None:
+    # Add a regular dependency for a package in main, and a git dependency for the same
+    # package in dev.
+    poetry.package.add_dependency(Factory.create_dependency("demo", "^0.1.1"))
+    poetry.package.add_dependency(
+        Factory.create_dependency(
+            "demo", {"git": "https://github.com/demo/demo.git"}, groups=["dev"]
+        )
+    )
+
+    demo_011 = get_package("demo", "0.1.1")
+    demo_011.description = "Demo package"
+    repo.add_package(demo_011)
+
+    pendulum_200 = get_package("pendulum", "2.0.0")
+    pendulum_200.description = "Pendulum package"
+    repo.add_package(pendulum_200)
+
+    # The git package is the one that gets into the lockfile.
+    assert isinstance(poetry.locker, TestLocker)
+    poetry.locker.mock_lock_data(
+        {
+            "package": [
+                {
+                    "name": "demo",
+                    "version": "0.1.2",
+                    "description": "Demo package",
+                    "optional": False,
+                    "python-versions": "*",
+                    "develop": False,
+                    "source": {
+                        "type": "git",
+                        "reference": MOCK_DEFAULT_GIT_REVISION,
+                        "resolved_reference": MOCK_DEFAULT_GIT_REVISION,
+                        "url": "https://github.com/demo/demo.git",
+                    },
+                },
+                {
+                    "name": "pendulum",
+                    "version": "2.0.0",
+                    "description": "Pendulum package",
+                    "optional": False,
+                    "platform": "*",
+                    "python-versions": "*",
+                    "checksum": [],
+                },
+            ],
+            "metadata": {
+                "python-versions": "*",
+                "platform": "*",
+                "content-hash": "123456789",
+                "files": {"demo": [], "pendulum": []},
+            },
+        }
+    )
+
+    # Nothing needs updating, there is no confusion between the git and not-git
+    # packages.
+    tester.execute("--outdated")
+    assert tester.io.fetch_output() == ""
+
+
+def test_url_dependency_is_not_outdated_by_repository_package(
+    tester: CommandTester,
+    poetry: Poetry,
+    installed: Repository,
+    repo: TestRepository,
+) -> None:
+    demo_url = "https://python-poetry.org/distributions/demo-0.1.0-py2.py3-none-any.whl"
+    poetry.package.add_dependency(
+        Factory.create_dependency(
+            "demo",
+            {"url": demo_url},
+        )
+    )
+
+    # A newer version of demo is available in the repository.
+    demo_100 = get_package("demo", "1.0.0")
+    repo.add_package(demo_100)
+
+    assert isinstance(poetry.locker, TestLocker)
+    poetry.locker.mock_lock_data(
+        {
+            "package": [
+                {
+                    "name": "demo",
+                    "version": "0.1.0",
+                    "description": "Demo package",
+                    "optional": False,
+                    "platform": "*",
+                    "python-versions": "*",
+                    "checksum": [],
+                    "source": {
+                        "type": "url",
+                        "url": demo_url,
+                    },
+                }
+            ],
+            "metadata": {
+                "python-versions": "*",
+                "platform": "*",
+                "content-hash": "123456789",
+                "hashes": {"demo": []},
+            },
+        }
+    )
+
+    # The url dependency on demo is not made outdated by the existence of a newer
+    # version in the repository.
+    tester.execute("--outdated")
+    assert tester.io.fetch_output() == ""
+
+
+def test_show_top_level(
+    tester: CommandTester, poetry: Poetry, installed: Repository
+) -> None:
+    poetry.package.add_dependency(Factory.create_dependency("cachy", "^0.2.0"))
+
+    cachy2 = get_package("cachy", "0.2.0")
+    cachy2.add_dependency(Factory.create_dependency("msgpack-python", ">=0.5 <0.6"))
+
+    installed.add_package(cachy2)
+
+    assert isinstance(poetry.locker, TestLocker)
+    poetry.locker.mock_lock_data(
+        {
+            "package": [
+                {
+                    "name": "cachy",
+                    "version": "0.2.0",
+                    "description": "",
+                    "category": "main",
+                    "optional": False,
+                    "platform": "*",
+                    "python-versions": "*",
+                    "checksum": [],
+                    "dependencies": {"msgpack-python": ">=0.5 <0.6"},
+                },
+                {
+                    "name": "msgpack-python",
+                    "version": "0.5.1",
+                    "description": "",
+                    "category": "main",
+                    "optional": False,
+                    "platform": "*",
+                    "python-versions": "*",
+                    "checksum": [],
+                },
+            ],
+            "metadata": {
+                "python-versions": "*",
+                "platform": "*",
+                "content-hash": "123456789",
+                "files": {"cachy": [], "msgpack-python": []},
+            },
+        }
+    )
+
+    tester.execute("--top-level")
+
+    expected = """cachy              0.2.0 \n"""
+
+    assert tester.io.fetch_output() == expected
+
+
+def test_show_top_level_with_explicitly_defined_depenancy(
+    tester: CommandTester, poetry: Poetry, installed: Repository
+) -> None:
+    poetry.package.add_dependency(Factory.create_dependency("a", "^0.1.0"))
+    poetry.package.add_dependency(Factory.create_dependency("b", "^0.2.0"))
+
+    a = get_package("a", "0.1.0")
+    a.add_dependency(Factory.create_dependency("b", "0.2.0"))
+    b = get_package("b", "0.2.0")
+
+    installed.add_package(a)
+    installed.add_package(b)
+
+    assert isinstance(poetry.locker, TestLocker)
+    poetry.locker.mock_lock_data(
+        {
+            "package": [
+                {
+                    "name": "a",
+                    "version": "0.1.0",
+                    "description": "",
+                    "category": "main",
+                    "optional": False,
+                    "platform": "*",
+                    "python-versions": "*",
+                    "checksum": [],
+                    "dependencies": {"b": "0.2.0"},
+                },
+                {
+                    "name": "b",
+                    "version": "0.2.0",
+                    "description": "",
+                    "category": "main",
+                    "optional": False,
+                    "platform": "*",
+                    "python-versions": "*",
+                    "checksum": [],
+                },
+            ],
+            "metadata": {
+                "python-versions": "*",
+                "platform": "*",
+                "content-hash": "123456789",
+                "files": {"a": [], "b": []},
+            },
+        }
+    )
+
+    tester.execute("--top-level")
+
+    expected = """a 0.1.0 \nb 0.2.0 \n"""
+
+    assert tester.io.fetch_output() == expected
+
+
+def test_show_error_top_level_with_tree(tester: CommandTester) -> None:
+    expected = "Error: Cannot use --tree and --top-level at the same time.\n"
+    tester.execute("--top-level --tree")
+    assert tester.io.fetch_error() == expected
+    assert tester.status_code == 1
+
+
+def test_show_error_top_level_with_single_package(tester: CommandTester) -> None:
+    expected = "Error: Cannot use --top-level when displaying a single package.\n"
+    tester.execute("--top-level some_package_name")
+    assert tester.io.fetch_error() == expected
+    assert tester.status_code == 1
+
+
+@pytest.mark.parametrize(
+    ("project_directory", "required_fixtures"),
+    [
+        (
+            "deleted_directory_dependency",
+            [],
+        ),
+    ],
+)
+def test_show_outdated_missing_directory_dependency(
+    tester: CommandTester,
+    poetry: Poetry,
+    installed: Repository,
+    repo: TestRepository,
+) -> None:
+    with (poetry.pyproject.file.path.parent / "poetry.lock").open(mode="rb") as f:
+        data = tomllib.load(f)
+
+    assert isinstance(poetry.locker, TestLocker)
+    poetry.locker.mock_lock_data(data)
+
+    poetry.package.add_dependency(
+        Factory.create_dependency(
+            "missing",
+            {"path": data["package"][0]["source"]["url"]},
+        )
+    )
+
+    with pytest.raises(ValueError, match="does not exist"):
+        tester.execute("")
