@@ -1,84 +1,120 @@
-from pathlib import Path
+from __future__ import annotations
 
-from packaging.tags import Tag
+import os
+import tempfile
+
+from pathlib import Path
+from typing import TYPE_CHECKING
+from zipfile import ZipFile
+
+import pytest
 
 from poetry.core.packages.utils.link import Link
+
+from poetry.factory import Factory
 from poetry.installation.chef import Chef
-from poetry.utils.env import MockEnv
+from poetry.repositories import RepositoryPool
+from poetry.utils.env import EnvManager
+from tests.repositories.test_pypi_repository import MockRepository
 
 
-def test_get_cached_archive_for_link(config, mocker):
+if TYPE_CHECKING:
+    from pytest_mock import MockerFixture
+
+    from poetry.utils.cache import ArtifactCache
+    from tests.conftest import Config
+    from tests.types import FixtureDirGetter
+
+
+@pytest.fixture()
+def pool() -> RepositoryPool:
+    pool = RepositoryPool()
+
+    pool.add_repository(MockRepository())
+
+    return pool
+
+
+@pytest.fixture(autouse=True)
+def setup(mocker: MockerFixture, pool: RepositoryPool) -> None:
+    mocker.patch.object(Factory, "create_pool", return_value=pool)
+
+
+def test_prepare_sdist(
+    config: Config,
+    config_cache_dir: Path,
+    artifact_cache: ArtifactCache,
+    fixture_dir: FixtureDirGetter,
+) -> None:
     chef = Chef(
-        config,
-        MockEnv(
-            version_info=(3, 8, 3),
-            marker_env={"interpreter_name": "cpython", "interpreter_version": "3.8.3"},
-            supported_tags=[
-                Tag("cp38", "cp38", "macosx_10_15_x86_64"),
-                Tag("py3", "none", "any"),
-            ],
-        ),
+        artifact_cache, EnvManager.get_system_env(), Factory.create_pool(config)
     )
+    archive = (fixture_dir("distributions") / "demo-0.1.0.tar.gz").resolve()
+    destination = artifact_cache.get_cache_directory_for_link(Link(archive.as_uri()))
 
-    mocker.patch.object(
-        chef,
-        "get_cached_archives_for_link",
-        return_value=[
-            Link("file:///foo/demo-0.1.0-py2.py3-none-any"),
-            Link("file:///foo/demo-0.1.0.tar.gz"),
-            Link("file:///foo/demo-0.1.0-cp38-cp38-macosx_10_15_x86_64.whl"),
-            Link("file:///foo/demo-0.1.0-cp37-cp37-macosx_10_15_x86_64.whl"),
-        ],
-    )
+    wheel = chef.prepare(archive)
 
-    archive = chef.get_cached_archive_for_link(
-        Link("https://files.python-poetry.org/demo-0.1.0.tar.gz")
-    )
-
-    assert Link("file:///foo/demo-0.1.0-cp38-cp38-macosx_10_15_x86_64.whl") == archive
+    assert wheel.parent == destination
+    assert wheel.name == "demo-0.1.0-py3-none-any.whl"
 
 
-def test_get_cached_archives_for_link(config, mocker):
+def test_prepare_directory(
+    config: Config,
+    config_cache_dir: Path,
+    artifact_cache: ArtifactCache,
+    fixture_dir: FixtureDirGetter,
+) -> None:
     chef = Chef(
-        config,
-        MockEnv(
-            marker_env={"interpreter_name": "cpython", "interpreter_version": "3.8.3"}
-        ),
+        artifact_cache, EnvManager.get_system_env(), Factory.create_pool(config)
     )
+    archive = fixture_dir("simple_project").resolve()
 
-    distributions = Path(__file__).parent.parent.joinpath("fixtures/distributions")
-    mocker.patch.object(
-        chef,
-        "get_cache_directory_for_link",
-        return_value=distributions,
-    )
+    wheel = chef.prepare(archive)
 
-    archives = chef.get_cached_archives_for_link(
-        Link("https://files.python-poetry.org/demo-0.1.0.tar.gz")
-    )
+    assert wheel.name == "simple_project-1.2.3-py2.py3-none-any.whl"
 
-    assert archives
-    assert set(archives) == {
-        Link(path.as_uri()) for path in distributions.glob("demo-0.1.0*")
-    }
+    assert wheel.parent.parent == Path(tempfile.gettempdir())
+    # cleanup generated tmp dir artifact
+    os.unlink(wheel)
 
 
-def test_get_cache_directory_for_link(config, config_cache_dir):
+def test_prepare_directory_with_extensions(
+    config: Config,
+    config_cache_dir: Path,
+    artifact_cache: ArtifactCache,
+    fixture_dir: FixtureDirGetter,
+) -> None:
+    env = EnvManager.get_system_env()
+    chef = Chef(artifact_cache, env, Factory.create_pool(config))
+    archive = fixture_dir("extended_with_no_setup").resolve()
+
+    wheel = chef.prepare(archive)
+
+    assert wheel.parent.parent == Path(tempfile.gettempdir())
+    assert wheel.name == f"extended-0.1-{env.supported_tags[0]}.whl"
+
+    # cleanup generated tmp dir artifact
+    os.unlink(wheel)
+
+
+def test_prepare_directory_editable(
+    config: Config,
+    config_cache_dir: Path,
+    artifact_cache: ArtifactCache,
+    fixture_dir: FixtureDirGetter,
+) -> None:
     chef = Chef(
-        config,
-        MockEnv(
-            marker_env={"interpreter_name": "cpython", "interpreter_version": "3.8.3"}
-        ),
+        artifact_cache, EnvManager.get_system_env(), Factory.create_pool(config)
     )
+    archive = fixture_dir("simple_project").resolve()
 
-    directory = chef.get_cache_directory_for_link(
-        Link("https://files.python-poetry.org/poetry-1.1.0.tar.gz")
-    )
+    wheel = chef.prepare(archive, editable=True)
 
-    expected = Path(
-        "{}/artifacts/ba/63/13/283a3b3b7f95f05e9e6f84182d276f7bb0951d5b0cc24422b33f7a4648".format(
-            config_cache_dir.as_posix()
-        )
-    )
+    assert wheel.parent.parent == Path(tempfile.gettempdir())
+    assert wheel.name == "simple_project-1.2.3-py2.py3-none-any.whl"
 
-    assert expected == directory
+    with ZipFile(wheel) as z:
+        assert "simple_project.pth" in z.namelist()
+
+    # cleanup generated tmp dir artifact
+    os.unlink(wheel)
