@@ -1,12 +1,35 @@
+from __future__ import annotations
+
+from typing import TYPE_CHECKING
+
 from poetry.core.packages.package import Package
+
 from poetry.factory import Factory
 from poetry.mixology.failure import SolveFailure
 from poetry.mixology.version_solver import VersionSolver
-from poetry.packages import DependencyPackage
 
 
-def add_to_repo(repository, name, version, deps=None, python=None):
-    package = Package(name, version)
+if TYPE_CHECKING:
+    from collections.abc import Mapping
+
+    from packaging.utils import NormalizedName
+    from poetry.core.factory import DependencyConstraint
+    from poetry.core.packages.project_package import ProjectPackage
+
+    from poetry.mixology.result import SolverResult
+    from poetry.repositories import Repository
+    from tests.mixology.version_solver.conftest import Provider
+
+
+def add_to_repo(
+    repository: Repository,
+    name: str,
+    version: str,
+    deps: Mapping[str, DependencyConstraint] | None = None,
+    python: str | None = None,
+    yanked: bool = False,
+) -> None:
+    package = Package(name, version, yanked=yanked)
     if python:
         package.python_versions = python
 
@@ -18,31 +41,39 @@ def add_to_repo(repository, name, version, deps=None, python=None):
 
 
 def check_solver_result(
-    root, provider, result=None, error=None, tries=None, locked=None, use_latest=None
-):
-    if locked is not None:
-        locked = {k: DependencyPackage(l.to_dependency(), l) for k, l in locked.items()}
+    root: ProjectPackage,
+    provider: Provider,
+    result: dict[str, str] | None = None,
+    error: str | None = None,
+    tries: int | None = None,
+    use_latest: list[NormalizedName] | None = None,
+) -> SolverResult | None:
+    solver = VersionSolver(root, provider)
+    with provider.use_latest_for(use_latest or []):
+        try:
+            solution = solver.solve()
+        except SolveFailure as e:
+            if error:
+                assert str(e) == error
 
-    solver = VersionSolver(root, provider, locked=locked, use_latest=use_latest)
+                if tries is not None:
+                    assert solver.solution.attempted_solutions == tries
 
-    try:
-        solution = solver.solve()
-    except SolveFailure as e:
-        if error:
-            assert str(e) == error
+            return None
 
-            if tries is not None:
-                assert solver.solution.attempted_solutions == tries
-
-            return
-
-        raise
+        except AssertionError as e:
+            if error:
+                assert str(e) == error
+                return None
+            raise
 
     packages = {}
     for package in solution.packages:
         packages[package.name] = str(package.version)
 
-    assert result == packages
+    assert packages == result
 
     if tries is not None:
         assert solution.attempted_solutions == tries
+
+    return solution
