@@ -1,15 +1,27 @@
+from __future__ import annotations
+
+from typing import TYPE_CHECKING
+from typing import Any
+
 from poetry.core.packages.package import Package
+
+from poetry.installation.operations.update import Update
 from poetry.puzzle.transaction import Transaction
 
 
-def check_operations(ops, expected):
+if TYPE_CHECKING:
+    from poetry.installation.operations.operation import Operation
+
+
+def check_operations(ops: list[Operation], expected: list[dict[str, Any]]) -> None:
     for e in expected:
         if "skipped" not in e:
             e["skipped"] = False
 
     result = []
     for op in ops:
-        if "update" == op.job_type:
+        if op.job_type == "update":
+            assert isinstance(op, Update)
             result.append(
                 {
                     "job": "update",
@@ -25,10 +37,10 @@ def check_operations(ops, expected):
 
             result.append({"job": job, "package": op.package, "skipped": op.skipped})
 
-    assert expected == result
+    assert result == expected
 
 
-def test_it_should_calculate_operations_in_correct_order():
+def test_it_should_calculate_operations_in_correct_order() -> None:
     transaction = Transaction(
         [Package("a", "1.0.0"), Package("b", "2.0.0"), Package("c", "3.0.0")],
         [
@@ -48,7 +60,7 @@ def test_it_should_calculate_operations_in_correct_order():
     )
 
 
-def test_it_should_calculate_operations_for_installed_packages():
+def test_it_should_calculate_operations_for_installed_packages() -> None:
     transaction = Transaction(
         [Package("a", "1.0.0"), Package("b", "2.0.0"), Package("c", "3.0.0")],
         [
@@ -79,7 +91,7 @@ def test_it_should_calculate_operations_for_installed_packages():
     )
 
 
-def test_it_should_remove_installed_packages_if_required():
+def test_it_should_remove_installed_packages_if_required() -> None:
     transaction = Transaction(
         [Package("a", "1.0.0"), Package("b", "2.0.0"), Package("c", "3.0.0")],
         [
@@ -111,7 +123,32 @@ def test_it_should_remove_installed_packages_if_required():
     )
 
 
-def test_it_should_update_installed_packages_if_sources_are_different():
+def test_it_should_not_remove_installed_packages_that_are_in_result() -> None:
+    transaction = Transaction(
+        [],
+        [
+            (Package("a", "1.0.0"), 1),
+            (Package("b", "2.0.0"), 2),
+            (Package("c", "3.0.0"), 0),
+        ],
+        installed_packages=[
+            Package("a", "1.0.0"),
+            Package("b", "2.0.0"),
+            Package("c", "3.0.0"),
+        ],
+    )
+
+    check_operations(
+        transaction.calculate_operations(synchronize=True),
+        [
+            {"job": "install", "package": Package("a", "1.0.0"), "skipped": True},
+            {"job": "install", "package": Package("b", "2.0.0"), "skipped": True},
+            {"job": "install", "package": Package("c", "3.0.0"), "skipped": True},
+        ],
+    )
+
+
+def test_it_should_update_installed_packages_if_sources_are_different() -> None:
     transaction = Transaction(
         [Package("a", "1.0.0")],
         [
