@@ -1,40 +1,60 @@
+from __future__ import annotations
+
 from pathlib import Path
+from typing import TYPE_CHECKING
+from typing import Any
 
 import pytest
 
-from entrypoints import EntryPoint
+from cleo.io.buffered_io import BufferedIO
+from deepdiff import DeepDiff
+from packaging.utils import canonicalize_name
+from poetry.core.constraints.version import parse_constraint
+from poetry.core.packages.vcs_dependency import VCSDependency
 
-from poetry.core.semver.helpers import parse_constraint
-from poetry.core.toml.file import TOMLFile
+from poetry.exceptions import PoetryException
 from poetry.factory import Factory
 from poetry.plugins.plugin import Plugin
+from poetry.repositories.exceptions import InvalidSourceError
 from poetry.repositories.legacy_repository import LegacyRepository
 from poetry.repositories.pypi_repository import PyPiRepository
+from poetry.repositories.repository_pool import Priority
+from poetry.toml.file import TOMLFile
+from tests.helpers import mock_metadata_entry_points
 
 
-fixtures_dir = Path(__file__).parent / "fixtures"
+if TYPE_CHECKING:
+    from cleo.io.io import IO
+    from pytest_mock import MockerFixture
+
+    from poetry.config.config import Config
+    from poetry.poetry import Poetry
+    from tests.types import FixtureDirGetter
 
 
 class MyPlugin(Plugin):
-    def activate(self, poetry, io):
-        io.write_line("Updating version")
-        poetry.package.set_version("9.9.9")
+    def activate(self, poetry: Poetry, io: IO) -> None:
+        io.write_line("Setting readmes")
+        poetry.package.readmes = (Path("README.md"),)
 
 
-def test_create_poetry():
-    poetry = Factory().create_poetry(fixtures_dir / "sample_project")
+def test_create_poetry(fixture_dir: FixtureDirGetter) -> None:
+    poetry = Factory().create_poetry(fixture_dir("sample_project"))
 
     package = poetry.package
 
-    assert package.name == "my-package"
+    assert package.name == "sample-project"
     assert package.version.text == "1.2.3"
     assert package.description == "Some description."
     assert package.authors == ["Sébastien Eustace <sebastien@eustace.io>"]
+    assert package.license is not None
     assert package.license.id == "MIT"
-    assert (
-        package.readme.relative_to(fixtures_dir).as_posix()
-        == "sample_project/README.rst"
-    )
+
+    for readme in package.readmes:
+        assert (
+            readme.relative_to(fixture_dir("sample_project")).as_posix() == "README.rst"
+        )
+
     assert package.homepage == "https://python-poetry.org"
     assert package.repository_url == "https://github.com/python-poetry/poetry"
     assert package.keywords == ["packaging", "dependency", "poetry"]
@@ -46,56 +66,58 @@ def test_create_poetry():
     for dep in package.requires:
         dependencies[dep.name] = dep
 
-    cleo = dependencies["cleo"]
+    cleo = dependencies[canonicalize_name("cleo")]
     assert cleo.pretty_constraint == "^0.6"
     assert not cleo.is_optional()
 
-    pendulum = dependencies["pendulum"]
+    pendulum = dependencies[canonicalize_name("pendulum")]
     assert pendulum.pretty_constraint == "branch 2.0"
     assert pendulum.is_vcs()
+    assert isinstance(pendulum, VCSDependency)
     assert pendulum.vcs == "git"
     assert pendulum.branch == "2.0"
     assert pendulum.source == "https://github.com/sdispater/pendulum.git"
     assert pendulum.allows_prereleases()
 
-    requests = dependencies["requests"]
+    requests = dependencies[canonicalize_name("requests")]
     assert requests.pretty_constraint == "^2.18"
     assert not requests.is_vcs()
     assert not requests.allows_prereleases()
     assert requests.is_optional()
     assert requests.extras == frozenset(["security"])
 
-    pathlib2 = dependencies["pathlib2"]
+    pathlib2 = dependencies[canonicalize_name("pathlib2")]
     assert pathlib2.pretty_constraint == "^2.2"
     assert parse_constraint(pathlib2.python_versions) == parse_constraint("~2.7")
     assert not pathlib2.is_optional()
 
-    demo = dependencies["demo"]
+    demo = dependencies[canonicalize_name("demo")]
     assert demo.is_file()
     assert not demo.is_vcs()
     assert demo.name == "demo"
     assert demo.pretty_constraint == "*"
 
-    demo = dependencies["my-package"]
+    demo = dependencies[canonicalize_name("my-package")]
     assert not demo.is_file()
     assert demo.is_directory()
     assert not demo.is_vcs()
     assert demo.name == "my-package"
     assert demo.pretty_constraint == "*"
 
-    simple_project = dependencies["simple-project"]
+    simple_project = dependencies[canonicalize_name("simple-project")]
     assert not simple_project.is_file()
     assert simple_project.is_directory()
     assert not simple_project.is_vcs()
     assert simple_project.name == "simple-project"
     assert simple_project.pretty_constraint == "*"
 
-    functools32 = dependencies["functools32"]
+    functools32 = dependencies[canonicalize_name("functools32")]
     assert functools32.name == "functools32"
     assert functools32.pretty_constraint == "^3.2.3"
     assert (
         str(functools32.marker)
-        == 'python_version ~= "2.7" and sys_platform == "win32" or python_version in "3.4 3.5"'
+        == 'python_version ~= "2.7" and sys_platform == "win32" or python_version in'
+        ' "3.4 3.5"'
     )
 
     assert "db" in package.extras
@@ -117,13 +139,55 @@ def test_create_poetry():
         "Programming Language :: Python :: 3.8",
         "Programming Language :: Python :: 3.9",
         "Programming Language :: Python :: 3.10",
+        "Programming Language :: Python :: 3.11",
         "Topic :: Software Development :: Build Tools",
         "Topic :: Software Development :: Libraries :: Python Modules",
     ]
 
 
-def test_create_poetry_with_packages_and_includes():
-    poetry = Factory().create_poetry(fixtures_dir / "with-include")
+@pytest.mark.parametrize(
+    ("project",),
+    [
+        ("simple_project",),
+        ("project_with_extras",),
+    ],
+)
+def test_create_pyproject_from_package(
+    project: str, fixture_dir: FixtureDirGetter
+) -> None:
+    poetry = Factory().create_poetry(fixture_dir(project))
+    package = poetry.package
+
+    pyproject: dict[str, Any] = Factory.create_pyproject_from_package(package)
+
+    result = pyproject["tool"]["poetry"]
+    expected = poetry.pyproject.poetry_config
+
+    # Extras are normalized as they are read.
+    extras = expected.pop("extras", None)
+    if extras is not None:
+        normalized_extras = {
+            canonicalize_name(extra): dependencies
+            for extra, dependencies in extras.items()
+        }
+        expected["extras"] = normalized_extras
+
+    # packages do not support this at present
+    expected.pop("scripts", None)
+
+    # remove any empty sections
+    sections = list(expected.keys())
+    for section in sections:
+        if not expected[section]:
+            expected.pop(section)
+
+    assert not DeepDiff(expected, result)
+
+
+def test_create_poetry_with_packages_and_includes(
+    fixture_dir: FixtureDirGetter,
+) -> None:
+    poetry = Factory().create_poetry(fixture_dir("with-include"))
 
     package = poetry.package
 
@@ -143,9 +207,11 @@ def test_create_poetry_with_packages_and_includes():
     ]
 
 
-def test_create_poetry_with_multi_constraints_dependency():
+def test_create_poetry_with_multi_constraints_dependency(
+    fixture_dir: FixtureDirGetter,
+) -> None:
     poetry = Factory().create_poetry(
-        fixtures_dir / "project_with_multi_constraints_dependency"
+        fixture_dir("project_with_multi_constraints_dependency")
     )
 
     package = poetry.package
@@ -153,112 +219,286 @@ def test_create_poetry_with_multi_constraints_dependency():
     assert len(package.requires) == 2
 
 
-def test_poetry_with_default_source(with_simple_keyring):
-    poetry = Factory().create_poetry(fixtures_dir / "with_default_source")
-
-    assert 1 == len(poetry.pool.repositories)
-
-
-def test_poetry_with_non_default_source(with_simple_keyring):
-    poetry = Factory().create_poetry(fixtures_dir / "with_non_default_source")
-
-    assert len(poetry.pool.repositories) == 2
-
-    assert not poetry.pool.has_default()
-
-    assert poetry.pool.repositories[0].name == "foo"
-    assert isinstance(poetry.pool.repositories[0], LegacyRepository)
-
-    assert poetry.pool.repositories[1].name == "PyPI"
-    assert isinstance(poetry.pool.repositories[1], PyPiRepository)
-
-
-def test_poetry_with_non_default_secondary_source(with_simple_keyring):
-    poetry = Factory().create_poetry(fixtures_dir / "with_non_default_secondary_source")
-
-    assert len(poetry.pool.repositories) == 2
-
-    assert poetry.pool.has_default()
-
-    repository = poetry.pool.repositories[0]
-    assert repository.name == "PyPI"
-    assert isinstance(repository, PyPiRepository)
-
-    repository = poetry.pool.repositories[1]
-    assert repository.name == "foo"
-    assert isinstance(repository, LegacyRepository)
-
-
-def test_poetry_with_non_default_multiple_secondary_sources(with_simple_keyring):
-    poetry = Factory().create_poetry(
-        fixtures_dir / "with_non_default_multiple_secondary_sources"
-    )
-
-    assert len(poetry.pool.repositories) == 3
-
-    assert poetry.pool.has_default()
-
-    repository = poetry.pool.repositories[0]
-    assert repository.name == "PyPI"
-    assert isinstance(repository, PyPiRepository)
-
-    repository = poetry.pool.repositories[1]
-    assert repository.name == "foo"
-    assert isinstance(repository, LegacyRepository)
-
-    repository = poetry.pool.repositories[2]
-    assert repository.name == "bar"
-    assert isinstance(repository, LegacyRepository)
-
-
-def test_poetry_with_non_default_multiple_sources(with_simple_keyring):
-    poetry = Factory().create_poetry(fixtures_dir / "with_non_default_multiple_sources")
-
-    assert len(poetry.pool.repositories) == 3
-
-    assert not poetry.pool.has_default()
-
-    repository = poetry.pool.repositories[0]
-    assert repository.name == "bar"
-    assert isinstance(repository, LegacyRepository)
-
-    repository = poetry.pool.repositories[1]
-    assert repository.name == "foo"
-    assert isinstance(repository, LegacyRepository)
-
-    repository = poetry.pool.repositories[2]
-    assert repository.name == "PyPI"
-    assert isinstance(repository, PyPiRepository)
-
-
-def test_poetry_with_no_default_source():
-    poetry = Factory().create_poetry(fixtures_dir / "sample_project")
+def test_poetry_with_default_source_legacy(
+    fixture_dir: FixtureDirGetter, with_simple_keyring: None
+) -> None:
+    io = BufferedIO()
+    poetry = Factory().create_poetry(fixture_dir("with_default_source_legacy"), io=io)
 
     assert len(poetry.pool.repositories) == 1
-
-    assert poetry.pool.has_default()
-
-    assert poetry.pool.repositories[0].name == "PyPI"
-    assert isinstance(poetry.pool.repositories[0], PyPiRepository)
+    assert "Found deprecated key" in io.fetch_error()
 
 
-def test_poetry_with_two_default_sources(with_simple_keyring):
+def test_poetry_with_default_source(
+    fixture_dir: FixtureDirGetter, with_simple_keyring: None
+) -> None:
+    io = BufferedIO()
+    poetry = Factory().create_poetry(fixture_dir("with_default_source"), io=io)
+
+    assert len(poetry.pool.repositories) == 1
+    assert io.fetch_error() == ""
+
+
+def test_poetry_with_default_source_and_pypi(
+    fixture_dir: FixtureDirGetter, with_simple_keyring: None
+) -> None:
+    io = BufferedIO()
+    poetry = Factory().create_poetry(fixture_dir("with_default_source_and_pypi"), io=io)
+
+    assert len(poetry.pool.repositories) == 2
+    assert poetry.pool.has_repository("PyPI")
+    assert isinstance(poetry.pool.repository("PyPI"), PyPiRepository)
+    assert poetry.pool.get_priority("PyPI") is Priority.PRIMARY
+    assert "Warning: Found deprecated key" not in io.fetch_error()
+
+
+def test_poetry_with_default_source_pypi(
+    fixture_dir: FixtureDirGetter, with_simple_keyring: None
+) -> None:
+    io = BufferedIO()
+    poetry = Factory().create_poetry(fixture_dir("with_default_source_pypi"), io=io)
+
+    assert len(poetry.pool.repositories) == 1
+    assert poetry.pool.has_repository("PyPI")
+    assert isinstance(poetry.pool.repository("PyPI"), PyPiRepository)
+    assert poetry.pool.get_priority("PyPI") is Priority.DEFAULT
+
+
+@pytest.mark.parametrize(
+    "project",
+    ("with_non_default_source_implicit", "with_non_default_source_explicit"),
+)
+def test_poetry_with_non_default_source(
+    project: str, fixture_dir: FixtureDirGetter, with_simple_keyring: None
+) -> None:
+    io = BufferedIO()
+    poetry = Factory().create_poetry(fixture_dir(project), io=io)
+
+    assert not poetry.pool.has_default()
+    assert poetry.pool.has_repository("PyPI")
+    assert poetry.pool.get_priority("PyPI") is Priority.SECONDARY
+    assert isinstance(poetry.pool.repository("PyPI"), PyPiRepository)
+    assert poetry.pool.has_repository("foo")
+    assert poetry.pool.get_priority("foo") is Priority.PRIMARY
+    assert isinstance(poetry.pool.repository("foo"), LegacyRepository)
+    assert {repo.name for repo in poetry.pool.repositories} == {"PyPI", "foo"}
+    error = io.fetch_error()
+    assert "Warning: In a future version of Poetry, PyPI will be disabled" in error
+
+
+def test_poetry_with_non_default_secondary_source_legacy(
+    fixture_dir: FixtureDirGetter, with_simple_keyring: None
+) -> None:
+    poetry = Factory().create_poetry(
+        fixture_dir("with_non_default_secondary_source_legacy")
+    )
+
+    assert poetry.pool.has_repository("PyPI")
+    assert isinstance(poetry.pool.repository("PyPI"), PyPiRepository)
+    assert poetry.pool.get_priority("PyPI") is Priority.DEFAULT
+    assert poetry.pool.has_repository("foo")
+    assert isinstance(poetry.pool.repository("foo"), LegacyRepository)
+    assert {repo.name for repo in poetry.pool.repositories} == {"PyPI", "foo"}
+
+
+def test_poetry_with_non_default_secondary_source(
+    fixture_dir: FixtureDirGetter, with_simple_keyring: None
+) -> None:
+    poetry = Factory().create_poetry(fixture_dir("with_non_default_secondary_source"))
+
+    assert poetry.pool.has_repository("PyPI")
+    assert isinstance(poetry.pool.repository("PyPI"), PyPiRepository)
+    assert poetry.pool.get_priority("PyPI") is Priority.DEFAULT
+    assert poetry.pool.has_repository("foo")
+    assert isinstance(poetry.pool.repository("foo"), LegacyRepository)
+    assert {repo.name for repo in poetry.pool.repositories} == {"PyPI", "foo"}
+
+
+def test_poetry_with_non_default_multiple_secondary_sources_legacy(
+    fixture_dir: FixtureDirGetter,
+    with_simple_keyring: None,
+) -> None:
+    poetry = Factory().create_poetry(
+        fixture_dir("with_non_default_multiple_secondary_sources_legacy")
+    )
+
+    assert poetry.pool.has_repository("PyPI")
+    assert isinstance(poetry.pool.repository("PyPI"), PyPiRepository)
+    assert poetry.pool.get_priority("PyPI") is Priority.DEFAULT
+    assert poetry.pool.has_repository("foo")
+    assert isinstance(poetry.pool.repository("foo"), LegacyRepository)
+    assert poetry.pool.has_repository("bar")
+    assert isinstance(poetry.pool.repository("bar"), LegacyRepository)
+    assert {repo.name for repo in poetry.pool.repositories} == {"PyPI", "foo", "bar"}
+
+
+def test_poetry_with_non_default_multiple_secondary_sources(
+    fixture_dir: FixtureDirGetter, with_simple_keyring: None
+) -> None:
+    poetry = Factory().create_poetry(
+        fixture_dir("with_non_default_multiple_secondary_sources")
+    )
+
+    assert poetry.pool.has_repository("PyPI")
+    assert isinstance(poetry.pool.repository("PyPI"), PyPiRepository)
+    assert poetry.pool.get_priority("PyPI") is Priority.DEFAULT
+    assert poetry.pool.has_repository("foo")
+    assert isinstance(poetry.pool.repository("foo"), LegacyRepository)
+    assert poetry.pool.has_repository("bar")
+    assert isinstance(poetry.pool.repository("bar"), LegacyRepository)
+    assert {repo.name for repo in poetry.pool.repositories} == {"PyPI", "foo", "bar"}
+
+
+def test_poetry_with_non_default_multiple_sources_legacy(
+    fixture_dir: FixtureDirGetter, with_simple_keyring: None
+) -> None:
+    poetry = Factory().create_poetry(
+        fixture_dir("with_non_default_multiple_sources_legacy")
+    )
+
+    assert not poetry.pool.has_default()
+    assert poetry.pool.has_repository("bar")
+    assert isinstance(poetry.pool.repository("bar"), LegacyRepository)
+    assert poetry.pool.has_repository("PyPI")
+    assert poetry.pool.get_priority("PyPI") is Priority.SECONDARY
+    assert isinstance(poetry.pool.repository("PyPI"), PyPiRepository)
+    assert poetry.pool.has_repository("foo")
+    assert isinstance(poetry.pool.repository("foo"), LegacyRepository)
+    assert {repo.name for repo in poetry.pool.repositories} == {"bar", "PyPI", "foo"}
+
+
+def test_poetry_with_non_default_multiple_sources(
+    fixture_dir: FixtureDirGetter, with_simple_keyring: None
+) -> None:
+    poetry = Factory().create_poetry(fixture_dir("with_non_default_multiple_sources"))
+
+    assert not poetry.pool.has_default()
+    assert poetry.pool.has_repository("PyPI")
+    assert isinstance(poetry.pool.repository("PyPI"), PyPiRepository)
+    assert poetry.pool.get_priority("PyPI") is Priority.SECONDARY
+    assert poetry.pool.has_repository("bar")
+    assert isinstance(poetry.pool.repository("bar"), LegacyRepository)
+    assert poetry.pool.has_repository("foo")
+    assert isinstance(poetry.pool.repository("foo"), LegacyRepository)
+    assert {repo.name for repo in poetry.pool.repositories} == {"PyPI", "bar", "foo"}
+
+
+def test_poetry_with_non_default_multiple_sources_pypi(
+    fixture_dir: FixtureDirGetter, with_simple_keyring: None
+) -> None:
+    io = BufferedIO()
+    poetry = Factory().create_poetry(
+        fixture_dir("with_non_default_multiple_sources_pypi"), io=io
+    )
+
+    assert len(poetry.pool.repositories) == 4
+    assert not poetry.pool.has_default()
+    assert poetry.pool.has_repository("PyPI")
+    assert isinstance(poetry.pool.repository("PyPI"), PyPiRepository)
+    assert poetry.pool.get_priority("PyPI") is Priority.PRIMARY
+    # PyPI must be between bar and baz!
+    expected = ["bar", "PyPI", "baz", "foo"]
+    assert [repo.name for repo in poetry.pool.repositories] == expected
+    error = io.fetch_error()
+    assert (
+        error.strip()
+        == "<warning>Warning: Found deprecated priority 'secondary' for source 'foo' in"
+        " pyproject.toml. Consider changing the priority to one of the"
+        " non-deprecated values: 'default', 'primary', 'supplemental', 'explicit'."
+    )
+
+
+def test_poetry_with_no_default_source(fixture_dir: FixtureDirGetter) -> None:
+    poetry = Factory().create_poetry(fixture_dir("sample_project"))
+
+    assert poetry.pool.has_repository("PyPI")
+    assert poetry.pool.get_priority("PyPI") is Priority.DEFAULT
+    assert isinstance(poetry.pool.repository("PyPI"), PyPiRepository)
+    assert {repo.name for repo in poetry.pool.repositories} == {"PyPI"}
+
+
+def test_poetry_with_supplemental_source(
+    fixture_dir: FixtureDirGetter, with_simple_keyring: None
+) -> None:
+    poetry = Factory().create_poetry(fixture_dir("with_supplemental_source"))
+
+    assert poetry.pool.has_repository("PyPI")
+    assert poetry.pool.get_priority("PyPI") is Priority.DEFAULT
+    assert isinstance(poetry.pool.repository("PyPI"), PyPiRepository)
+    assert poetry.pool.has_repository("supplemental")
+    assert poetry.pool.get_priority("supplemental") is Priority.SUPPLEMENTAL
+    assert isinstance(poetry.pool.repository("supplemental"), LegacyRepository)
+    assert {repo.name for repo in poetry.pool.repositories} == {"PyPI", "supplemental"}
+
+
+def test_poetry_with_explicit_source(
+    fixture_dir: FixtureDirGetter, with_simple_keyring: None
+) -> None:
+    poetry = Factory().create_poetry(fixture_dir("with_explicit_source"))
+
+    assert len(poetry.pool.repositories) == 1
+    assert len(poetry.pool.all_repositories) == 2
+    assert poetry.pool.has_repository("PyPI")
+    assert poetry.pool.get_priority("PyPI") is Priority.DEFAULT
+    assert isinstance(poetry.pool.repository("PyPI"), PyPiRepository)
+    assert poetry.pool.has_repository("explicit")
+    assert isinstance(poetry.pool.repository("explicit"), LegacyRepository)
+    assert {repo.name for repo in poetry.pool.repositories} == {"PyPI"}
+
+
+def test_poetry_with_explicit_pypi_and_other(
+    fixture_dir: FixtureDirGetter, with_simple_keyring: None
+) -> None:
+    io = BufferedIO()
+    poetry = Factory().create_poetry(fixture_dir("with_explicit_pypi_and_other"), io=io)
+
+    assert len(poetry.pool.repositories) == 1
+    assert len(poetry.pool.all_repositories) == 2
+    error = io.fetch_error()
+    assert error == ""
+
+
+@pytest.mark.parametrize(
+    "project", ["with_explicit_pypi_no_other", "with_explicit_pypi_and_other_explicit"]
+)
+def test_poetry_with_pypi_explicit_only(
+    project: str, fixture_dir: FixtureDirGetter, with_simple_keyring: None
+) -> None:
+    with pytest.raises(PoetryException) as e:
+        Factory().create_poetry(fixture_dir(project))
+    assert str(e.value) == "At least one source must not be configured as 'explicit'."
+
+
+def test_poetry_with_two_default_sources_legacy(
+    fixture_dir: FixtureDirGetter, with_simple_keyring: None
+) -> None:
     with pytest.raises(ValueError) as e:
-        Factory().create_poetry(fixtures_dir / "with_two_default_sources")
+        Factory().create_poetry(fixture_dir("with_two_default_sources_legacy"))
 
-    assert "Only one repository can be the default" == str(e.value)
+    assert str(e.value) == "Only one repository can be the default."
 
 
-def test_validate():
-    complete = TOMLFile(fixtures_dir / "complete.toml")
-    content = complete.read()["tool"]["poetry"]
+def test_poetry_with_two_default_sources(
+    fixture_dir: FixtureDirGetter, with_simple_keyring: None
+) -> None:
+    with pytest.raises(ValueError) as e:
+        Factory().create_poetry(fixture_dir("with_two_default_sources"))
+
+    assert str(e.value) == "Only one repository can be the default."
+
+
+def test_validate(fixture_dir: FixtureDirGetter) -> None:
+    complete = TOMLFile(fixture_dir("complete.toml"))
+    pyproject: dict[str, Any] = complete.read()
+    content = pyproject["tool"]["poetry"]
 
     assert Factory.validate(content) == {"errors": [], "warnings": []}
 
 
-def test_validate_fails():
-    complete = TOMLFile(fixtures_dir / "complete.toml")
-    content = complete.read()["tool"]["poetry"]
+def test_validate_fails(fixture_dir: FixtureDirGetter) -> None:
+    complete = TOMLFile(fixture_dir("complete.toml"))
+    pyproject: dict[str, Any] = complete.read()
+    content = pyproject["tool"]["poetry"]
     content["this key is not in the schema"] = ""
 
     expected = (
@@ -269,34 +509,60 @@ def test_validate_fails():
     assert Factory.validate(content) == {"errors": [expected], "warnings": []}
 
 
-def test_create_poetry_fails_on_invalid_configuration():
+def test_create_poetry_fails_on_invalid_configuration(
+    fixture_dir: FixtureDirGetter,
+) -> None:
     with pytest.raises(RuntimeError) as e:
-        Factory().create_poetry(
-            Path(__file__).parent / "fixtures" / "invalid_pyproject" / "pyproject.toml"
-        )
+        Factory().create_poetry(fixture_dir("invalid_pyproject") / "pyproject.toml")
 
     expected = """\
 The Poetry configuration is invalid:
   - 'description' is a required property
+  - Project name (invalid) is same as one of its dependencies
 """
-    assert expected == str(e.value)
+    assert str(e.value) == expected
 
 
-def test_create_poetry_with_local_config(fixture_dir):
+def test_create_poetry_with_local_config(fixture_dir: FixtureDirGetter) -> None:
     poetry = Factory().create_poetry(fixture_dir("with_local_config"))
 
     assert not poetry.config.get("virtualenvs.in-project")
     assert not poetry.config.get("virtualenvs.create")
     assert not poetry.config.get("virtualenvs.options.always-copy")
+    assert not poetry.config.get("virtualenvs.options.no-pip")
+    assert not poetry.config.get("virtualenvs.options.no-setuptools")
     assert not poetry.config.get("virtualenvs.options.system-site-packages")
 
 
-def test_create_poetry_with_plugins(mocker):
-    mocker.patch(
-        "entrypoints.get_group_all",
-        return_value=[EntryPoint("my-plugin", "tests.test_factory", "MyPlugin")],
-    )
+def test_create_poetry_with_plugins(
+    mocker: MockerFixture, fixture_dir: FixtureDirGetter
+) -> None:
+    mock_metadata_entry_points(mocker, MyPlugin)
 
-    poetry = Factory().create_poetry(fixtures_dir / "sample_project")
+    poetry = Factory().create_poetry(fixture_dir("sample_project"))
 
-    assert "9.9.9" == poetry.package.version.text
+    assert poetry.package.readmes == (Path("README.md"),)
+
+
+@pytest.mark.parametrize(
+    ("source", "expected"),
+    [
+        ({}, "Missing [name] in source."),
+        ({"name": "foo"}, "Missing [url] in source 'foo'."),
+        (
+            {"name": "PyPI", "url": "https://example.com"},
+            "The PyPI repository cannot be configured with a custom url.",
+        ),
+    ],
+)
+def test_create_package_source_invalid(
+    source: dict[str, str],
+    expected: str,
+    config: Config,
+    fixture_dir: FixtureDirGetter,
+) -> None:
+    with pytest.raises(InvalidSourceError) as e:
+        Factory.create_package_source(source, config=config)
+        Factory().create_poetry(fixture_dir("with_source_pypi_url"))
+
+    assert str(e.value) == expected
